% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{Adj.Rsq.blblm}
\alias{Adj.Rsq.blblm}
\title{Get the Adjusted R^2}
\usage{
Adj.Rsq.blblm(object, confidence = FALSE, level = 0.95)
}
\arguments{
\item{object}{A variable with the class "blblm"}

\item{confidence}{Boolean for whether a confidence interval should be reported}

\item{level}{The significance level for a confidence interval (1 - alpha)}
}
\value{
An estimate for the Adjusted R^2 (or a confidence interval vector)
}
\description{
Get an estimate for the adjusted coefficient of determination, with the option of a confidence interval.
}
