% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbLM.R
\name{blb.sigma.ci}
\alias{blb.sigma.ci}
\title{Bag of Little Bootstrap Confidence Interval for Sigma (Linear Model)}
\usage{
blb.sigma.ci(data, y, x, m, r, alpha)
}
\arguments{
\item{data}{a dataframe or tibble object}

\item{y}{the name of the response variable}

\item{x}{the name of the indepedent variables}

\item{m}{the number of group to split}

\item{r}{the number of iteration for each boot}

\item{alpha}{the significance level}
}
\value{
returns a matrix of the bootstrap ci
}
\description{
Bag of Little Bootstrap Confidence Interval for Sigma (Linear Model)
}
\examples{
blb.sigma.ci(iris, "Sepal.Length",  "Petal.Width", 10, 15, 0.05)
blb.sigma.ci(iris, "Sepal.Length", c("Sepal.Width", "Petal.Width"), 10, 15, 0.05)
}
