% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blb_coef}
\alias{blb_coef}
\title{Coefficients}
\usage{
blb_coef(object, formula)
}
\arguments{
\item{object}{Fitted linear regression model}

\item{formula}{Linear regression formula}
}
\value{
Model coefficients
}
\description{
Computes coefficients from fit with lmC
}
