% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blb_logreg}
\alias{blb_logreg}
\title{Logistic Regression}
\usage{
blb_logreg(formula, data, m = 3, B = 100)
}
\arguments{
\item{formula}{The formula to be used.}

\item{data}{The dataframe to be used.}

\item{m}{Number of splits to make in the dataframe,}

\item{B}{Number of bootstraps to be performed.}
}
\description{
Computes the coefficients for the logistic regression.
}
\examples{
data <- iris
labels <- rep(0:1,75)
data$Species <- labels
fit <- blb_logreg(Species ~ Petal.Length * Sepal.Length, data = data, m = 2, B = 100)
coef(fit)
}
