% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blb_sigmaC}
\alias{blb_sigmaC}
\title{Sigma}
\usage{
blb_sigmaC(object)
}
\arguments{
\item{object}{Fitted linear regression model}
}
\value{
Sigma value
}
\description{
Computes sigma value from fit with lmC
}
