% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Linear Regression Coefficients}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{the fitted lm object}
}
\description{
An internal function to pull linear regression
coefficient estimates from the linear model
}
