% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\alias{blbglm_files}
\alias{glm_subsample_boot}
\title{Generalized Linear Regression with Little Bag of Bootstraps}
\usage{
blbglm(formula, family, data, m = 10, B = 5000)

blbglm_files(formula, file_paths, B = 5000)

glm_subsample_boot(formula, family, df, n, B)
}
\arguments{
\item{formula}{mpg ~ wt * hp.see \link[stats:formula]{formula}}

\item{family}{\code{stats::binomial()}, see \link[stats:family]{family}}

\item{data}{All data.}

\item{m}{The number of subsamples.Integer.}

\item{B}{Times of bootstrap}

\item{file_paths}{The path to csv/tsv files, which are subsamples of the dataset.}

\item{df}{data.frame from a split subsample}

\item{n}{The nrow of all data}
}
\value{
list of coef and sigma
}
\description{
If \code{family = gaussian()}, \code{blbglm} is the same as \code{blblm}.
}
\examples{
\dontrun{
fit <- blbglm(vs ~ mpg+hp, family = binomial(),data = mtcars, m = 3, B = 100)
coef(fit)
confint(fit, c("wt", "hp"))
}

\dontrun{
file_pahts <- list.files(pattern = "*.csv")
blblm_files(vs ~ mpg+hp, family = binomial(),file_pahts, B = 100)
}
n <- nrow(mtcars)
data_list <- split_data(mtcars, 3)
glm_subsample_boot(vs ~ mpg+hp, family = binomial(), data_list[[1]], n, 100)
}
