% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm_file.R
\name{blbglm_file}
\alias{blbglm_file}
\title{Fit a general linear regression model with file imput with more than one CPUs}
\usage{
blbglm_file(formula, family = gaussian, file, B = 5000)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{family}{which indicates which kinds of glm we use}

\item{file}{which is usual is the data frame, which is the data we imput.}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}
}
\value{
blblm object
}
\description{
Fit the linear regression model with file input with more than one CPUs
}
\examples{
#library(parallel)
#cl<- makeCluster(4)
# fit<- blbglm.file(Species~., data = read.csv("iris.csv"), m = 10, B = 5000,cl)
# stopCluster(cl)
}
