% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm_file_par.R
\name{blbglm_file_par}
\alias{blbglm_file_par}
\title{Fit a general linear regression model with file imput with more than one CPUs}
\usage{
blbglm_file_par(formula, family = gaussian, file, B = 5000, cl)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{file}{which is usual is a list of names of file, which is the data we imput.}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}

\item{cl}{The muticluster we get from parallel}

\item{m}{which is numeric variables, indicates the number of splits we need}
}
\value{
blblm object
}
\description{
Fit the linear regression model with file input with more than one CPUs
}
\examples{
#library(parallel)
#cl<- makeCluster(4)
# fit<- blbglm_file_par(Species~., data = read.csv("iris.csv"), m = 10, B = 5000,cl)
# stopCluster(cl)
}
\seealso{
Other family which indicates which kinds of glm we use: 
\code{\link{blbglm_file}()},
\code{\link{blbglm_par}()},
\code{\link{blbglm}()}
}
\concept{family which indicates which kinds of glm we use}
