% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm_par.R
\name{blbglm_par}
\alias{blbglm_par}
\title{Fit a general linear regression model  with more than one CPUs}
\usage{
blbglm_par(formula, family = gaussian, data, m = 10, B = 5000, cl)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{family}{which indicates which kinds of glm we use}

\item{data}{which is usual is the data frame, which is the data we imput.}

\item{m}{which is numeric variables, indicates the number of splits we need}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}

\item{cl}{The muticluster we get from parallel}
}
\value{
blblm object
}
\description{
Fit the linear regression model  with more than one CPUs
}
\examples{
#library(parallel)
#cl<- makeCluster(4)
# fit<- blbglm_par(Species~., data = read.csv("iris.csv"), m = 10, B = 5000)
# stopCluster(cl)
}
