% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm_par_c}
\alias{blbglm_par_c}
\title{Generate logistic regression with boostrap(without sigma)}
\usage{
blbglm_par_c(formula, data, m = 10, B = 5000, cl)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{m}{number of subdata}

\item{B}{number of bootstrips}

\item{cl}{cluster specified}
}
\value{
class blblm
}
\description{
create a logistic regression with the method of bootstrips, with people selected feature. Now, support parallel with user specified cluster
}
\examples{
\dontrun{
cl <- makec(2)
x=blblm_par(mpg~wt,data=mtcars,m=3,B=100,cl=cl)
stopCluster(cl)
coef(x)
}
}
