% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm_parallel}
\alias{blbglm_parallel}
\title{fit a blbglm_parallel class}
\usage{
blbglm_parallel(
  formula,
  data,
  family = binomial(link = "logit"),
  m = 10,
  B = 5000,
  cluster = 1
)
}
\arguments{
\item{formula}{formula indicating y and x variables}

\item{data}{dataframe}

\item{family}{family of glm, logit model as default.}

\item{m}{number of splits}

\item{B}{number of bootstraps}

\item{cluster}{number of clusters to use for parallel computing. when = 1 will be running on single thread.}
}
\description{
fit bootstrap GLM models and utilize parallel computing
}
\examples{
#proj = blblm::blbglm_parallel(vs~mpg+disp+gear, mtcars, m = 3, B = 100, cluster = 2)

}
