% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{blbglmcoef}
\alias{blbglmcoef}
\title{Extracts Coefficients
compute the coefficients from fit}
\usage{
blbglmcoef(fit)
}
\arguments{
\item{fit}{a fitted blbglm model}
}
\value{
numeric values
}
\description{
Extracts Coefficients
compute the coefficients from fit
}
