% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\docType{package}
\name{blbglmpar-package}
\alias{blbglmpar}
\alias{blbglmpar-package}
\title{blbglmpar: Fitting Generalized Linear Models Using Bootstrap}
\description{
blbglmpar is used to fit generalized linear models, 
    specified by giving a symbolic description of the linear predictor and a description of the error distribution. blbglmpar also has parallelization capabilities.
}
\details{
Generalized Linear Model with Little Bag of Bootstraps and Parallelization
}
