% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{blbglmsigma}
\alias{blbglmsigma}
\title{Extract Residual Standard Deviation}
\usage{
blbglmsigma(fit)
}
\arguments{
\item{fit}{a fitted blbglm model}
}
\value{
a numeric value
}
\description{
compute sigma from fit
}
