% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbgm}
\alias{blbgm}
\title{Construct the logistic regression with bootstraps and each kind of parament might used.}
\usage{
blbgm(formula, data, m = 10, B = 5000, cl = 1)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{m}{m parts user want to split the data}

\item{B}{the number of replications}

\item{cl}{the number of cores that user want to use.}
}
\description{
Construct the logistic regression with bootstraps and each kind of parament might used.
}
