% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of little bootstrap for linear regression}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{file_list}{a character vector. file names that containing data.}

\item{path}{character. the path of the file list.}

\item{m}{the number of subsamples.}

\item{B}{the repeat time of bootstrapping for each subsample.}

\item{parallel}{logical. If TRUE the function will use parallelization.}

\item{num_cl}{numeric. The number of clusters to be created for parallelization.}

\item{logit}{logical. If TRUE the function will do logistic regression.}
}
\value{
an object of class "blblm"
}
\description{
R statistical functions for linear regression with little bag of bootstraps.
}
\details{
Bag of little bootstraps works by averaging the results of bootstrapping multiple subsets. BLB is well suited to modern parallel architectures.
}
