% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblm_par.R
\name{blblm.par}
\alias{blblm.par}
\title{Title}
\usage{
\method{blblm}{par}(formula, data, B = 5000, cl)

\method{blblm}{par}(formula, data, B = 5000, cl)
}
\arguments{
\item{formula}{the regression formula}

\item{data}{user-specified datasets}

\item{B}{the number of bootstraps}

\item{cl}{the number of clusters that the users choose}
}
\description{
Title
}
\examples{
blblm.par(distance ~ air_time, c("flights/1.csv","flights/2.csv"), B = 100,cl=4)
}
