% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\alias{blblm_files}
\alias{lm_subsample_boot}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000)

blblm_files(formula, file_paths, B = 5000)

lm_subsample_boot(formula, df, n, B)
}
\arguments{
\item{formula}{mpg ~ wt * hp.see \link[stats:formula]{formula}}

\item{data}{All data.}

\item{m}{The number of subsamples.Integer.}

\item{B}{Times of bootstrap}

\item{file_paths}{The path to csv/tsv files, which are subsamples of the dataset.}

\item{df}{data.frame from a split subsample}

\item{n}{The nrow of all data}
}
\value{
list of coef and sigma
}
\description{
Linear Regression with Little Bag of Bootstraps

Subsample Parallel lm bootstrap
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
coef(fit)
confint(fit, c("wt", "hp"))


\dontrun{
file_pahts = list.files(pattern = "*.csv")
blblm_files(mpg ~ wt * hp, file_pahts, B = 100)
}

n =  nrow(mtcars)
data_list = split_data(mtcars, 3)
lm_subsample_boot(mpg ~ wt * hp,data_list[[1]], n,100)
}
