% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm1}
\alias{blblm1}
\title{generalized linear model}
\usage{
blblm1(
  formula,
  data,
  m = 10,
  B = 5000,
  ifgeneral = FALSE,
  family = "gaussian",
  wrkrs = 6
)
}
\arguments{
\item{formula}{formula (old)}

\item{data}{data to be used}

\item{m}{split count}

\item{ifgeneral}{TRUE is generalized model to be used}

\item{family}{GML family parameter}

\item{wrkrs}{number of workers}

\item{Bbootstrap}{depth}
}
\description{
generalized linear model
}
