% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_benchmarks}
\alias{blblm_benchmarks}
\title{Benchmark parallelization}
\usage{
blblm_benchmarks(formula, data, m, B)
}
\arguments{
\item{formula}{Formula for the regression}

\item{data}{The input data frame.}

\item{m}{The number of splits to make on the data. m > 1.}

\item{B}{The number of bootstraps to execute. B > 1.}
}
\value{
Returns a dataframe with results for both memory and time usage.
}
\description{
Run blblm with the number of cores varied from 1 to max available and compare their benchmarks.
}
\details{
An example you can run is:
blblm_benchmarks(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
note: This cannot be ran like a normal example due to an issue with the future package.
}
