% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_csv}
\alias{blblm_csv}
\title{blblm() with CSV files}
\usage{
blblm_csv(filenames, formu, B = 5000, useParallel = FALSE, useFreqs = TRUE)
}
\arguments{
\item{filenames}{List (or vector) of CSV files}

\item{formu}{Regression formula}

\item{B}{Number of bootstraps per subsample}

\item{useParallel}{Boolean for using parallel functions (future::plan() must be set before using this)}

\item{useFreqs}{Boolean for generating frequencies for each row of data (used in weighted linear regression)}
}
\value{
A list with (1) elements separated by filename and bootstrap trial and (2) the model formula
(Note: The filenames are used as list names in the output).
Within each trial, the output includes the model coefficients, sigma, R^2, Adjusted R^2, and Weighted R^2 (if weights are used).
}
\description{
Rather than use a data frame variable as input, the data is read in from specified filenames.
Since the data is already split, a shortened version of blblm() is applied separately to the data within each file.
}
