% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_f}
\alias{blblm_f}
\title{Generate linear regression with boostrap}
\usage{
blblm_f(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{m}{number of subdata}

\item{B}{number of bootstrips}
}
\value{
class blblm
}
\description{
create a linear regression with the method of bootstrips, with people selected feature, and this function is using future map so you need to use it with planf
to specified the number of workers you want
}
\examples{
\dontrun{
planf(4)
x=blblm_f(mpg~wt,data=mtcars,m=3,B=100)
coef(x)}
}
