% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_fast.R
\name{blblm_fast}
\alias{blblm_fast}
\title{Fit a linear regression model for a fast method}
\usage{
blblm_fast(x, y, m = 10, B = 5000)
}
\arguments{
\item{x}{which is usual matrix as the predictive variable, if you want to include the intercetion, please include the 1 vector in the first colmun}

\item{y}{the reaponse variable}

\item{m}{which is numeric variables, indicates the number of splits we need}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}
}
\value{
blblm object
}
\description{
Fit the linear regression model with matrix input
}
\examples{
#x<- matrix(c(1,1,1,4,5,3,1,3,5),3,3)
#y<- c(2,4,5)
# fit<- blblm_fast(x, y)
}
