% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_file.R
\name{blblm_file}
\alias{blblm_file}
\title{Fit a linear regression model with file input}
\usage{
blblm_file(formula, file, m = 10, B = 5000)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{file}{which is usual the list of name of .csv files, which is the data we imput.}

\item{m}{which is numeric variables, indicates the number of splits we need}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}
}
\value{
blblm object
}
\description{
Fit the linear regression model with data frame input with file input
}
\examples{
# file<- c("file01.csv","file02.csv")
# fit<- blblm_file(mpg~., file, m = 10, B = 5000,cl)
}
