% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_file_par.R
\name{blblm_file_par}
\alias{blblm_file_par}
\title{Fit a linear regression model with file input and with more than one CPU works}
\usage{
blblm_file_par(formula, file, B = 5000, cl)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{file}{which is usual the list of name of .csv files, which is the data we imput.}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}

\item{cl}{The muticluster we get from parallel}
}
\value{
blblm object
}
\description{
Fit the linear regression model with data frame input with file input and with more than one CPU works.
}
\examples{
# library(parallel)
# cl<- makeCluster(4)
# file<- c("file01.csv","file02.csv")
# fit<- blblm_file_par(mpg~., file, B = 5000,cl)
# stopCluster(cl)
}
