% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_general}
\alias{blblm_general}
\title{Fit a linear regression model}
\usage{
blblm_general(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{data}{which is usual is the data frame, which is the data we imput.}

\item{m}{which is numeric variables, indicates the number of splits we need}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}
}
\value{
blblm object
}
\description{
Fit the linear regression model with data frame input
}
\examples{
# fit<- blblm_general(mpg~., data = mtcars, m = 10, B = 5000)
}
