% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_glm}
\alias{blblm_glm}
\title{Creating the glm for blb fucntion for Linear regression}
\usage{
blblm_glm(formula, data, m = 10, B = 200, parallel = FALSE)
}
\arguments{
\item{formula}{in the model}

\item{data}{part of the dataset}

\item{m}{a subsample number}

\item{B}{number for bootstrap}

\item{parallel}{Boolean}
}
\value{
model with cluster and blb techniques
}
\description{
This function computes the bootstrap samplefor glm
}
