% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_log}
\alias{blblm_log}
\title{Little Bag of Bootstraps with Logistic Regression
Same as blblm but with logistic regression
Also note that confint and predict are not implemented}
\usage{
blblm_log(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{The linear regression formula, in the form of y ~ x1 + x2 + ...}

\item{data}{Dataset you would like to analyze}

\item{m}{The number of subsamples you want the data to be split into, default is 10}

\item{B}{The number of Bootstrap trials you want to run on each subsample, default is 5000}
}
\value{
An object that contains the regression formula ($formula) and all of the estimates from each Bootstrap trial in a vector.
}
\description{
Little Bag of Bootstraps with Logistic Regression
Same as blblm but with logistic regression
Also note that confint and predict are not implemented
}
\examples{
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
pred_log <- blblm_log(admit ~ gre + gpa + rank, data = mydata)

}
