% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_mod}
\alias{blblm_mod}
\title{Bag of Little Bootstraps Algorithm}
\usage{
blblm_mod(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{m}{number of subsets for the data}

\item{B}{number of bootstraps}

\item{parallel}{TRUE/FALSE for parallel computing}
}
\value{
Returns the bag of little bootstrap confidence intervals for sigma for each boot of the linear model
}
\description{
To run calculations parallelly for faster speed, first detect num of cores by running the following command:
\code{future::availableCores()}
Then specify number of cores you want use in the following line:
\verb{plan(multisession, worker = #_cores)}
}
\examples{
blblm_mod(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
blblm_mod(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = TRUE)
blblm_mod(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
}
