% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_multi_file}
\alias{blblm_multi_file}
\title{Little Bag of Bootstraps with Linear Regression (Multiple Files)
Same as blblm but with a vector of file names, the data from each file will be its own subsample}
\usage{
blblm_multi_file(formula, data, B = 5000)
}
\arguments{
\item{formula}{The linear regression formula, in the form of y ~ x1 + x2 + ...}

\item{data}{Vector of file names, each files serves as a subsample}

\item{B}{The number of Bootstrap trials you want to run on each subsample, default is 5000}
}
\description{
Little Bag of Bootstraps with Linear Regression (Multiple Files)
Same as blblm but with a vector of file names, the data from each file will be its own subsample
}
\examples{
set.seed(141)
1:10 \%>\% walk(function(i) {
  dt <- tibble(x = rnorm(5), y = rnorm(5))
  write_csv(dt, file.path("files", sprintf("file\%02d.csv", i)))
})
file_names <- file.path("files", list.files("files"))
fit_multi <- blblm_multi_file(y ~ x, file_names)
}
