% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_p}
\alias{blblm_p}
\title{Linear Regression with Little Bag of Bootstraps and parallel computation}
\usage{
blblm_p(formula, data, m = 10, B = 5000, n_cores = 4)
}
\arguments{
\item{formula}{a linear formula}

\item{data}{a dataframe or dataset}

\item{m}{number of subsets}

\item{B}{bumber of bootstrap}

\item{n_cores}{number of cores}
}
\value{
an invisible blblm object
}
\description{
Linear Regression with Little Bag of Bootstraps and parallel computation
}
