% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{Creating the blblm parallel fucntion for Linear regression}
\usage{
blblm_par(formula, data, m, B, cl)
}
\arguments{
\item{formula}{in the model}

\item{data}{part of the dataset}

\item{m}{a subsample number}

\item{B}{number for bootstrap}

\item{cl}{the clusters}
}
\value{
model with cluster and blb techniques
}
\description{
This function computes the bootstrap sample with parallelization
}
\examples{
fit2 <- blblm_par(mpg~wt * hp, data = mtcars, m = 3, B = 100,cl = 3)
coef(fit2)
confint(fit2,c("wt", "hp"))
}
