% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par_user}
\alias{blblm_par_user}
\title{Parallel regression, User specified data}
\usage{
blblm_par_user(formula, path, ncores, B)
}
\arguments{
\item{formula}{The linear regression formula.}

\item{path}{The path to the folder containing the csv files.}

\item{ncores}{The desires number of cores to use. ncores > 0.}

\item{B}{The number of bootstraps to be performed. B > 1.}
}
\description{
Allows the usewr to computer coefficients for the linear regression using a specified
number of cores, and using their own input csv files.
}
\details{
An example:
dir.create("files", showWarnings = FALSE)
set.seed(141)
1:100 \%>\% walk(function(i) {
dt <- tibble(x = rnorm(5000), y = rnorm(5000))
write_csv(dt, file.path("files", sprintf("file\%02d.csv", i)))
})
fit <- blblm_par_user(formula = y ~ 2 * x, path = 'files',ncores = 2, B = 100)
coef(fit)

Note: This cannot be run in check() due to an error with the future package.
}
