% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_parallel}
\alias{blblm_parallel}
\title{fit a blblm_parallel class}
\usage{
blblm_parallel(formula, data, m = 10, B = 5000, cluster = 1)
}
\arguments{
\item{formula}{formula indicating y and x variables}

\item{data}{dataframe}

\item{m}{number of splits}

\item{B}{number of bootstraps}

\item{cluster}{number of clusters to use for parallel computing. when = 1 will be running on single thread.}
}
\value{
blblm_parallel object
}
\description{
fit bootstrap LM models and utilize parallel computing
}
\examples{
# proj = blblm::blblm_parallel(disp~mpg, mtcars, m = 3, B = 100, cluster = 4)

}
