% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_parallel_list}
\alias{blblm_parallel_list}
\title{blblm_parallel on a list of csv files with same structure}
\usage{
blblm_parallel_list(formula, file_names, B = 5000, cluster = 1)
}
\arguments{
\item{formula}{formula indicating y and x variables}

\item{file_names}{dataframe}

\item{B}{number of bootstraps}

\item{cluster}{number of clusters to use for parallel computing. when = 1 will be running on single thread. more than 2 cores is not compatible in packages by default}
}
\description{
fit bootstrap LM models and utilize parallel computing on a list of files. notice the files should be csv and all column names should exactly match.
}
\examples{
# listnames = c('~/blblm/files/data01.csv', '~/blblm/files/data02.csv',
# '~/blblm/files/data03.csv', '~/blblm/files/data04.csv')
# projec = blblm_parallel_list(mpg ~ cyl + hp, listnames, B = 100, cluster = 1)
}
