% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_pl}
\alias{blblm_pl}
\title{Linear Regression with one or more cpus}
\usage{
blblm_pl(formula, data, m, B, Cluster)
}
\arguments{
\item{formula}{a formula}

\item{data}{dataframe}

\item{m}{numeric}

\item{B}{numeric}

\item{Cluster}{numeric}
}
\value{
linear regression
}
\description{
Linear Regression with one or more cpus
}
\examples{
blblm_pl(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, 4)
}
