% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_single_file}
\alias{blblm_single_file}
\title{Little Bag of Bootstraps with Linear Regression (Single File)}
\usage{
blblm_single_file(formula, data, B = 5000)
}
\arguments{
\item{formula}{The linear regression formula, in the form of y ~ x1 + x2 + ...}

\item{data}{File name in CSV format you want to analyze}

\item{B}{The number of Bootstrap trials you want to run on each subsample, default is 5000}
}
\value{
An object that contains the regression formula ($formula) and all of the estimates from each Bootstrap trial in a vector.
}
\description{
This function performs in the same way as blblm, but using a single file instead of a predefined dataset.
Also this function does NOT split files into subsamples, it assumes your "subsample" is the single file.
}
\examples{
# dir.create("files", showWarnings = FALSE)
set.seed(141)
1:10 \%>\% walk(function(i) {
  dt <- tibble(x = rnorm(5), y = rnorm(5))
  write_csv(dt, file.path("files", sprintf("file\%02d.csv", i)))
})
file_names <- file.path("files", list.files("files"))
setting up files to use - single file that puts together all multiple files
write.csv(file_names \%>\% lapply(read_csv) \%>\% bind_rows, 'files/total.csv')
blblm_single_file(y ~ x, 'files/total.csv', B = 10)
}
