% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{blblogcoef}
\alias{blblogcoef}
\title{Compute the coefficients from fit}
\usage{
blblogcoef(fit)
}
\arguments{
\item{fit}{fitted blblm model}
}
\value{
numeric value
}
\description{
Compute the coefficients from fit
}
