% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{blblogsigma}
\alias{blblogsigma}
\title{Compute sigma from fit}
\usage{
blblogsigma(fit)
}
\arguments{
\item{fit}{fitted blblm model}
}
\value{
numeric value
}
\description{
Compute sigma from fit
}
