% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blbmethods-package}
\alias{blbmethods}
\alias{blbmethods-package}
\title{blbmethods: Bag of Little Bootstraps for Various Models}
\description{
Implements Bag of Little Boostrap Algorithm for Linear Regression, Random Forests, and Logistic Regression.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: David Mao \email{dlmao@ucdavis.edu}

}
