% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbrf.R
\name{blbrf}
\alias{blbrf}
\title{Bag of Little Boostrap for Random Forest}
\usage{
blbrf(formula, data, m = 10, B = 500, nthreads = 1)
}
\arguments{
\item{formula}{object of class "formula"}

\item{data}{a dataframe}

\item{m}{an integer}

\item{B}{an integer}

\item{nthreads}{an integer}
}
\value{
blbrf
}
\description{
Implementation of the Bag of Little Bootstrap algorithm for Random Forest.
Includes methods blb estimates for predictions. Based on the R Ranger package.
}
\examples{
blbrf(mpg ~ wt * hp, data = mtcars, m = 3, B = 10000, nthreads = 8)
}
