% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Linear Regression Sigma}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{the fitted lm object}
}
\description{
An internal function to pull linear regression
coefficient sigma from the linear model
}
