% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigmaC}
\alias{blbsigmaC}
\title{Compute sigma from fit with lmC (compatible).
This functions is used for compatability with the rcpp version of lm.}
\usage{
blbsigmaC(fit)
}
\arguments{
\item{fit}{the fitted model returned from blblm}
}
\description{
Compute sigma from fit with lmC (compatible).
This functions is used for compatability with the rcpp version of lm.
}
