% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blglml.R
\docType{package}
\name{blglml}
\alias{blblm}
\alias{blblm-package}
\title{blglml}
\description{
This pakage is going to bootstrap for linear regression
 and logistic regression
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
