% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_blblm.R
\name{choice_blblm}
\alias{choice_blblm}
\title{User choice for blblm with or without parallelization}
\usage{
choice_blblm(formula, data, m, B, work = 4)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{from dataset}

\item{m}{integer number of subsamples data is separated into}

\item{B}{integer number of bootstrap samples}

\item{work}{integer number of workers to be used if parallization is used}
}
\value{
list of linear regression coefficients and standard deviation for each bootstrap sample
}
\description{
Input is read from user whether parallelization will be used to implement bag of little
bootstraps(blblm)
}
\examples{
choice_blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, work = 4)
}
