% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Coefficients of Bag of Little Bootstraps}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{the fitted object}

\item{...}{additional parameters related to coef.lm}
}
\description{
The mean coefficients for the BLB
}
