% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blbglm_parallel}
\alias{coef.blbglm_parallel}
\title{coefficients of blbglm_parallel project}
\usage{
\method{coef}{blbglm_parallel}(object, ...)
}
\arguments{
\item{object}{a fitted blbglm_parallel class}

\item{...}{others}
}
\description{
coefficients method of blbglm_parallel class returning coefficients
}
\examples{
#proj = blblm::blbglm_parallel(vs~mpg+disp, mtcars, m = 3, B = 100, cluster = 2)
#coef(proj)
}
