% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Calculate the estimates coef}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{\code{blblm} object}

\item{...}{}
}
\description{
Calculate the estimates coef
}
