% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm_parallel}
\alias{coef.blblm_parallel}
\title{coefficients of blblm_parallel project}
\usage{
\method{coef}{blblm_parallel}(object, ...)
}
\arguments{
\item{object}{a fitted blblm_parallel class}

\item{...}{others}
}
\description{
coefficients method of blblm_parallel class returning coefficients
}
\examples{
# proj = blblm::blblm_parallel(disp~mpg, mtcars, m = 3, B = 100, cluster = 4)
# coef(proj)
}
