% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{coef.blblog}
\alias{coef.blblog}
\title{Computing the Coefficients for the Model}
\usage{
\method{coef}{blblog}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{additional arguments affecting the result}
}
\description{
Computing the Coefficients for the Model
}
