% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblogreg}
\alias{coef.blblogreg}
\title{Coefficients for blblogreg}
\usage{
\method{coef}{blblogreg}(object, ...)
}
\arguments{
\item{object}{model fit produced by blblogreg(formula, data, ...)}

\item{...}{other arguments.}
}
\description{
Outputs the coefficients for the fitted object (after blblogreg algorithm)
}
