% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Confidence Interval of Bag of Little Bootstraps}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the fitted object}

\item{parm}{parameters for requested conf interval}

\item{level}{confidence level}

\item{...}{additional parameters related to confint.lm}
}
\description{
The confidence interval of coefficients for the BLB
}
