% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Calculate the confidence level of independent variables}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{\code{blblm} object}

\item{parm}{variables that are used to fit the blblm model.}

\item{level}{Confidence level, default 0.95.}

\item{...}{}
}
\description{
Calculate the confidence level of independent variables
}
