% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm_parallel}
\alias{confint.blblm_parallel}
\title{confidence intervals of blblm_parallel coefficients}
\usage{
\method{confint}{blblm_parallel}(object, parm = NULL, level = 0.95, cluster = 1, ...)
}
\arguments{
\item{object}{a fitted blblm_parallel class}

\item{parm}{a vector of parameter names or NULL for all parameters in the model}

\item{level}{confidence level}

\item{cluster}{number of clusters to use for parallel computing. when = 1 will be running on single thread.}

\item{...}{others}
}
\description{
coefficients method of blblm_parallel class returning coefficients
}
\examples{
# proj = blblm::blblm_parallel(disp~mpg, mtcars, m = 3, B = 100, cluster = 2)
# confint(proj, parm = NULL, level = 0.99, cluster = 1)
}
