% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{blblogreg Confidence Interval}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{model fit produced by blblogreg(formula, data, ...)}

\item{parm}{character vector, parameters to give confidence interval for. Defaults to all predictors.}

\item{level}{numeric, confidence interval level.}

\item{...}{other arguments.}
}
\description{
Constructs a confidence interval for parameters of the logistic model
}
