% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confints.blb}
\alias{confints.blb}
\title{confints.blb}
\usage{
confints.blb(object, level = 0.95)
}
\arguments{
\item{object}{object}

\item{level}{numeric}
}
\value{

}
\description{
computes confidence interval of bootstrap coefficients
}
