% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blglml.R
\name{data}
\alias{data}
\title{read data in to one list}
\usage{
data(filename, n)

data(filename, n)
}
\arguments{
\item{filename}{"name"}

\item{n}{integer}
}
\description{
read data in to one list

read data in to one list
}
