% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{est_df}
\alias{est_df}
\title{Create Dataframe of Estimate Values}
\usage{
est_df(object, value)
}
\arguments{
\item{object}{the fitted object}

\item{value}{the estimate value requested}
}
\description{
An internal function to compute build a dataframe
based on requested estimate value
}
