% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_lm}
\alias{fast_lm}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
fast_lm(X, y)
}
\arguments{
\item{X}{matrix for explanatory}

\item{y}{column vector for response variable}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
