% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastlm}
\alias{fastlm}
\alias{fastLm_impl}
\title{An Rcpp function that constructs a linear regression model}
\usage{
fastLm_impl(X, y)
}
\arguments{
\item{X}{numeric vector}

\item{y}{numeric vector}
}
\value{
numeric scalar
}
\description{
An Rcpp function that constructs a linear regression model
}
