% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastLm_impl}
\alias{fastLm_impl}
\title{A fast implementation of weighted linear regression based on the similar RcppArmadillo implementation.}
\usage{
fastLm_impl(X, y, w)
}
\arguments{
\item{X}{A matrix of independent variables}

\item{y}{A vector representing the resulting dependent variable}

\item{w}{A vector of weights of equal length as the number of observations of x and y given to the model.}
}
\description{
A fast implementation of weighted linear regression based on the similar RcppArmadillo implementation.
}
