% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastwLm_impl}
\alias{fastwLm_impl}
\title{An Rcpp function to quickly perform weighted linear regression}
\arguments{
\item{x}{design matrix}

\item{y}{numeric vector of response variables}

\item{w}{numeric vector of weights}
}
\value{
list
}
\description{
An Rcpp function to quickly perform weighted linear regression
}
