% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{future_blblogreg}
\alias{future_blblogreg}
\title{Bag of Little Bootstraps Logistic Regression Model with Parallelization}
\usage{
future_blblogreg(formula, data, m = 10, B = 5000, w = 4)
}
\arguments{
\item{formula}{A symbolic description of the logistic regression model to be fitted.}

\item{data}{A data set, such as a data frame or list, containing the variables in the model.}

\item{m}{Number indicating how many sub-samples the data will be split into.}

\item{B}{Number of bootstrap samples.}

\item{w}{Number of workers.}
}
\value{
Coefficients of each fit.
}
\description{
Fits a logistic regression model on a given data set using the bag of little bootstraps algorithm and parallelization for efficiency.
}
