% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{Linear Regression Model}
\usage{
glm1(formula, data, family, n, freqs)
}
\arguments{
\item{formula}{the formula to be modelled}

\item{data}{the source data}

\item{family}{the model family}

\item{n}{the sample size}

\item{freqs}{multinomial dist weights}
}
\description{
An internal function to compute linear regression estimates
based on given number of repetitions
}
