% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1_c}
\alias{glm1_c}
\title{Generate logistic regression with boostrap(without sigma)}
\usage{
glm1_c(formula, data, freqs)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{freqs}{frequency}
}
\value{
estimate the regression estimates based on given the number of repetitions
estimate the regression estimates based on given the number of repetitions
}
\description{
estimate the regression estimates based on given the number of repetitions
}
