% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{GLM_blblm}
\alias{GLM_blblm}
\title{GLM model}
\usage{
GLM_blblm(formula, data, m, B)
}
\arguments{
\item{formula}{fit linear model}

\item{data}{a data set}

\item{m}{integer
split data into m parts of approximated equal sizes}

\item{B}{repeat B times}
}
\description{
GLM model
}
