% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Linear Regression Model of Each Bootstrap}
\usage{
glm_each_boot(formula, data, family, n)
}
\arguments{
\item{formula}{the formula to be modelled}

\item{data}{the source data}

\item{family}{the model family}

\item{n}{the sample size}
}
\description{
An internal function to compute linear regression estimates
for each boot of a given bag of bootstraps
}
