% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm_file.R
\name{glm_each_boot_file}
\alias{glm_each_boot_file}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot_file(formula, file, family)
}
\description{
compute the regression estimates for a blb dataset
}
