% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample_c}
\alias{glm_each_subsample_c}
\title{Generate logistic regression with boostrap(without sigma)}
\usage{
glm_each_subsample_c(formula, data, n, B)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{n}{number of row of the data}

\item{B}{number of bootstrips}
}
\value{
vector
compute the estimates
}
\description{
compute the estimates
}
