% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{gm1}
\alias{gm1}
\title{estimate the logistic estimates based on given number of repetitions}
\usage{
gm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{freqs}{how many multinomial probabilities exists}

\item{family}{the model family}
}
\description{
estimate the logistic estimates based on given number of repetitions
}
