% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{gm_each_boot}
\alias{gm_each_boot}
\title{compute the logistic regression estimates for a blb dataset}
\usage{
gm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{n}{the number of random vectors to draw}
}
\description{
compute the logistic regression estimates for a blb dataset
}
