% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{in the model}

\item{data}{part of the parameters}

\item{freqs}{part of the paramters}
}
\value{
model with cluster and blb techniques
}
\description{
estimate the regression estimates based on given the number of repetitions
}
