% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1C}
\alias{lm1C}
\title{estimate the regression estimates using lm in rcpp}
\usage{
lm1C(formula, data, freqs)
}
\arguments{
\item{formula}{a lm formula}

\item{data}{a data frame containing the features mentioned in formula}

\item{freqs}{the wights computed from a multinomial distribution in lm_each_boot}
}
\description{
estimate the regression estimates using lm in rcpp
}
