% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm2}
\alias{lm2}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm2(formula, data, freqs)
}
\arguments{
\item{formula}{fomula: V1~V2}

\item{data}{dataframe}

\item{freqs}{multinomially distributed random number vectors}
}
\value{
list
}
\description{
estimate the regression estimates based on given the number of repetitions
}
