% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_each_boot1}
\alias{lm_each_boot1}
\title{lm_each_boot}
\usage{
lm_each_boot1(formula, data, n)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{n}{number of multiprobability}

\item{how}{many parts that we want to split}
}
\description{
lm_each_boot
}
