% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_fast.R
\name{lm_each_boot_fast}
\alias{lm_each_boot_fast}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot_fast(data, n)
}
\description{
compute the regression estimates for a blb dataset
}
