% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_fast_file.R
\name{lm_each_boot_fast_file}
\alias{lm_each_boot_fast_file}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot_fast_file(file)
}
\description{
compute the regression estimates for a blb dataset
}
