% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{in the model}

\item{data}{the data for part of the paramters}

\item{n}{subsample numeric number}

\item{B}{number for bootstrap}
}
\value{
model with cluster and blb techniques
}
\description{
compute the estimates
}
