% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_fast.R
\name{lm_fast}
\alias{lm_fast}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm_fast(data)
}
\description{
estimate the regression estimates based on given the number of repetitions
}
