% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmC}
\alias{lmC}
\title{Compute coefficients and sigma in a weighted linear regression}
\usage{
lmC(X, y, W, n, p)
}
\arguments{
\item{X}{design matrix with rank p.}

\item{y}{response vector}

\item{W}{weighted diagonal matrix}

\item{n}{the size of whole data}

\item{p}{the rank of X}
}
\value{
a list including coef and sigma
}
\description{
In order to speed up the function of lm in R, I write this Rcpp file to solve
coefficients and sigma on my own. Firstly, it needs RcppArmadillo to make basic
linear computation. We should first input design matrix X, response vector y,
weights diagnoal matrix W, the size of whole data and the rank p.
Then coef = (X^T W X)^(-1)(X^T W y) and sigma = sqrt(W * (X*coef - y)^2/(n-p)).
}
