% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmW}
\alias{lmW}
\title{Weighted Least Squares in RcppArmadillo.}
\usage{
lmW(X, y, w)
}
\arguments{
\item{X}{numeric matrix}

\item{y}{numeric vector}

\item{w}{numeric vector}
}
\value{
list
}
\description{
Weighted Least Squares in RcppArmadillo.
}
