% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{log1}
\alias{log1}
\title{Logistic Model for blblm
estimate the logistic regression estimates based on given the number of repetitions}
\usage{
log1(formula, data, freqs)
}
\arguments{
\item{formula}{Formula of logistic regression}

\item{data}{Data from a particular Bootstrap trial}

\item{freqs}{Numeric vector that describes how often each sample should be included in the model (elements should add up to the total number of rows in the data)}
}
\description{
Logistic Model for blblm
estimate the logistic regression estimates based on given the number of repetitions
}
