% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{log_each_boot}
\alias{log_each_boot}
\title{Logistic Regression for a Bootstrap Trial
Compute the logistic regression estimates for a blb dataset}
\usage{
log_each_boot(formula, data, n)
}
\arguments{
\item{formula}{Formula of logistic regression}

\item{data}{Subsample data}

\item{n}{Number of rows in data}
}
\description{
Logistic Regression for a Bootstrap Trial
Compute the logistic regression estimates for a blb dataset
}
