% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{log_each_subsample}
\alias{log_each_subsample}
\title{Logistic regression on Subsamples
Compute the estimates (log regression)}
\usage{
log_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{Formula of logistic regression}

\item{data}{Subsample data}

\item{n}{Number of rows in data}

\item{B}{The number of Bootstrap trials you want to run on each subsample}
}
\description{
Logistic regression on Subsamples
Compute the estimates (log regression)
}
