% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{logreg1}
\alias{logreg1}
\title{Fit Logistic regresion model}
\usage{
logreg1(formula, data)
}
\arguments{
\item{formula}{Logistic regression model}

\item{data}{Dataframes to have their coefficients estimated.}
}
\description{
Fit Logistic regresion model
}
