% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{logreg_each_subsample}
\alias{logreg_each_subsample}
\title{Replicate the input data B times.}
\usage{
logreg_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{Regression formula.}

\item{data}{The data.}

\item{n}{Number of total data points to be replicated.}

\item{B}{Number of bootstraps.}
}
\description{
Replicate the input data B times.
}
