% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblr.R
\name{lr1}
\alias{lr1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lr1(formula, data, freqs)
}
\description{
estimate the regression estimates based on given the number of repetitions
}
