% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_blblm.R
\name{lst_blblm}
\alias{lst_blblm}
\title{Bag of Little Bootstraps using List of File Names}
\usage{
lst_blblm(formula, data, B = 5000, file_names)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{dataset}

\item{B}{integer bootstrap sample number}

\item{file_names}{list of file names the data is split into}
}
\value{
list of linear regression coefficients and standard deviation for each bootstrap sample
}
\description{
Instead of splitting the data within the blblm function, the data is already split up into multiple files.
This function will read in a list of those file names as an extra parameter and carry out the blblm function.
}
