% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_parblblm.R
\name{lst_parblblm}
\alias{lst_parblblm}
\title{Bag of Little Bootstraps using List of File Names with Parallelization}
\usage{
lst_parblblm(formula, data, B = 5000, file_names, work = 4)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{dataset}

\item{B}{integer bootstrap sample number}

\item{file_names}{list of file names the data is split into}

\item{work}{integer number of workers for parallelization}
}
\value{
list of linear regression coefficients and standard deviation for each bootstrap sample
}
\description{
Data has to already have been split into multiple files. The data is then read in
through a list of file names containing the data as a parameter to be specified. The blblm function
with parallelization is then used to return linear regression coefficients and standard deviation.
The amount of workers to be used can also be specified as a parameter.
}
