% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_mean}
\alias{map_mean}
\title{Applies a function to an object via map and then calculcates the mean of the result}
\usage{
map_mean(.x, .f, ...)
}
\arguments{
\item{.x}{object}

\item{.f}{function}

\item{...}{further arguments}
}
\value{
numeric mean
}
\description{
Applies a function to an object via map and then calculcates the mean of the result
}
