% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_rbind}
\alias{map_rbind}
\title{Applies a function to an object via map and binds the rows}
\usage{
map_rbind(.x, .f, ...)
}
\arguments{
\item{.x}{object}

\item{.f}{function}

\item{...}{further arguments}
}
\value{
tibble
}
\description{
Applies a function to an object via map and binds the rows
}
