% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{mean_lwr_upr}
\alias{mean_lwr_upr}
\title{Mean of confidence interval}
\usage{
mean_lwr_upr(x, level = 0.95)
}
\arguments{
\item{x}{fit data}

\item{level}{confidence level}
}
\description{
computes mean of confidence interval from bootstrap
}
