% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blblm}
\alias{par_blblm}
\title{Bag of Little Bootstraps Linear Model (parallelization)}
\usage{
par_blblm(formula, data, m = 10, B = 5000, cl = 2)
}
\arguments{
\item{formula}{Linear regression formula}

\item{data}{Chosen dataset}

\item{m}{Number of times to split data for bootstrapping}

\item{B}{umber of bootstrap samples}

\item{cl}{umber of cores to use (Made using `cl = makeCluster()`)}
}
\value{
Model coefficients
}
\description{
Computes coefficeints for a linear model using blb (multi core)
}
