% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blblm_chunk}
\alias{par_blblm_chunk}
\title{Exact to par_blblm but makes us of chunking}
\usage{
par_blblm_chunk(formula, data, m = 10, B = 5000, cl, ck)
}
\arguments{
\item{formula}{an lm formula}

\item{data}{the dataset containing the features mentioned in formula}

\item{m}{the number of subsamples the data is to be split into}

\item{B}{the number of bootstrap samples to take}

\item{cl}{this is equvalent to makeCluster and is created by the user outside the funtion.}

\item{ck}{the chunk size}
}
\value{
list
}
\description{
Exact to par_blblm but makes us of chunking
}
\examples{
\dontrun{
cl <- makeCluster(4)
par_blblm_chunk(mpg ~ hp, mtcars, 10, 100, cl, 2)
stopCluster(cl)
}

}
