% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{parallelization}
\alias{parallelization}
\title{Gives the user the choice if he/she would like to use parallelization in the blblm calculations.
User can specify number of workers.}
\usage{
parallelization(parallel = FALSE, workers)
}
\arguments{
\item{parallel}{logical Indicate whether or not you want to use parallelization.}

\item{workers}{numeric Specify number of workers to distribute tasks.}
}
\value{
logical TRUE if using parallelization, FALSE if not.
}
\description{
Gives the user the choice if he/she would like to use parallelization in the blblm calculations.
User can specify number of workers.
}
\examples{
parallelization(FALSE)
}
