% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parblblm.R
\name{parblblm}
\alias{parblblm}
\title{Bag of Little Bootstraps using Parallelization}
\usage{
parblblm(formula, data, m = 10, B = 5000, work = 4)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{dataset}

\item{m}{integer number of subsamples data is split into}

\item{B}{integer bootstrap number}

\item{work}{integer number of workers}
}
\value{
list of linear regression coefficients and standard deviation for each bootstrap sample
}
\description{
Bag of little bootstraps for linear regression using parallelization. The user can prespecify the amount
of workers used
}
