% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{parLB_blblm}
\alias{parLB_blblm}
\title{Bag of Little Bootstraps Linear Model (parallelization and load balancing/dynamic scheduling)}
\usage{
parLB_blblm(formula, data, m = 10, B = 5000, cl = 2)
}
\arguments{
\item{formula}{Linear regression formula}

\item{data}{Chosen dataset}

\item{m}{umber of times to split data for bootstrapping}

\item{B}{Number of bootstrap samples}

\item{cl}{Number of cores to use (Made using `cl = makeCluster()`)}
}
\value{
Model coefficients
}
\description{
Computes coefficeints using blb and dynamic scheduling (multi core)
}
