% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parralel.R
\name{parr.blbglm}
\alias{parr.blbglm}
\title{parr.blbglm}
\usage{
parr.blbglm(formula, family, data, B, cores)
}
\arguments{
\item{formula}{formula}

\item{family, }{family}

\item{data}{data.frame}

\item{B}{numeric}

\item{cores}{numeric}
}
\value{
list
}
\description{
computes bootstraps with parallel computing
}
\details{
Linear Regression with Little Bag of Bootstraps
}
\examples{
parr.blbglm(Survived~Class+Sex,binomial, data.frame(Titanic),100,2)
}
