% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parr.R
\name{parr.blblm}
\alias{parr.blblm}
\title{parr.blblm}
\usage{
parr.blblm(formula, data, B, cores)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}

\item{B}{numeric}

\item{cores}{numeric}
}
\value{
list
}
\description{
computes bootstraps with parallel computing
}
\details{
Linear Regression with Little Bag of Bootstraps
}
\examples{
parr.blblm(wt~mpg,mtcars,1000,2)
}
