% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Predicting with Bag of Little Bootstraps}
\usage{
\method{predict}{blbglm}(object, newdata, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the fitted object}

\item{newdata}{data used to predict}

\item{confidence}{boolean request for confidence interval}

\item{level}{confidence level}

\item{...}{additional parameters related to predict.lm}
}
\description{
Prediction with the mean estimates of the bag of boots
on new data.
}
