% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict the the new data use fitted blblm model.}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE,
  level = 0.95, ...)
}
\arguments{
\item{object}{\code{blblm} object}

\item{new_data}{New data.frame of varables}

\item{confidence}{Boolean, TRUE OR FALSE.}

\item{level}{Confidence level, default 0.95.}

\item{...}{}
}
\description{
Predict the the new data use fitted blblm model.
}
