% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm_parallel}
\alias{predict.blblm_parallel}
\title{predict blblm_parallel}
\usage{
\method{predict}{blblm_parallel}(object, new_data, confidence = FALSE, level = 0.95, cluster = 1, ...)
}
\arguments{
\item{object}{a blblm_parallel object}

\item{new_data}{a data frame}

\item{confidence}{boolean specify if you need confidence interval}

\item{level}{confidence level}

\item{cluster}{integer > 1 will uses future for parallel computing}

\item{...}{others}
}
\description{
prediction using blblm models and parallel computing
}
\examples{
# proj1 = blblm::blblm_parallel(disp~mpg, mtcars, m = 3, B = 100, cluster = 4)
# predict(proj1, data.frame('mpg' = c(24,21)), confidence = TRUE, cluster = 4)

}
