% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{predict.blblog}
\alias{predict.blblog}
\title{Obtain the predictions/prediction intervals}
\usage{
\method{predict}{blblog}(
  object,
  new_data,
  confidence = FALSE,
  level = 0.95,
  par = FALSE,
  type = "link",
  ...
)
}
\arguments{
\item{object}{object of class "blblog"}

\item{new_data}{data frame indicating the new values to be used for prediction}

\item{confidence}{a logical value, FALSE by default, indicating whether or not user wants the prediction interval in addition to the predicted value}

\item{level}{numeric specify the level of the prediction interval}

\item{par}{logical. Indicate whether or not user wants parallelization to be run. FALSE by default.}

\item{type}{character indicating whether user wants "link", or "response"}

\item{...}{optional arguments to be passed}
}
\value{
numeric value indicating the value of the model (if type = "link"), or the probability of a response happening(if type="response")
}
\description{
We use this function \code{predict} in order to find coefficients for all variables involved. This method works best for binary classification
}
\details{
\code{predict} gives two different values for logistic regression, depending on the \code{type} used. If "response" is given,
then \code{predict} gives the probability that the predicted value belongs to either class 1 or 2. Using "link" gives instead the
predicted logit values.
}
\examples{
newiris = iris[c(1:100), ]
fit = blblog(Species ~ Sepal.Length, data = newiris, m = 2, B = 2)
newdata = data.frame(Sepal.Length=c(5.1))
predict(fit, newdata, type = "link")
fit = blblog(Species ~ Sepal.Length, data = iris, m = 2, B = 2)
newdata = data.frame(Sepal.Length=c(5.1))
predict(fit, newdata, type = "response")
}
