% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblogreg}
\alias{predict.blblogreg}
\title{blblogreg model prediction}
\usage{
\method{predict}{blblogreg}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{model fit produced by blblogreg(formula, data, ...)}

\item{new_data}{data frame with same column dimensions as in object.}

\item{confidence}{logical, TRUE gives a confidence interval, FALSE gives a single estimate.}

\item{level}{numeric, confidence interval level (only if confidence = TRUE).}

\item{...}{other arguments.}
}
\description{
Makes a prediction (of probability of response = 1) using the fitted logistic model (object) on new data with similar dimensions
}
