% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbrf.R
\name{predict.blbrf}
\alias{predict.blbrf}
\title{predict.blbrf}
\usage{
\method{predict}{blbrf}(object, new_data, type = "prediction", level = 0.95, ...)
}
\arguments{
\item{object}{blbrf}

\item{new_data}{dataframe of new data entries}

\item{level}{double level of confidence interval}

\item{confidence}{boolean return confidence interval}
}
\description{
predict.blbrf
}
