% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predicts.blb}
\alias{predicts.blb}
\title{predicts.blb}
\usage{
predicts.blb(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{object}

\item{new_data}{numeric}

\item{confidence}{boolean}

\item{level}{numeric}

\item{...}{numeric}
}
\value{

}
\description{
makes prediction based off bootstrap coefficients
}
