% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{prepare_files}
\alias{prepare_files}
\title{Prepare the subfiles, select the specific files}
\usage{
prepare_files(data, m, select_names)

prepare_files(data, m, select_names)
}
\arguments{
\item{data}{dataset for the model}

\item{m}{number of subsets}

\item{select_names}{character list for file names}
}
\description{
Prepare the subfiles, select the specific files

Prepare the subfiles, select the specific files
}
