% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Print}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{the print object}

\item{...}{additional parameters related to print.lm}
}
\description{
Print fitted object model
}
