% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blbglm_parallel}
\alias{print.blbglm_parallel}
\title{print a blbglm_parallel project}
\usage{
\method{print}{blbglm_parallel}(x, ...)
}
\arguments{
\item{x}{a blbglm_parallel project}

\item{...}{other arguments}
}
\description{
print a blbglm_parallel project
}
\examples{
#proj = blblm::blbglm_parallel(vs~mpg+disp, mtcars, m = 3, B = 100, cluster = 2)
#print(proj)
}
