% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blbgm}
\alias{print.blbgm}
\title{Get the expression of formula}
\usage{
\method{print}{blbgm}(x, ...)
}
\arguments{
\item{x}{linear regression formula}

\item{...}{further arguments passed from or to other methods}
}
\description{
Get the expression of formula
}
