% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{the print object}

\item{...}{additional parameters related to print.lm}
}
\description{
Print fitted object model
}
