% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm_parallel}
\alias{print.blblm_parallel}
\title{print method of blblm_parallel}
\usage{
\method{print}{blblm_parallel}(x, ...)
}
\arguments{
\item{x}{a blblm_parallel fitted class}

\item{...}{others}
}
\description{
print blblm_parallel
}
\examples{
# proj = blblm::blblm_parallel(disp~mpg, mtcars, m = 3, B = 100, cluster = 4)
# print(proj)
}
