% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{print.blblog}
\alias{print.blblog}
\title{Prints the Logistic Model}
\usage{
\method{print}{blblog}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments affecting the result}
}
\value{
formula of blblog
}
\description{
Prints the Logistic Model
}
