% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblogreg}
\alias{print.blblogreg}
\title{Print blblogreg}
\usage{
\method{print}{blblogreg}(x, ...)
}
\arguments{
\item{x}{model fit produced by blblogreg(formula, data, ...)}

\item{...}{other arguments.}
}
\description{
Prints the formula from the blblogreg function
}
