% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blglm}
\alias{print.blglm}
\title{prints blglm model}
\usage{
\method{print}{blglm}(x, ...)
}
\arguments{
\item{x}{the item to be printed}

\item{...}{additional parameters to be passed in}
}
\description{
prints blglm model
}
