% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbrf.R
\name{rf_each_subsample}
\alias{rf_each_subsample}
\title{compute the estimates}
\usage{
rf_each_subsample(formula, data, B, nthreads)
}
\arguments{
\item{formula}{formula}

\item{data}{split dataset}

\item{B}{integer number of repetitions}

\item{nthreads}{integer number of workers}

\item{n}{integer length of dataset}
}
\description{
compute the estimates
}
