% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{Rsq.blblm}
\alias{Rsq.blblm}
\title{Get R^2}
\usage{
Rsq.blblm(object, confidence = FALSE, level = 0.95)
}
\arguments{
\item{object}{A variable with the class "blblm"}

\item{confidence}{Boolean for whether a confidence interval should be reported}

\item{level}{The significance level for a confidence interval (1 - alpha)}
}
\value{
An estimate for R^2 (or a confidence interval vector for R^2)
}
\description{
Get an estimate for the coefficient of determination, with the option of a confidence interval
}
