% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-settings.R
\name{set_workers}
\alias{set_workers}
\title{Set the paralle options.}
\usage{
set_workers(n = 2L)
}
\arguments{
\item{n}{Number of parallel workers.}
}
\description{
Set the paralle options.
}
\examples{

\dontrun{
# require(furrr)
set_workers(4)
}

}
