% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{Sigma of Bag of Little Bootstraps}
\usage{
\method{sigma}{blbglm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the fitted object}

\item{confidence}{boolean request for confidence interval}

\item{level}{confidence level}

\item{...}{additional parameters related to sigma.lm}
}
\description{
The mean sigma for the BLB
}
