% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blbgm}
\alias{sigma.blbgm}
\title{Get the confidence interval of parament}
\usage{
\method{sigma}{blbgm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{estimated object}

\item{confidence}{Whether return the confidence interval of sigma}

\item{level}{alpha value}

\item{...}{further arguments passed from or to other methods}
}
\description{
Get the confidence interval of parament
}
