% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Calculate the sigma confidence level}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{\code{blblm} object}

\item{confidence}{Boolean, TRUE OR FALSE.}

\item{level}{Confidence level, default 0.95.}

\item{...}{}
}
\description{
Calculate the sigma confidence level
}
