% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm_parallel}
\alias{sigma.blblm_parallel}
\title{MSE of blblm_parallel object}
\usage{
\method{sigma}{blblm_parallel}(object, confidence = FALSE, level = 0.95, cluster = 1, ...)
}
\arguments{
\item{object}{a fitted blblm_parallel class}

\item{confidence}{bool indicating if you want a confidence interval of sigma}

\item{level}{confidence level}

\item{cluster}{number of clusters to use for parallel computing. when = 1 will be running on single thread.}

\item{...}{others}
}
\description{
coefficients method of blblm_parallel class returning coefficients
}
\examples{
# proj = blblm::blblm_parallel(disp~mpg, mtcars, m = 3, B = 100, cluster = 4) # not run
# sigma(proj, confidence = TRUE, level = 0.99, cluster = 1) # not run
}
