% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{sigma.blblog}
\alias{sigma.blblog}
\title{Computing Sigma}
\usage{
\method{sigma}{blblog}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{confidence}{logical value indicating whether we construct the confidence interval}

\item{level}{integer between 0 and 1 indicating the confidence level}

\item{...}{additional arguments affecting the result}
}
\value{
numeric
}
\description{
Computes the standard deviation/error of Blblm using Polymorphism
}
