% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_data}
\alias{split_data}
\title{Split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{A data.frame or \link[tibble:tibble-package]{tibble} to be splited}

\item{m}{The number of subsamples.Integer.}
}
\value{
a list of subsamples \link[tibble:tibble-package]{tibble}
}
\description{
The subsamples may not in the same size.
}
\examples{
split_data(mtcars, 3)
}
