% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{split_data_glm}
\alias{split_data_glm}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data_glm(data, m)
}
\arguments{
\item{data}{data frame}

\item{m}{numeric, split data into m parts of approximated equal sizes}
}
\description{
split data into m parts of approximated equal sizes
}
