% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_sample}
\alias{split_sample}
\title{split data into m parts of approximated equal sizes}
\usage{
split_sample(data, m, min_subsample_size, even_split)
}
\description{
split data into m parts of approximated equal sizes
}
