% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Coefficients for blblm model}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{A fit blblm Model}
}
\description{
calls coef function to compute coefficients
}
\details{
compute the coefficients from fit
}
