% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Extract the coefficients from a model fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{The model fit to be extracted from.}
}
\description{
Extract the coefficients from a model fit
}
