% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Compute the coefficients from fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{lm, glm}
}
\value{
numeric coefficent estimates
}
\description{
Compute the coefficients from fit
}
