% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Calculate Coefficients}
\usage{
blbcoef(object)

blbcoef(object)
}
\arguments{
\item{object}{fitted model}
}
\value{
coefficients for the fit
compute the coefficients from fit

coefficients for the fit
compute the coefficients from fit
}
\description{
Calculate coefficients for input fitted model

Calculate coefficients for input fitted model
}
