% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{BLB Coefficients}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{fitted blblm model}
}
\value{
vector
}
\description{
Computes the coefficients from model fit
}
\examples{
blbcoef(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
}
