% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{bag of little boostraps coefficient value
compute the coefficients from fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{fitted values, use to get residuals/coefficients}
}
\description{
bag of little boostraps coefficient value
compute the coefficients from fit
}
