% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{compute the coefficients from fit}
\usage{
blbcoef(fit, formula)
}
\arguments{
\item{fit}{the fitted model returned from blblm}

\item{formula}{the formula used to fit the model}
}
\description{
compute the coefficients from fit
}
