% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R, R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Compute the coefficients from fit
Obtain the coefficients of the parameters of fit}
\usage{
blbcoef(fit)

blbcoef(fit)
}
\arguments{
\item{fit}{a fitted blblm model}
}
\value{
numeric value, return the value of the coefficients of the parameters

numeric value, return the value of the coefficients of the parameters
}
\description{
Compute the coefficients from fit
Obtain the coefficients of the parameters of fit

Compute the coefficients from fit
Obtain the coefficients of the parameters of fit
}
