% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblogreg.R
\name{blbcoef}
\alias{blbcoef}
\title{coefficient of blb}
\usage{
blbcoef(fit)

blbcoef(fit)
}
\arguments{
\item{fit}{fit}
}
\value{
coefficient of the fit
compute the coefficients from fit

coefficient of the fit
compute the coefficients from fit
}
\description{
coefficient of blb
}
