% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{coefficient of blblm}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{the linear regression function
compute the coefficients from fit}
}
\description{
compute the coefficients from fit
}
