% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{blbfcoef}
\usage{
blbcoef(fit, formula)
}
\arguments{
\item{fit}{individual lm model fit with cpp armadillo function}

\item{formula}{formula to fit the bootstraped model}
}
\value{
coef
}
\description{
blbfcoef
}
