% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{compute the coefficients from fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{the model returned from the main blblm function}
}
\description{
compute the coefficients from fit
}
