% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Computing Coefficients}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{a list containing information about the fitted model.}
}
\value{
coefficients of the model.
}
\description{
Compute the coefficients from fit.
}
