% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Blblm Coefficients
compute the coefficients from fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{Blblm object}
}
\description{
Blblm Coefficients
compute the coefficients from fit
}
