% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{get coefficient}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{regression model}
}
\value{
vector of coefficient
compute the coefficients from fit
}
\description{
compute the coefficients from fit
}
