% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{compute the coefficients from fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{in the parameters}
}
\value{
model with cluster
}
\description{
compute the coefficients from fit
}
