% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Coefficients (C++)}
\usage{
blbcoef(object, formula)
}
\arguments{
\item{object}{Fitted linear regression model}

\item{formula}{Linear regression formula}
}
\value{
Model coefficients
}
\description{
Computes coefficients from fit with lmC
}
