% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Grabs coefficients}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{The fitted linear regression model from lm().}
}
\description{
Grabs the estimated regression coefficients from the fitted linear regression model.
}
