% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblogreg.R
\name{blbcoef}
\alias{blbcoef}
\title{Bag of Little Bootstraps Coefficients}
\usage{
blbcoef(fit)

blbcoef(fit)
}
\arguments{
\item{fit}{Fitted logistic regression model.}
}
\value{
Coefficients from the given fit.

compute the coefficients from fit

Coefficients from the given fit.

compute the coefficients from fit
}
\description{
Extracts the coefficients of a given linear model.

Extracts the coefficients of a given logistic regression model.
}
