% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{compute the coefficients from fit}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{a model, generally linear}
}
\value{
the coefficient of the relationship between the variables in the
original fit model
}
\description{
compute the coefficients from fit
}
\examples{
fit = blblm(Sepal.Width ~ Sepal.Length, data = iris, B = 10)
blbcoef(fit)
}
