% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{Extract Model Coefficients for blb}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{an object for which extracting blb coefficients is meaningful.}
}
\description{
Extract model coefficients from blb object
}
