% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{compute the coefficients from fit}
\usage{
blbcoef(fit)

blbcoef(fit)
}
\arguments{
\item{fit}{the result of lm()}
}
\value{
the coef of fit

the coef of fit
}
\description{
compute the coefficients from fit

compute the coefficients from fit
}
