% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{blbglm}
\alias{blbglm}
\title{Logistic Regression with Little Bag of Bootstraps with the Option to Use Parallel Programming}
\usage{
blbglm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{a formula}

\item{data}{a dataframe}

\item{m}{numeric}

\item{B}{numeric}

\item{parallel}{logical}
}
\value{
blbglm
}
\description{
This function also has the option to use parallel programming so if you are using parallel programming, need to specify your workers first. For example, plan(multisession, workers = 4).
}
