% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{Bag of Little Bootstraps General Linear Model}
\usage{
blbglm(
  formula,
  family = gaussian(),
  data = NULL,
  filepaths = NULL,
  read_function = read.csv,
  m = 10,
  B = 5000,
  min_subsample_size = NULL,
  even_split = NULL,
  use_plan = TRUE,
  ...
)
}
\arguments{
\item{formula}{The formula to use for the linear models.}

\item{family}{The family of the linear models (see glm()). Defaults to
gaussian() for linear regression.}

\item{data}{An object containing data (which can be subsetted and passed to
glm()).}

\item{filepaths}{A vector or list of filepaths to read from.}

\item{read_function}{The function to use to read files.}

\item{m}{The number of subsamples to create.}

\item{B}{The number of bootstraps for each subsample.}

\item{min_subsample_size}{The minimum size of each subsample. For small
numbers of observations per subsample, mitigates weird behaviors from
fitting glm() on small data sets.}

\item{even_split}{Whether to split subsamples so that they are as equally
sized as possible.}

\item{use_plan}{Whether to use the plan set by future::plan(). use_plan =
FALSE is faster for sequential processing.}

\item{...}{Additional arguments to pass to read_function.}
}
\value{
A blbglm object, containing the estimates from each bootstrap, the
formula, and the family.
}
\description{
This function implements the bag of little bootstraps for the general linear
model.
}
\details{
Linear and logistic regression are properly supported, but if your
data is appropriately formed, there's no reason you can't use other general
linear models with blbglm(). (Corrections to predictions, where necessary,
are only made for the logistic case.)

Parallelized processing is supported through future::plan(). blbglm()
defaults to using future_map(), respecting any plan() the user has set.
However, disabling the use of future_map() for sequential processing is
usually faster; set use_plan = FALSE for this.

blbglm() can also work with filepaths in lieu of a data object. Provide a
vector of filepaths and blbglm() will read them for you. (Filepath reading
is faster when blbglm() is running in a parallelized mode.) In this case,
each file is considered to be one subsample (and m is ignored).

By default, blbglm() assigns an equal number of observations to each
subsample (ties are broken in favor of earlier numbered subsamples). It is
possible to disable this behavior so that blbglm() assigns each observation
to a subsample at random, without regard for the number of observation in
each subsample. The minimum number of observations in each sample is
controlled by min_subsample_size, but it should be at least 2 + the number
of dependent variables.
}
\examples{
blbglm(mpg ~ wt, data = mtcars, B = 100, even_split = TRUE, use_plan = FALSE)
blbglm(Species ~ Sepal.Length,
  family = binomial(), data = iris,
  m = 3, B = 100, min_subsample_size = 30
)
}
