% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Fit a general linear regression model}
\usage{
blbglm(formula, family = gaussian, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{data}{which is usual is the data frame, which is the data we imput.}

\item{m}{which is numeric variables, indicates the number of splits we need}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}
}
\value{
blblm object
}
\description{
Fit the linear regression model with data frame input
}
\examples{
# fit<- blbglm(Species~., data = read.csv("iris.csv"), m = 10, B = 5000)
}
\seealso{
Other family which indicates which kinds of glm we use: 
\code{\link{blbglm_file_par}()},
\code{\link{blbglm_file}()},
\code{\link{blbglm_par}()}
}
\concept{family which indicates which kinds of glm we use}
