% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{Generalized Linear Model with BLB}
\usage{
blbglm(formula, data, m = 10, B = 5000, family)
}
\arguments{
\item{formula}{regression formula}

\item{data}{a list of files, or a vector}

\item{m}{prefered size of data seperation}

\item{B}{number of bootstraps}

\item{family}{the glm family}
}
\description{
Generalized Linear Model with BLB
}
