% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Bag of Little Bootstraps for Binomial Logistic Regression}
\usage{
blbglm(formula, data, m = 10, B = 5000, nthreads = 1)
}
\arguments{
\item{formula}{An object of class "formula".}

\item{data}{A data frame containing the variables of the model.}

\item{m}{An integer denoting the number of splits}

\item{B}{an integer denoting the number of bootstrap samples per split}

\item{nthreads}{an integer denoting number of workers}
}
\value{
blbglm
}
\description{
Implementation of the Bag of Little Bootstrap algorithm for binomial Logistic
Regression. Includes methods blb estimates for common glm statistics.
}
\examples{
blbglm(Species ~ ., data = iris, m = 2, B = 5000, nthreads = 8)
}
