% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{blbglm}
\usage{
blbglm(formula, family, data, times)
}
\arguments{
\item{formula}{formula}

\item{family}{family}

\item{data}{data.frame}

\item{times}{numeric}
}
\value{
list
}
\description{
computes bootstraps for glm
}
\details{
Linear Regression with Little Bag of Bootstraps
}
