% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{blbglm}
\alias{blbglm}
\title{Bag of Little Bootstraps for Logistic Regression
Use Bag of Little Bootstraps to do Logistic Regression}
\usage{
blbglm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame or a character vector}

\item{m}{an integer giving the number of groups to split the data in}

\item{B}{an integer that represents the number of bootstraps}

\item{parallel}{logical; if TRUE, run plan(multiprocess, workers = 4) before running the blbglm function. In this example, workers = 4 but you can change the value of workers to be the number of workers you want.}
}
\value{
blbglm
}
\description{
Bag of Little Bootstraps for Logistic Regression
Use Bag of Little Bootstraps to do Logistic Regression
}
