% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Bag of Little Bootstraps Logistic Regression Algorithm}
\usage{
blbglm(formula, data, m = 2, B = 100, parallel = FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{m}{number of subsets for the data}

\item{B}{number of bootstraps}

\item{parallel}{TRUE/FALSE for parallel computing}
}
\value{
Returns the bag of little bootstrap confidence intervals for sigma for each boot of the logistic model
}
\description{
To run calculations parallelly for faster speed, first detect num of cores by running the following command:
\code{future::availableCores()}
Then specify number of cores you want use in the following line:
\verb{plan(multisession, worker = #_cores)}
}
