% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Bag of Little Bootstraps for Generalized Linear Models}
\usage{
blbglm(formula, data, family = "gaussian", m = 10, B = 5000)
}
\arguments{
\item{formula}{the formula to be modelled}

\item{data}{the source data}

\item{family}{the model family}

\item{m}{the number of subsamples}

\item{B}{the number of bootstraps performed per subsample}
}
\description{
The  primary namesake function subsets the data into m sections,
then bootstraps each subsample B times.
}
