% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Bag of Little Bootstrap Logistic Regression}
\usage{
blbglm(formula, data, m = 5, B = 1000, parallel = FALSE)
}
\arguments{
\item{formula}{the regression model you would like to create.}

\item{data}{the data you would like to analyze.}

\item{m}{the number of sub-samples you want to divide your data into.}

\item{B}{number of bootstraps performed on each sub-sample.}

\item{parallel}{indicates if you would like to utilize multiple cores to run this function.}
}
\value{
blbglm returns a list of B number of estimated coefficients for the regression, B number of standard deviations of the errors, and the formula used to fit the model.
}
\description{
Bag of Little Bootstraps is used to fit a logistic regression model.
If parallel = TRUE, prior to the use of this function, users will need to run the function "plan" from the furrr package which specifies the number of workers (cores) being utilized for parallelization.
}
