% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Fitting Logistic Linear Models with Bag of Little Bootstrap}
\usage{
blbglm(formula, data, family, m = 10, B = 5000, nthreads = 1)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted (for example, y ~ x).}

\item{data}{a data frame, list or environment containing variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{m}{an integer specifying the number of subsamples to be used for the bag of little bootstrap process. By default, \code{m = 10}.}

\item{B}{an integer specifying the number of bootstraps for each subsample. By default, \code{B = 5000}.}

\item{nthreads}{an integer which indicates the number of CPUs to be used for parallelization. By default, \code{nthreds = 1}.}
}
\value{
\code{blbglm} returns an object of class "blbglm".
}
\description{
\code{blbglm} is used to fit logistic linear models with bag of little bootstrap.
It can be used to carry out regression, analysis of variance, coefficients, confident interval,
and to compute prediction based on the regression results.
}
\examples{
blbglm(Species ~ Sepal.Length * Sepal.Width, data = iris, family = binomial(), m = 3, B = 100)
# using parallelization
blbglm(Species ~ Sepal.Length * Sepal.Width, data = iris, family = binomial(), 3, 100, nthreads = 1)
}
