% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{Build generalized linear model}
\usage{
blbglm(formula, data, family = gaussian(), m = 10, B = 5000)
}
\arguments{
\item{formula}{linear model}

\item{data}{data.frame/string vector}

\item{family}{string type of glm}

\item{m}{integer}

\item{B}{integer}
}
\description{
Set up glm using blb method, default family is gaussian
}
