% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{bag of little bootstrap for logistic regression model}
\usage{
blbglm(formula, data, m = 10, B = 5000, family, Parallel = FALSE)
}
\arguments{
\item{formula}{formula to use in blbglm}

\item{data}{data}

\item{m}{number of splits}

\item{B}{number of boostraps}

\item{family}{glm family to specify}

\item{Parallel}{boolean value to specify whether to use parallelization}
}
\value{
blbglm object
}
\description{
bag of little bootstrap for logistic regression model
}
\examples{
blbglm(Species ~ Sepal.Length * Sepal.Width, iris[1:100,], 3, 100, family = binomial)
blbglm(Species ~ Sepal.Length * Sepal.Width, iris[1:100,], 3, 100, binomial, TRUE)
}
