% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Fit generalized linear models using bootstrap}
\usage{
blbglm(formula, data, m = 10, B = 200, family = gaussian)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe, list}

\item{m}{numeric}

\item{B}{numeric}

\item{family}{family}
}
\value{
blbglm
}
\description{
Fit generalized linear models using bootstrap
}
\examples{
fit <- blbglm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, family = gaussian)
}
