% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{BLBGLM}
\usage{
blbglm(formula, data, m = 3, B = 100, seed_value = 100)
}
\arguments{
\item{formula, }{the formula form of the desired Logistic regression.}

\item{data, }{the data to pull the formula variables from.}

\item{m, }{the number of subsets of "little bootstraps" to run, defaults to \code{10}.}

\item{B, }{the number of times to bootstrap the data, defaults to \code{5000}.}

\item{seed_value, }{an initial seed value, defaults to \code{100}, for reproduceability.}
}
\description{
BLBGLM
}
\examples{
blbglm(am ~ wt * hp, data = mtcars, m = 3, B = 100)
}
