% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{The Bag of Little Bootstraps for Logistic Linear Regression Model}
\usage{
blbglm(formula, data, parallel = TRUE, m = 10, B = 5000)
}
\arguments{
\item{formula}{the generalized linear regression formula}

\item{data}{data frame}

\item{parallel}{whether users want to use parallelization to use more than one CPUs. If you choose to use parallization, run plan(multisession, worker=i) first, where i is the number of CPUs you are willing to use.}

\item{m}{numeric, split data into m parts of approximated equal sizes}

\item{B}{numeric, repeat the process B times}
}
\description{
The Bag of Little Bootstraps for Logistic Linear Regression Model
}
\examples{
fitaaaa <- blbglm(formula=Sepal.Length~Sepal.Width+Petal.Length,data=iris,m=10,B=100,parallel=TRUE)
summary(fitaaaa)
}
