% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{Generalized linear models with Little Bag of Bootstraps}
\usage{
blbglm(formula, family = gaussian, file, m = 10, B = 5000, cl = 1)
}
\arguments{
\item{formula}{formula}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{file}{dataframe or a list of file of datasets}

\item{m}{number of parts data split into}

\item{B}{number of bootstraps}

\item{cl}{number of clusters to use}
}
\value{
list
}
\description{
Generalized linear models with Little Bag of Bootstraps
}
