% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{blbglm}
\alias{blbglm}
\title{Bag of little bootsraps for generalized linear model.}
\usage{
blbglm(
  formula,
  data,
  m = 10,
  B = 5000,
  parallel = FALSE,
  select_names = NULL,
  family
)
}
\arguments{
\item{formula}{a formula fitting a linear regression model.}

\item{data}{dataset for the model}

\item{m}{number of subsets}

\item{B}{number of simulations}

\item{parallel}{a logical operator, T for parallelization}

\item{select_names}{character list for file names}

\item{family}{family type for models, such as binomial, gaussian, gamma etc.}
}
\value{
.
}
\description{
Bag of little bootsraps for generalized linear model.
}
