% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Linear Regression with a Bag of Little Bootstraps}
\description{
Functions related to using little bootstraps to make linear models. Support for parallelism and CSV file reading is included.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
