% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of Little Bootstraps}
\description{
The data is first separated into m subsamples. Then for each subsample,
    we resample until we have the size of the original data r times, find the bootstap
    statistic(linear regression in this case), and compute the desired statistic(linear regression
    coefficients and residual standard deviation) from the bootstrap statistics. Then the average of 
    the m statistics are averaged.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
