% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblogreg.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Fitting Linear Models by Applying Bag of Little Bootstrap}
\description{
This package performs bag of little bootstrap (blb) on lm function. In other words, it is a blb version of lm which improves the accuracy and the efficiency of assessing the quality of estimators.

This package performs bag of little bootstrap (blb) on lm function. In other words, it is a blb version of lm which improves the accuracy and the efficiency of assessing the quality of estimators.
}
\details{
Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: Xiaoxuan Hu \email{xiahu@ucdavis.edu}

Authors:
\itemize{
  \item Randy Lai
}


\strong{Maintainer}: Xiaoxuan Hu \email{xiahu@ucdavis.edu}

Authors:
\itemize{
  \item Randy Lai
}

}
