% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\title{blblm: Bag of Little Bootstraps for Regression}
\description{
This package includes many of the functions featured in the lm function.
    What differs between the functions provided in this package and lm is that this package
    runs the lm function using bag of little bootstraps. The fucntions allows for single core
    and multi-core use and allows the users to pick the number of splits and bootstraps to be done
    on the data.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
