% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblm_logistic.R
\docType{package}
\name{blblm-package}
\title{blblm: A Bag of Little Bootstrap Package}
\description{
Use the bag of little bootstraps algorithm to compute the linear regression and logistic regression model.The user is also given the use parallel programming for each of these main functions. If the user wishes to use parallelization, then the user must specifiy `parallel = TRUE` in the arguments, otherwise it will use the default `parallel = FALSE`. The user must specify number of CPUs before using parallel programming. Please run `plan(multisession, worker = cl)` where cl is the number of CPUs before using the function with parallelization.

Use the bag of little bootstraps algorithm to compute the linear regression and logistic regression model.The user is also given the use parallel programming for each of these main functions. If the user wishes to use parallelization, then the user must specifiy `parallel = TRUE` in the arguments, otherwise it will use the default `parallel = FALSE`. The user must specify number of CPUs before using parallel programming. Please run `plan(multisession, worker = cl)` where cl is the number of CPUs before using the function with parallelization.
}
\details{
Linear Regression with Little Bag of Bootstraps with the Option to Use Parallel Programming

Logistic Regression with Little Bag of Bootstraps with the Option to Use Parallel Programming
}
