% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of Little Bootstraps for Linear Models}
\description{
Implements the bag of little bootstraps algorithm for linear regression models.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
