% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of Little Bootstraps for Generalized Linear Models}
\description{
Implements the bag of little bootstraps algorithm for the general linear model.
}
\details{
Generalized Linear Models with Bag of Little Bootstraps
}
