% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Calculate Bag of Little Bootstraps Model}
\description{
This package contains functions used to calculate linear regression
    model coefficients,coefficent standard errors, confidence intervals, 
    and predictions. This package utilizes bag of little bootstraps to 
    perform calculations, and is also capable of handling parallelization.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
