% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\title{blblm: Computes a Linear Regression Using the Bag of Little Bootstraps Algorithm}
\description{
This package uses the Bag of Little Bootstraps Algorithm to return a list of m lists, 
    each with B different parameter estimates (sigma and coef). Calling the methods sigma(), coef(), confint(), etc. will combine the
    results into a single value or term.
}
\author{
\strong{Maintainer}: Austin Way \email{arway@ucdavis.edu}

}
