% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blbglm_file.R, R/blbglm_file_par.R,
%   R/blbglm_par.R, R/blblm.R, R/blblm_fast.R, R/blblm_fast_file.R,
%   R/blblm_fast_file_par.R, R/blblm_fast_par.R, R/blblm_file.R,
%   R/blblm_file_par.R, R/blblm_par.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: The multiple methods of doing linear regression and general linear regression based on bag of little bootstraps}
\description{
The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.

The package can fit the linear regression and general linear model based on bag of little bootstraps; And the package can do the fast linear regression based on C++.
}
\details{
Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}


\strong{Maintainer}: Ruichen Xu \email{rucxu@ucdavis.edu}

}
