% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of Little Bootstraps for Linear Model}
\description{
This package applies bag of little bootstraps on linear model. It also does whatever the lm package does. For example, it has the function that helps to calculte coefficents of a linear regression model.

This package applies bag of little bootstraps on linear model. It also does whatever the lm package does. For example, it has the function that helps to calculte coefficents of a linear regression model.
}
\details{
Logistic Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
