% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blglml.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: bootstrap (Title Case)}
\description{
This pakage is going to bootstrap for linear regression
 and logistic regression
    Use four spaces when indenting paragraphs within the Description.

This pakage is going to bootstrap for linear regression
 and logistic regression
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
