% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Linear/Logstic Regression with Bootstrap and parallelization}
\description{
This package help users to compute linear/logstic regression models and parameters 
    using Little Bag of Bootstraps technique. Users also have the option to use parallization
    while computing the fitted model. To use multiple CPU's, users will need to run 
    `suppressWarnings(plan(multiprocess, workers = 4))` and change number of workers to specify
    number of CPU's want to use.

This package help users to compute linear/logstic regression models and parameters 
    using Little Bag of Bootstraps technique. Users also have the option to use parallization
    while computing the fitted model. To use multiple CPU's, users will need to run 
    `suppressWarnings(plan(multiprocess, workers = 4))` and change number of workers to specify
    number of CPU's want to use.
}
\details{
Logstic Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
