% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of Little Bootstraps}
\description{
Bag of Little Bootstraps linear model contains the functions requisit of estimating a statistic by using
    the bag of little bootstraps method and parallelization. the user can specify the number of cores desired in the blblm_par function. 
    the user can also pass a list of data files.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
