% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{Calculate some statistics via Linear Regression With a Bag of Little Bootstraps Implementation}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame, list, or environment(iterable)}

\item{m}{numeric which indicates the number of subsamples. Default is 10.}

\item{B}{numeric which is bootstrapped samples per subsample. Default is 5000.}

\item{par}{logical. By default, this is FALSE, meaning only one core is used. If user wants to use multiple cores, use par=TRUE, and before running this function, please run \code{plan(multiprocess, workers = w)}, where w is the number of cores on the user's computer.
If any warnings emerge and you wish to hide them, be sure to run:\code{suppressWarnings(plan(multiprocess, workers = w))} and \code{options(future.rng.onMisuse = "ignore")}}
}
\value{
bblm returns an object of class bblm
}
\description{
We use the bag of little bootstraps algorithm in order to calculate the statistics generated by weighted linear regression.
}
\details{
The object returned is of class bblm. Obtaining the coefficients and other estimates is similar to obtaining them
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, par = TRUE)
}
