% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Linear Regression with Little Bag of Bootstraps}
\description{
Linear Regression with Bag of Little Bootstraps, returns a coefficient estimator or confidence interval. 
    Bag of little bootstraps = subsample + bootstrap to yield a robust, computationally efficient means of assesing the quality of estimators.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
