% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag of Little Bootstraps}
\description{
The bag of little bootstraps lm package contains various functions needed to estimate a statistic using
    the bag of little bootstraps method. Additionally, this package allows the user to make use of lm
    formulas when finding statistics for a particular dataset. This package has been modified to include 
    a parallelization version allowing the user to specify the number of cores desired in the primary
    blblm function.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
