% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R, R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag of Little Bootstraps}
\description{
This package uses bag of little bootstraps to calculate estimators 
    such as the intercept and slope of a fitted model and confidence intervals.
    There is also an option to use bag of little bootstraps on a logistic 
    regression model.

This package uses bag of little bootstraps to calculate estimators 
    such as the intercept and slope of a fitted model and confidence intervals.
    There is also an option to use bag of little bootstraps on a logistic 
    regression model.
}
\details{
Logistic Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
