% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: What the Package Does (Title Case)}
\description{
This package helps to train the linear regression model for dataset with the method of bootstraps.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
