% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Linear Regression with Little Bag of Bootstraps}
\description{
Use little bag of bootstraps to do linear regression, including functions required while running blblm().
}
\details{
Linear Regression with Little Bag of Bootstraps
}
