% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of Little Bootstraps for Linear Regression Model}
\description{
The package implements The bag of little bootstraps for linear
    regression models for more than one CPU if the user chooses the
    package to do so. Then, it computes estimates from the linear
    regression model for its coefficients, its sigma, and its
    confidence interval.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
