% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\title{blblogreg}
\description{
This package calculates bootstraps of linear regression.
}
\details{
this is a package that calculate logistic regression with a little bag of bootstraps.
Logistic Regression with Little Bag of Bootstraps
}
