% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Bag of little bootstraps for linear regression model}
\description{
This package is able to take a dataset or a list of datasets and compute several results using little bag bootstraping. These result includes computations of best fit model, coefficient, sigma, upper and lower bound, and predictions.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
