% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bootstrapping on linear models}
\description{
Package uses bootstrapping to both estimate paramters and predict responses.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
