% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blbglm_cluster.R, R/blblm.R,
%   R/blblm_cluster.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Finding the linear regression and logistic regression to make people choose whether they want clusteroing}
\description{
Build 4 function to find out the linear and logistic regression model of the dataset. Two of them are providing the option to decide whether they want cluster or not.

Build 4 function to find out the linear and logistic regression model of the dataset. Two of them are providing the option to decide whether they want cluster or not.

Build 4 function to find out the linear and logistic regression model of the dataset. Two of them are providing the option to decide whether they want cluster or not.

Build 4 function to find out the linear and logistic regression model of the dataset. Two of them are providing the option to decide whether they want cluster or not.
}
\details{
Logistic Regression with Little Bag of Bootstraps

Logistic Regression with Cluster with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps

Linear Regression with cluster with Little Bag of Bootstraps
}
