% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag Of Little Bootstrap With Linear Regression and Confidence Interval}
\description{
This package perform bag of little bootstrap linear regression with the improvment 
             of parallelization and use of Rcpp in linear regression to speed up the functions,
             with the addtional functions to calculate weighted sigma, confidence interval, predict
             new data using the same file.
}
\details{
Linear Regression with Bag of Bootstraps
}
