% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag of Little Bootstraps Algorithm}
\description{
Bootstrapping allows a simple assessment of the quality of estimators, and blblm
    is an alternative procedure that incorporates both bootstrapping and subsampling methods. This package
    works well on large datasets because it is well-suited for parallel and distributed computing. The benefits
    of blblm are it is compuationally friendly, flexible, and maintains most of the statistical properties of 
    regular bootstrapping.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
