% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: "The Parallelized Way of Regression Analysis"}
\description{
"This package applies parallel computing to conduct Regression 
    Analysis on continuous datasets. The core functions build faster linear
    regression models and conduct confidence intervals to measure where the true
    mean datapoint lies between an interval."
}
\details{
Linear Regression with Little Bag of Bootstraps
}
