% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bootstrap Linear Regression with optional Parallel Processing}
\description{
Package allow users to perform bootstrap linear regression in a timely manner. This is acheived through the use of C functions to perform linear regression. Parallel Processing is also avaliable for use and the number of cores that are used can be specified as aparameter of the blblm function.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
