% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\title{blblm: Bag Of Little Bootstraps Linear Model}
\description{
Using the method of bag of little bootstraps to estimate the model and its confidence intervals. 
    Bootstrapping mimics the process of obtaining new data sets, so that we can estimate the variability of our estimate without generating additional samples.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
