% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\title{blblm: Bag Of Little Bootstraps Linear Model}
\description{
This package uses the method of Bag of Little Bootstraps to estimate the model,its confidence intervals, and its standard deviation. 
    Bootstrapping mimics the process of obtaining new data sets to approximate the variability of our estimates without needing to collect additional samples.
    Bag of Little Bootstraps combines the features of both sub-sampling and bootstrapping to produce a more robust and efficient way to assess the quality of estimators.
    Bag of Little Bootstrap is also ideal because it can be easily integrated into parallel processing, utilizing multiple cores to complete separate tasks to speed up runtime.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
