% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\title{blblm: Bag of Little Bootstraps for Regression Models}
\description{
Fit linear regression models and logistic regression models 
    using bag of little bootstraps.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
