% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Build Linear Model Using Bag of Little Bootstraps}
\description{
This package is used to build linear regression/ generalized linear regression models by using 
    the Bag of Little Bootstrap (blb) method to increase the accuracy of estimated parameters. Besides setting
    up the model, this package also provides an estimation of confidence interval on estimated parameters, 
    sigma, and predicted outcome.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
