% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\alias{Linear}
\alias{Regression}
\alias{with}
\alias{Little}
\alias{Bag}
\alias{of}
\alias{Bootstraps}
\title{blblm: Bag of little Bootstrap on regressions}
\description{
Using the method of Bag of little Bootstrap to calculate the regression. Include in this pacakge. A blblm function for blb on linear regression and a blbglm function on logistic regression.
}
