% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Linear Regression with Little Bag of Bootstraps}
\description{
This package is used for Linear Regression with Little Bag of Bootstraps.
    Users will be able to use four main functions blblm,blblm_par, blblm_glm, blblm_glm_par.
    When useres want to use parralle function, they need to first load library(parallel) and create cluster themselves.
    blblm and blblm_par serve to simply use linear regression model.
    blblm_glm and blblm_glm_par are general linear regression model with gaussian model.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
