% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Perform bag of little bootstrap for linear regression model}
\description{
BLB is an awesome algorithm to use to increase the precision and accurity of getting coefficients. Therefore, I'd like to rewrite lm function using the algorithm, to make the prediction more acurates as well as the coefficients and sigmas.Remember, eric qian is sb.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: Xuanbin Chen \email{jaunechen@qq.com}

}
