% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/logistic_regression.R
\docType{package}
\name{blblm-package}
\title{blblm: Computes Bag of Little Bootstraps For Linear and Logistic Regression}
\description{
Uses the bag of little bootstraps algorithm to compute the
    the linear regression and logistic regression model. It is more computationally efficient since
    it uses both bootstrap and subsampling. It splits the data, computes the
    estimates, computes the regression estimates for a bag of little bootstraps 
    dataset, along with computing the coefficient, sigma, and confidence intervals
    for the fit.
    Use four spaces when indenting paragraphs within the Description.

Uses the bag of little bootstraps algorithm to compute the
    the linear regression and logistic regression model. It is more computationally efficient since
    it uses both bootstrap and subsampling. It splits the data, computes the
    estimates, computes the regression estimates for a bag of little bootstraps 
    dataset, along with computing the coefficient, sigma, and confidence intervals
    for the fit.
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Linear Regression with Little Bag of Bootstraps

Logistic Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: Megan Goodson \email{magoodson@ucdavis.edu}


\strong{Maintainer}: Megan Goodson \email{magoodson@ucdavis.edu}

}
