% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag of Little Bootstraps for Linear Models}
\description{
The `blblm` package is used for statistical analysis of linear models. The functions in the package return the
    coefficients, sigma values, predicted values, and the confidence interval. The user can also enable parallelization and
    dynamic scheduling to increase performance on larger datasets. The functions included in the package are based off linear
    regression and least squares techniques.
}
\details{
Bag of Little Bootstraps for Linear Models
}
