% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Enabling Parallelization for Regression Analysis}
\description{
A package that applies parallelization when conducting any regression analysis. Provides quicker 
                usage time, along with the use of confidence intervals to pinpoint accurate data.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
