% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Utilizes Bag of Little Bootstraps Functionality for Multiple Model Estimation}
\description{
This package utilizes the bag of little bootstraps method in order to estimate different model parameters. The package will run the specified model fit both on subsets of the data as specified by m as well as bootstrapping a number of times as specified by B. It will then be able to provide not only coefficient and sigma values, but also confidence intervals for such values.
}
\details{
Linear and Logistic Regression models with Little Bag of Bootstraps
}
