% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{The Bag of Little Bootstraps for Linear Regression Model}
\description{
blblm is to compute the bag of little bootstraps for linear regression model.
    blblm can return the coefficient, sigma, confident interval and prodiction of the bootstraps for linear regression model, and also it can print out the model.
}
\details{
This package implement the linear regression with little bag of bootstraps by using blblm(formula,data,parallel,m,B)
In this package, there are also corresponding functions from blblm, such as \code{coef.blblm}, \code{sigma.blblm},\code{confint.blblm}and \code{predict.blblm}
}
\author{
\strong{Maintainer}: Xiaoqi Na \email{xna@ucdavis.edu}

}
