% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblr.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\alias{_PACKAGE}
\title{blblm: Linear Regression and Logistic Regression using Little Bag of Bootstraps}
\description{
This package has functions for running little bag of bootstraps on
    linear regression models and logistic regression model on datasets.
    For the linear regression model, users can determine the regression coefficients (beta heads),
    the estimate of variance of errors (sigma), confidence interval and prediction interval on new data.
    For logistic regression model, users can determine the regression coefficients (beta heads),
    as well as confidence interval on particular coefficient or prediction interval on coefficients.
    For both function inside the pakage, users can also decide if they want to use parallel computing 
    by changing the option in function.

This package has functions for running little bag of bootstraps on
    linear regression models and logistic regression model on datasets.
    For the linear regression model, users can determine the regression coefficients (beta heads),
    the estimate of variance of errors (sigma), confidence interval and prediction interval on new data.
    For logistic regression model, users can determine the regression coefficients (beta heads),
    as well as confidence interval on particular coefficient or prediction interval on coefficients.
    For both function inside the pakage, users can also decide if they want to use parallel computing 
    by changing the option in function.
}
\details{
Linear Regression with Little Bag of Bootstraps

_PACKAGE
Logistic Regression with Little Bag of Bootstraps
}
