% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm-package}
\alias{_PACKAGE}
\title{Using Little Bag of Bootstraps for Linear Regression}
\description{
This package do a linear regression with little bag of bootstraps.
}
