% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblog.R
\docType{package}
\name{blblm-package}
\title{blblm: Regression Model with Bag of Little Bootstrap}
\description{
This package allows to use bag of little bootstrap to run linear regression model and logistic regression model. The user can decide to use parallelization to run in several CPUs.

This package allows to use bag of little bootstrap to run linear regression model and logistic regression model. The user can decide to use parallelization to run in several CPUs.
}
\details{
Linear Regression with Little Bag of Bootstraps

Logistic Regression with Little Bag of Bootstraps
}
