% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{Bootstrap for linear regression model}
\description{
Bag of little bootstraps for linear regression model.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
