% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Linear Regression with Little Bag of Bootstraps}
\description{
Linear Regression with Little Bag of Bootstraps
    The users can do the linear regression with bootstraps and output the estimate coefficients and sigma.
    It can also use parallelization if the user choose to use blblm_pl.
    After generating the model, the users can use coef, sigma, confint, and predict to further explore the model.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
