% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm-package.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: Linear Regression with Little Bag of Bootstraps}
\description{
Linear Regression with Little Bag of Bootstraps,
    which is implemented in parallel.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: Ziqing Luo \email{zqluo@ucdavis.edu}

}
