% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag of little bootstraps algorithm}
\description{
Bag of little bootstraps algorithm for linear regression and 
    generalized linear models. This package includes improvement that allows users to 
    use parallelization and specify a list of file of datasets. For GLM, users can specify the familytype
    in order to choose the desired model.

Bag of little bootstraps algorithm for linear regression and 
    generalized linear models. This package includes improvement that allows users to 
    use parallelization and specify a list of file of datasets. For GLM, users can specify the familytype
    in order to choose the desired model.
}
\details{
Generalized Linear Models with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
