% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear modeling with a bag of little bootstraps}
\usage{
blblm(dat, formu, m = 10, B = 5000, useParallel = FALSE, useFreqs = TRUE)
}
\arguments{
\item{dat}{Data Frame (or equivalent)}

\item{formu}{Regression formula}

\item{m}{Number of subsamples}

\item{B}{Number of bootstraps per subsample}

\item{useParallel}{Boolean for using parallel functions (future::plan() must be set before using this)}

\item{useFreqs}{Boolean for generating frequencies for each row of data (used in weighted linear regression)}
}
\value{
A list with (1) elements separated by filename and bootstrap trial and (2) the model formula
(Note: The filenames are used as list names in the output).
Within each trial, the output includes the model coefficients, sigma, R^2, Adjusted R^2, and Weighted R^2 (if weights are used)
}
\description{
This will produce mutliple Linear Regression models by splitting the data and using little bootstraps.
}
\examples{
blblm(dat = mtcars, formu = mpg ~ wt * hp, m = 5, B = 91)
blblm(dat = mtcars, formu = mpg ~ wt * hp, m = 3, B = 100, useParallel = TRUE)
}
