% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstraps Version of Linear Regression}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be converted to that class).}

\item{data}{a data.frame, environment, or list.}

\item{m}{the number of splits on the data, m = 10 by default.}

\item{B}{the number of bootstraps, B = 5000 by default.}

\item{parallel}{boolean, specify whether to use parallelization or not, parallel = FALSE by default}
}
\value{
a blblm object
}
\description{
blblm performs bag of little bootstrap (blb) on lm function. If users set "parallel = TRUE" in the blblm() function, the plan() function is usually needed for users to specify in front of the blblm() function.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 10, B = 10000, parallel = TRUE)
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
