% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of little Bootstraps for lm}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{An object of class "formula". Can also be seen as a symbolic description of a model that is to be fit}

\item{data}{A dataframe containing the data used in formula.}

\item{m}{Represents the amount of times the data given is split.}

\item{B}{Represents the amount of times the data is run on bootstrap}

\item{parallel}{A boolean which denotes whether or not the user would like to use parallelization. If TRUE, blb will be run with
the maximum number of threads available. If FALSE, then a single thread will be used.}
}
\description{
This functions aims to improve the results produced by the lm function with the use of blb. The functions allows
for the user to set the number of splits they want in their data along with the number of times they want to bootstrap their data.
This function also allows for parallelization. NOTE: set up workers before running function if you're planning to use parallelization.
e.g plan(multiprocess, workers = 4).
}
\examples{
fit = blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
