% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{blblm function to compute Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{a formula}

\item{data}{a dataframe or a character vector}

\item{m}{an integer that is the number of subsets in the data}

\item{B}{an integer that is the number of bootstraps}

\item{parallel}{a logical character}
}
\value{
blblm
}
\description{
This function also has the option to use parallel programming so if you are using parallel programming, need to specify your workers first. For example, plan(multisession, workers = 4).
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
