% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Compute Linear Regression Model using Bag of Little Bootstraps algorithm.
The user can specify whether or not she would like to use parallelization.
Parallelization is done using future_map.
If a character vector of files is passed in, each file is read in the workers.
If an entire data frame is passed in, the data is split and it will either use future_map or regular map to calculate coefficents and sigma.}
\usage{
blblm(formula, data, parallel = FALSE, m = 10, B = 5000)
}
\arguments{
\item{formula}{Specify formula for model.}

\item{data}{character or data.frame Character vector listing files or whole dataset.}

\item{parallel}{logical TRUE if you specified TRUE in parallelization function, FALSE otherwise.}

\item{m}{numeric Number of subsamples to split the data if whole dataset is specified.}

\item{B}{numeric Number of bootstrap estimates.}
}
\value{
class blblm object Can extract coefficent, sigma estimates, along with formula.
}
\description{
Compute Linear Regression Model using Bag of Little Bootstraps algorithm.
The user can specify whether or not she would like to use parallelization.
Parallelization is done using future_map.
If a character vector of files is passed in, each file is read in the workers.
If an entire data frame is passed in, the data is split and it will either use future_map or regular map to calculate coefficents and sigma.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
