% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstraps Regression}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE, seed = NaN)
}
\arguments{
\item{formula}{Regression Formula}

\item{data}{Data Frame}

\item{m}{Integer (Number of Subsamples)}

\item{B}{Integer (Number of Replications)}

\item{parallel}{Boolean - If true, will use all available CPU cores}

\item{seed}{Optional seed, will allow for testing output.}
}
\description{
Bag of Little Bootstraps Regression
}
\details{
Uses Little Bag of Bootstraps to Compute Linear Regression with a specified number of subsamples and replications
}
