% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression BLB Original version}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression formula}

\item{data}{a list of files, or a vector}

\item{m}{prefered size of data seperation}

\item{B}{number of bootstraps}
}
\description{
Linear Regression BLB Original version
}
