% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstraps for fitting Linear Models}
\usage{
blblm(formula, data, m = 10, B = 5000, nthreads = 1)
}
\arguments{
\item{formula}{An object of class "formula".}

\item{data}{A data frame containing the variables of the model.}

\item{m}{An integer denoting the number of splits}

\item{B}{an integer denoting the number of bootstrap samples per split}

\item{nthreads}{an integer denoting number of workers}
}
\value{
blblm
}
\description{
Implementation of the Bag of Little Bootstrap algorithm for linear regression.
Includes methods blb estimates for common regression statistics.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 10000, nthreads = 8)
}
