% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bind two factors}
\usage{
blblm(formula, data, m = 10, B = 5000, nthreads = 1)
}
\arguments{
\item{formula}{object of class "formula"}

\item{data}{a dataframe}

\item{m}{an integer}

\item{B}{an integer}

\item{nthreads}{an integer}
}
\value{
blblm
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 10000, nthreads = 8)
}
