% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Builds blblm object}
\usage{
blblm(formula, data, m = 10, B = 5000, workers = 1, fastlm = TRUE)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe}

\item{m}{integer}

\item{B}{integer}

\item{workers}{integer}

\item{fastlm}{boolean}
}
\value{
object class blblm
}
\description{
Creates linear regression objects with little bag of bootstraps
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, fastlm = TRUE)
}
