% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear regression with Bag of Little Bootstraps (BLB)}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{an object of class "\link{formula}": a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model}

\item{m}{an integer specifying the number of rows (Default 10) the data will be sliced into.}

\item{B}{an integer specifying the number of bootstraps (Default 5000) done within each sub-sample.}
}
\value{
blblm returns an object of class "blblm".

Like object of class "lm", the generic accessor functions \code{coefficients}, \code{confint}, \code{sigma}, \code{predict} extract various useful features of the value returned by \code{blblm}.
}
\description{
Linear regression with Bag of Little Bootstraps (BLB)
}
\examples{
library(blblm)
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)

}
