% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear regression with Bag of Little Bootstraps (BLB)}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = TRUE)
}
\arguments{
\item{formula}{an object of class "\link{formula}": a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model}

\item{m}{an integer specifying the number of chunks (Default 10) the data will be sliced into.}

\item{B}{an integer specifying the number of bootstraps (Default 5000) done within each sub-sample.}

\item{parallel}{logical, use parallel (Default) or not.}
}
\value{
blblm returns an object of class "blblm".

Like object of class "lm", the generic accessor functions \code{coefficients}, \code{confint}, \code{sigma}, \code{predict} extract various useful features of the value returned by \code{blblm}.
}
\description{
Linear regression with Bag of Little Bootstraps (BLB)
}
\examples{
data(mtcars)
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
closeAllConnections()
}
