% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear regression with Bag of Little Bootstraps (BLB)}
\usage{
blblm(
  formula,
  data,
  m = 10L,
  B = 5000L,
  use.legacy = NA,
  parallel = availableCores()
)
}
\arguments{
\item{formula}{an object of class \link{formula}: a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model}

\item{m}{an integer specifying the number of chunks (Default 10) the data will be sliced into.}

\item{B}{an integer specifying the number of bootstraps (Default 5000) done within each sub-sample.}

\item{use.legacy}{an logical specifying whether use the legacy version of function (Default NA to use optimized one).}

\item{parallel}{an integer specifying number of parallel processes (Default \code{\link[future:re-exports]{future::availableCores()}}) used. Set parallel = 1L to turn off.}
}
\value{
blblm returns an object of class "blblm".

Like object of class "lm", the generic accessor functions \link{coef}, \link{confint}, \link{sigma}, \link{predict} extract various useful features of the value returned by \code{blblm}.
}
\description{
to be filled
}
\examples{
data(mtcars)
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
closeAllConnections() # close all connections to clean up
}
