% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Bag of Little Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{The linear regression model you want}

\item{data}{The data you want to perform blblm on}

\item{m}{= number of observations}

\item{B}{= times to perform boostrap on the data, more times = more accurate; but will take longer.}

\item{parallel}{set to \verb{ = TRUE} if you want to activate speed up blb if your system can handle it.}
}
\description{
Linear Regression with Bag of Little Bootstraps
}
\references{
https://link.springer.com/chapter/10.1007/978-3-030-01418-6_53
}
