% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{blblm}
\usage{
blblm(formula, data, times)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}

\item{times}{numeric}
}
\value{
list
}
\description{
computes bootstraps
}
\details{
Linear Regression with Little Bag of Bootstraps
}
