% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Calculate coefficeints using bag of little bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{the features of interest from the data}

\item{data}{the dataset of interest}

\item{m}{the number of subsamples of the data}

\item{B}{the number of bootstrap samples to take}
}
\value{
the fitted lm for the bag of little bootstraps
}
\description{
Calculate coefficeints using bag of little bootstraps
}
\examples{
blblm(hp~mpg,mtcars,10,100)
}
