% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m, B, Cluster)
}
\arguments{
\item{formula}{a formula}

\item{data}{dataframe}

\item{m}{numeric}

\item{B}{numeric}

\item{Cluster}{SOCKcluster}
}
\value{
linear regression
}
\description{
Linear Regression with Little Bag of Bootstraps
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
