% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm}
\alias{blblm}
\alias{_PACKAGE}
\alias{blblm-package}
\title{blblm}
\usage{
blblm(formula, data, m = 10, B = 5000, ncpu = 1)
}
\arguments{
\item{formula}{the formula of choice - linear regression with weight in this case.}

\item{data}{data of choice}

\item{m}{a positive number, the number of items to choose from.}

\item{B}{B a posive integer. It represents a number of bootstrap samples.}

\item{ncpu}{set as 1 in default which means the algorithm is implementing 1 CPU. The user can input any number of CPUs they would like to use.}
}
\value{
logistic regression model
}
\description{
This package calculates bootstraps of linear regression.
}
\details{
this is a package the calculate linera regression with a little bag of bootstraps.
Linear Regression with Little Bag of Bootstraps
}
