% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Retruns model fitted with bootstrap sampling}
\usage{
blblm(formula, data, m = 10, B = 5000, model = "lr", para = FALSE, core = 2)
}
\arguments{
\item{formula}{model specification of response and predictor variables.}

\item{data}{dataset to fit}

\item{m}{integer}

\item{B}{integer}

\item{model}{linear model}

\item{para}{logical indicating parallel computing}

\item{core}{integer number of cores for paralle}
}
\description{
Retruns model fitted with bootstrap sampling
}
