% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Blb lm}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{the interest user want from data}

\item{data}{the dataset}

\item{m}{numbers of subsamples created}

\item{B}{the number of bootstraps user want}
}
\description{
linear regression that creates subsample and bootstrap
}
