% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Parallelization of Linear Regression Model}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = TRUE)
}
\arguments{
\item{formula}{linear regression model}

\item{data}{dataset}

\item{m}{sample size}

\item{B}{random iterations}

\item{parallel}{chose to use parallel functions}
}
\description{
Parallelization of Linear Regression Model
}
