% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Bag of Little Bootstraps one or more cpu}
\usage{
blblm(formula, data, m = 10, B = 5000, n_cl = 1)
}
\arguments{
\item{formula}{Bootstraped linear regression model}

\item{data}{The dataset to perform blblm}

\item{m}{number of observations}

\item{B}{number of bootstraps}

\item{n_cl}{number of clusters for parallelization}
}
\description{
Linear Regression with Bag of Little Bootstraps one or more cpu
}
