% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm}
\alias{blblm}
\alias{blblm-package}
\title{Linear Modeling with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{the formula to be modelled}

\item{data}{the source data}

\item{m}{the number of subsamples}

\item{B}{the number of bootstraps performed per subsample}
}
\description{
The bag of little bootstraps (BLB) is a procedure 
    that takes the best of subsampling and bootstrapping to yield 
    a robust and efficient analytical tool. 
    BLB's primary use is for assessing the quality of estimators.

The  primary namesake function subsets the data into m sections,
then bootstraps each subsample B times.
}
\examples{
fit <- blblm(Sepal.Length ~ Petal.Width + Petal.Length, data = iris, m = 10, B = 15)

# An example with an interaction
blblm(Sepal.Length ~ Petal.Width + Petal.Length + Petal.Width * Petal.Length,
                          data = iris, m = 10, B = 15)

}
