% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{blblm}
\usage{
blblm(formula, data, m = 10, B = 5000, n_cores = 1)
}
\arguments{
\item{formula}{This function performs bootstraped lineear regression}

\item{data}{data is a data frame to perform lm}

\item{m}{number of subsets made from data}

\item{B}{number of bootstraps}

\item{n_cores}{specify number of cores if parallelization is used.}
}
\description{
Package allows users to perform bootstrap linear regression in a timely manner. This is acheived through the use of C functions to perform linear regression. Parallel Processing is also avaliable for use and the number of cores that are used can be specified as a parameter of the blblm function.
}
