% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstrap Linear Regression}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{the regression model you would like to create.}

\item{data}{the data you will use to pass into the function}

\item{m}{the number of parts you want your data to be split into.}

\item{B}{number of bootstraps on each m split}

\item{parallel}{perform function with parallel or not (default: FALSE)}
}
\value{
list of 2, B number of estimates and model fit
}
\description{
Create a linear regression model using bag of little bootstraps, specifying the number of parts data will be split and the number of bootstrap at each split.
requires users to set plan prior to calling function when parallel = TRUE
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)

}
