% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstrap Linear Regression}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{the regression model you would like to create.}

\item{data}{the data you would like to analyze.}

\item{m}{the number of sub-samples you want to divide your data into.}

\item{B}{number of bootstraps performed on each sub-sample.}

\item{parallel}{indicates if you would like to utilize multiple cores to run this function.}
}
\value{
blblm returns a list of B number of estimated coefficients for the regression, B number of standard deviations of the errors, and the formula used to fit the model.
}
\description{
Bag of Little Bootstraps is used to fit a linear regression model.
If parallel = TRUE, prior to the use of this function, users will need to run the function "plan" from the furrr package which specifies the number of workers (cores) being utilized for parallelization.
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)

}
