% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Fitting Linear Models with Bag of Little Bootstrap}
\usage{
blblm(formula, data, m = 10, B = 5000, nthreads = 1, fast_option = FALSE)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted (for example, y ~ x).}

\item{data}{a data frame, list or environment containing variables in the model.}

\item{m}{an integer specifying the number of subsamples to be used for the bag of little bootstrap process. By default, \code{m = 10}.}

\item{B}{an integer specifying the number of bootstraps for each subsample. By default, \code{B = 5000}.}

\item{nthreads}{an integer which indicates the number of CPUs to be used for parallelization. By default, \code{nthreds = 1}.}

\item{fast_option}{logical. \code{TRUE} if using Rcpp code in \code{lm1()}. By default, \code{fast_option = FALSE}.}
}
\value{
\code{blblm} returns an object of class "blblm".
}
\description{
\code{blblm} is used to fit linear models with bag of little bootstrap.
It can be used to carry out regression, analysis of variance, coefficients, confident interval,
and to compute prediction based on the regression results.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
blblm(mpg ~ wt * hp, data = mtcars, m = 10, B = 1000, nthreads = 4) # using parallelization
}
