% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Build Linear Regression Model}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{linear regression model}

\item{data}{data frame or vector of file names}

\item{m}{number of subsamples}

\item{B}{number of bootstraps}
}
\description{
Build the linear Regression model,
using the bag of little bootstrap method to find the coefficients
}
