% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Little Bag of Bootstraps with Linear Regression}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{The linear regression formula, in the form of y ~ x1 + x2 + ...}

\item{data}{Dataset you would like to analyze}

\item{m}{The number of subsamples you want the data to be split into, default is 10}

\item{B}{The number of Bootstrap trials you want to run on each subsample, default is 5000}
}
\value{
An object that contains the regression formula ($formula) and all of the estimates from each Bootstrap trial in a vector.
}
\description{
This function splits data up into m subsamples.
On each sample, B bootstrap trials are run to find estimates of the linear regression coefficients and standard deviation.
All those estimates, as well as the formula, are contained in the object returned.
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
