% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Generate linear regression with boostrap}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{m}{number of subdata}

\item{B}{number of bootstrips}
}
\value{
class blblm
}
\description{
create a linear regression with the method of bootstrips, with people selected feature
}
\examples{
x=blblm(mpg~wt,data=mtcars,m=3,B=100)
coef(x)
}
