% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{bag of little bootstrap for linear regression model}
\usage{
blblm(formula, data, m = 10, B = 5000, Parallel = FALSE)
}
\arguments{
\item{formula}{formula to use in blblm}

\item{data}{data}

\item{m}{number of splits}

\item{B}{number of boostraps}

\item{Parallel}{boolean value to specify whether to use parallelization}
}
\value{
blblm object
}
\description{
bag of little bootstrap for linear regression model
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, Parallel = FALSE)
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, Parallel = TRUE )
}
