% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Creating the blblm fucntion for Linear regression}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{in the model}

\item{data}{part of the paramterst}

\item{m}{a subsample number}

\item{B}{number for bootstrap}
}
\value{
model with cluster and blb techniques
}
\description{
This function computes the bootstrap sample without parallelization
}
\examples{
fit<-blblm(mpg~wt * hp, data = mtcars, m = 3, B = 100)
coef(fit)
confint(fit, c("wt", "hp"))
}
