% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Computes Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{linear regression formula}

\item{data}{dataframe}

\item{m}{an integer giving  the number of subsets for the data}

\item{B}{an integer giving the number of bootstraps}

\item{parallel}{logical value indicating TRUE or FALSE for parallelization}
}
\value{
list of estimates and formula
}
\description{
Give a formula, data, value for m, and value for B. The user should
run plan(multisession, workers = 4) if they want to use parallelization
for example, parallel = TRUE, and they can change the number of workers to
any numeric value they wish to use.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
