% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstraps Linear Model (single core)}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{Linear regression formula}

\item{data}{Chosen dataset}

\item{m}{Number of times to split data for bootstrapping}

\item{B}{Number of bootstrap samples}
}
\value{
Model coefficients
}
\description{
Computes coefficeints for a linear model using blb (single core)
}
