% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstraps Linear Model}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE, LB = FALSE, cl = NULL)
}
\arguments{
\item{formula}{Linear regression formula}

\item{data}{Chosen dataset}

\item{m}{Number of times to split data for bootstrapping}

\item{B}{Number of bootstrap samples}

\item{parallel}{Enables parallelization}

\item{LB}{Enables load balancing (Dynamic scheduling)}

\item{cl}{Number of cores (Made using `cl = makeCluster()`)}
}
\value{
Model coefficients
}
\description{
Calls the specified function to compute coefficeints for a linear model using blb, options for parallelization and load balancing
}
