% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression Coefficients}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{Formula for the linear regression.}

\item{data}{The dataframe to be used.}

\item{m}{The number of splits for the dataframe. m > 1}

\item{B}{The number of bootstraps to use. B > 1}
}
\description{
Calculates the linear regression coefficients of an input formula, data, number of splits m,
and number of bootstraps, B. Single-core program.
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
coef(fit)
}
