% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Bootstraps for Linear Regression}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = F, workers = 1)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{m}{an integer, the number of equal-size groups for bag of little bootstraps algorithm.}

\item{B}{an integer, the number of iterations over bag of little bootstraps algorithm.}

\item{parallel}{logical, indicates parallelization (TRUE) or not (FALSE).}

\item{workers}{an integer, the number of cores to use in bag of little bootstraps algorithm.}
}
\description{
blblm takes in arguments for traditional linear regression and three additional arguments,
then returns a fitted least squares model and confidence intervals corresponding to bootstrap method.
}
