% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{BLBLM}
\usage{
blblm(
  formula,
  data,
  m = 10,
  B = 5000,
  seed_value = 100,
  num_workers = detectCores(),
  parallelize = TRUE,
  fast = TRUE
)
}
\arguments{
\item{formula, }{the formula form of the desired linear regression.}

\item{data, }{the data to pull the formula variables from.}

\item{m, }{the number of subsets of "little bootstraps" to run, defaults to \code{10}.}

\item{B, }{the number of times to bootstrap the data, defaults to \code{5000}.}

\item{seed_value, }{an initial seed value, defaults to \code{100}, for reproduceability.}

\item{num_workers, }{the number of workers if parallelized, defaults to the number of computer cores. This value doesn't matter if parallelize is set to \code{FALSE}.}

\item{parallelize, }{whether or not to parallelize the function, defaults to \code{TRUE}.}

\item{fast, }{whether or not to use the faster rcpp version of the weighted linear regression function. Defaults to \code{TRUE}.}
}
\description{
BLBLM
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, num_workers = 4, parallelize = FALSE, fast = FALSE)
}
