% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{blblm is used to implement the bag of little bootstraps for the linear regression model.}
\usage{
blblm(
  formula,
  data,
  m = 10,
  B = 5000,
  parallel = FALSE,
  nthreads = detectCores()
)
}
\arguments{
\item{formula}{an object of class formula giving relationship of variables in model}

\item{data}{An optional data frame, list or environment containing variables for the linear model.}

\item{m}{An optional integer to specify the number of splits (m) in the data set.}

\item{B}{An optional integer for each subsample, sample observations with replacement, repeat for (B) times.}

\item{parallel}{logical. If TRUE the linear model with be fitted using parallel computing. Function will follow single process if FALSE.}

\item{nthreads}{n optional integer to specify the number of cores used only if parallel is TRUE.}
}
\description{
blblm is used to implement the bag of little bootstraps for the linear regression model.
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
fit2 <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = TRUE, nthreads = 4)


}
