% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{the main function to implement linear regression with little bag of bootstraps}
\usage{
blblm(
  formula,
  files = NULL,
  n = NULL,
  data = NULL,
  m = 10,
  B = 5000,
  method = "lm",
  cluster = NULL,
  seed = 141
)
}
\arguments{
\item{formula}{formula for linear regression}

\item{files}{allow user to inpute the path of files to do distribution, only support csv file}

\item{n}{if users input files, they should also input the size of whole data}

\item{data}{data for linear regression}

\item{m}{the number of subsamples divided by the whole data}

\item{B}{the time of little bag of bootstrap}

\item{method}{the method to compute coef and sigma, choice: "lm","lmR","lmC", default as "lm"}

\item{cluster}{a set of copies of R running in parallel, default as NULL only one core}

\item{seed}{set.seed to make the result reproducible, default as 141}
}
\value{
list including estimates of coefficients and sigma
in each time of little bag of bootsrap in each subsample
and the formula.
}
\description{
This function will devide the original data into several subsamples
and do little bag of bootsrap.
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
