% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{blblm}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE, threads = 4)
}
\arguments{
\item{formula}{linear regression formula}

\item{data}{dataset}

\item{m}{number of split, default = 10}

\item{B}{number of bootstrap, default = 5000}

\item{parallel}{estimation being calculated parallely, default = FALSE}

\item{threads}{number of workers using to run the estimation, default = 4}
}
\description{
blblm
}
