% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Little Bag of Bootstrap}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{The formula for linear regression (response ~ terms)}

\item{data}{The data use for linear regression}

\item{m}{Number of subsample (split data into m groups)}

\item{B}{Number of bootstrap}

\item{parallel}{Determine if using parallelization (logical)}
}
\value{
blblm: regression result
}
\description{
\code{blblm} is used to do a linear regression with
implementing little bag of bootstrap. The default setting is in \code{usage}
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = TRUE)
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
}
