% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{a linear formula}

\item{data}{a dataframe or dataset}

\item{m}{number of subsets}

\item{B}{bumber of bootstrap}
}
\value{
an invisible blblm object
}
\description{
Linear Regression with Little Bag of Bootstraps
}
