% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{"blblm()" is a function to model linear regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000, nthreads = 1)
}
\arguments{
\item{formula}{A symbolic description of the model}

\item{data}{dataframe}

\item{m}{integer,number of subsample you want}

\item{B}{integer, bootstrap iteration times}

\item{nthreads}{the number of cores you want to use when processing}
}
\value{
object
}
\description{
"blblm()" is a function to model linear regression with Little Bag of Bootstraps
}
\examples{
blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
