% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Title Blblm without parallelization}
\usage{
blblm(formula, data, m = 3, B = 5000)
}
\arguments{
\item{formula}{formula for regression}

\item{data}{dataset or a list of files}

\item{m}{number of subset or number of files}

\item{B}{number of boostrap samples}
}
\value{
fitted lm
}
\description{
Title Blblm without parallelization
}
\examples{
\dontrun{
blblm(mpg~wt,mtcars, 3,1000)
}
}
