% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, file, m = 10, B = 5000, cl = 1)
}
\arguments{
\item{formula}{formula}

\item{file}{dataframe or a list of file of datasets}

\item{m}{number of parts data split into}

\item{B}{number of bootstraps}

\item{cl}{number of clusters to use}
}
\value{
list
}
\description{
Linear Regression with Little Bag of Bootstraps
}
