% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_glm}
\alias{blblm_glm}
\title{Generate linear regression with boostrap}
\usage{
blblm_glm(formula, data, m = m, B = B)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{m}{number of subdata}

\item{B}{number of bootstrips}
}
\value{
blblm
}
\description{
Generate linear regression with boostrap
}
\examples{
blblm_glm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
}
