% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_par.R
\name{blblm_par}
\alias{blblm_par}
\title{Fit a linear regression model with more than one CPU works}
\usage{
blblm_par(formula, data, m = 10, B = 5000, cl)
}
\arguments{
\item{formula}{The linear regression model we fitted}

\item{data}{which is usual is the data frame, which is the data we imput.}

\item{m}{which is numeric variables, indicates the number of splits we need}

\item{B}{which is a numeric variable, indicates number of bootstraps we need}

\item{cl}{The muticluster we get from parallel}
}
\value{
blblm object
}
\description{
Fit the linear regression model with data frame input with more than one CPU works
}
\examples{
#library(parallel)
#cl<- makeCluster(4)
# fit<- blblm_par(mpg~., data = mtcars, m = 10, B = 5000,cl)
#stopCluster(cl)
}
