% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{Compute Linear regression Model}
\usage{
blblm_par(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{variale formula for model}

\item{data}{data frame we are working on}

\item{m}{number of subsamples}

\item{B}{number of bootstrap samples}
}
\value{
fitted model using mutliple clusters and blb technique
}
\description{
Find linear regression model using subsamples and bootstrap samples with option for parallelization
Users will need to make clusters own their own.Recommanded code is plan(multisession, worker = 4).
}
\examples{
blblm_par(mpg ~ wt * hp, data= mtcars, m=3, B=100)
}
