% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{Title}
\usage{
blblm_par(formula, data, m = 10, B = 5000, cluster)
}
\arguments{
\item{formula}{regression formula}

\item{data}{a list of files, or a vector}

\item{m}{prefered size of data seperation}

\item{B}{number of bootstraps}

\item{cluster}{parallel cluster}
}
\value{

}
\description{
Title
}
