% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{blb parallel}
\usage{
blblm_par(formula, data, m, B, cl)
}
\arguments{
\item{formula}{the interest user want from data}

\item{data}{the dataset}

\item{m}{numbers of subsamples created}

\item{B}{the number of bootstraps user want}

\item{cl}{the cluster user create}
}
\description{
It allows user to do bag of little bootstrap linear regression applying cluster to speed the process
}
