% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{Generate linear regression with boostrap}
\usage{
blblm_par(formula, data, m = m, B = B, cl)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data we are dealing with}

\item{m}{number of subdata}

\item{B}{number of bootstrips}

\item{cl}{cluster/integer}
}
\value{
blblm_par
}
\description{
Generate linear regression with boostrap
}
