% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{Calculate Regression With Parallelization}
\usage{
blblm_par(formula, data, m, B, ncores)
}
\arguments{
\item{formula}{The regression formula}

\item{data}{The input data frame.}

\item{m}{Number of splits to split the data into. m > 1.}

\item{B}{Number of bootstraps to be performed in estimating the coefficients. B > 1.}

\item{ncores}{Number of cores to use. Must be between 1 and the maximum number of cores you have.}
}
\description{
Calculates the linear regression coefficients of a dataframe and formula using a specified number
of cores, using B bootstraps and with the data split into m parts.
}
\details{
An example is below:
fit <- blblm_par(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, ncores = 2)
coef(fit)
Note: This cannot be ran during check() due to an error with the future package.
}
