% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm_par}
\alias{blblm_par}
\title{Blblm with parallelization}
\usage{
blblm_par(formula, data, m = 3, B = 5000, cl = cl)
}
\arguments{
\item{formula}{regression formula}

\item{data}{dataset}

\item{m}{number of subset}

\item{B}{number of boostrap samples}

\item{cl}{cluster created}
}
\value{
fitted lm
}
\description{
Blblm with parallelization
}
\examples{
\dontrun{
cl<- makeCluster(4)
invisible(clusterEvalQ(cl, {library(tidyverse)
NULL}))
blblm_par(mpg~wt,mtcars, 3,1000 , cl)
stopCluster(cl)}
}
