% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{blblog}
\alias{blblog}
\title{Calculate some statistics via Logistic Regression With a Bag of Little Bootstraps Implementation}
\usage{
blblog(formula, data, m = 10, B = 5000, par = FALSE)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame, list, or environment(iterable)}

\item{m}{numeric which indicates the number of subsamples. Default is 10.}

\item{B}{numeric which is bootstrapped samples per subsample. Default is 5000.}

\item{par}{logical. Indicate whether or not user wants parallelization to be run. FALSE by default.}
}
\value{
blblog returns an object of class blblog
}
\description{
We use the bag of little bootstraps algorithm in order to calculate the statistics generated by weighted logistic regression.
}
\details{
This function is used to fit a bag of little bootstraps version of logistic regression. \code{blblog} returns a fitted object of class "blblog, with a size of length two.
}
\examples{
fit <- blblog(Species ~ Sepal.Length, data = iris, m = 3, B = 100)
}
