% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{blblog}
\alias{blblog}
\title{Compute logistic regression model with Bag of Little Bootstraps}
\usage{
blblog(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{logistic regression formula}

\item{data}{data frame}

\item{m}{integer indicating the number of subsets}

\item{B}{integer indicating the number of bootstrap}

\item{parallel}{boolean which contains only two value, TRUE or FALSE}
}
\value{
fitted model by using blb and parallezation
}
\description{
Find logistic regression model using subsamples and bootstrap samples with parallelization.
The user can specify the number of workers by plan(multisession, workers = 4) if they choose to run in parallezation.
For example, when they write parallel = TRUE, they need can change the number of clusters as they needed.
}
