% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{blblogcoef}
\alias{blblogcoef}
\title{compute the coefficients from object}
\usage{
blblogcoef(fit)
}
\arguments{
\item{fit}{fitted model from blblog}
}
\value{
the coefficients
}
\description{
compute the coefficients from object
}
