% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{blblogreg}
\alias{blblogreg}
\title{Bag of Little Bootstraps Version of Logistic Regression}
\usage{
blblogreg(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be converted to that class).}

\item{data}{a data.frame, environment, or list.}

\item{m}{the number of splits on the data, m = 10 by default.}

\item{B}{the number of bootstraps, B = 5000 by default.}

\item{parallel}{boolean, specify whether to use parallelization or not, parallel = FALSE by default}
}
\value{
a blblogreg object
}
\description{
blblogreg performs bag of little bootstrap (blb) on glm() function. If users set "parallel = TRUE" in the blblogreg() function, the plan() function is usually needed for users to specify in front of the blblogreg() function.
}
\examples{
library(ISLR)
blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = Smarket, parallel = TRUE)
blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = Smarket, m = 3, B = 100)
}
