% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblogreg}
\alias{blblogreg}
\title{The following package is used to fit logistics models}
\usage{
blblogreg(formula, data, m = 10, B = 5000, family, parallel = TRUE)
}
\arguments{
\item{formula}{the formula to use for the function; an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{m}{factor. Number of splits}

\item{B}{factor. Number of bootstraps}

\item{family}{specify which glm family is used.Use binomial for logistic regression.}

\item{parallel}{logical. Specifies whether to use parallelization. If TRUE, parallelization is used.}
}
\value{
object of class "blblogreg"
}
\description{
The following package is used to fit logistics models
}
\examples{
blblogreg(mpg ~wt *hp, data = mtcars, m = 3, B = 100, family = binomial(), parallel = TRUE)
}
