% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{blblogsigma}
\alias{blblogsigma}
\title{compute sigma from object}
\usage{
blblogsigma(fit)
}
\arguments{
\item{fit}{fitted model from blblog}
}
\value{
the standard errors
}
\description{
compute sigma from object
}
