% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blbrf.R
\docType{package}
\name{blbmethods-package}
\alias{blbmethods}
\alias{blbmethods-package}
\title{blbmethods: Implement Bag of Little Bootstraps for Linear Model}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Linear Regression with Little Bag of Bootstraps

Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: David Mao \email{dlmao@ucdavis.edu}


\strong{Maintainer}: David Mao \email{dlmao@ucdavis.edu}

}
