% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbrf.R
\name{blbrf}
\alias{blbrf}
\title{Bind two factors}
\usage{
blbrf(formula, data, m = 10, B = 500, nthreads = 1)
}
\arguments{
\item{formula}{object of class "formula"}

\item{data}{a dataframe}

\item{m}{an integer}

\item{B}{an integer}

\item{nthreads}{an integer}
}
\value{
blbrf
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
blbrf(mpg ~ wt * hp, data = mtcars, m = 3, B = 10000, nthreads = 8)
}
