% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{sigma}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{A fit model}
}
\description{
call returns the sigma value of the given fit model
}
\details{
compute sigma from fit
}
