% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Compute sigma (residual standard error) from fit}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{lm}
}
\value{
numeric sigma estimate
}
\description{
Compute sigma (residual standard error) from fit
}
