% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Calculate signma}
\usage{
blbsigma(object)
}
\arguments{
\item{object}{fitted model}
}
\value{
sigma for the fit
compute sigma from fit
}
\description{
Calculate sigma value for input fitted model
}
