% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Sigma value computed from bag of litte bootstraps
ompute sigma from fit}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{fitted values, use to get residuals/coefficients}
}
\description{
Sigma value computed from bag of litte bootstraps
ompute sigma from fit
}
