% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{compute sigma from fit (not compatible with lm1C)}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{the fitted model returned from blblm}
}
\description{
compute sigma from fit (not compatible with lm1C)
}
