% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R, R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Residual Standard Deviation Sigma
Computes the Residual Standard Deviation Sigma}
\usage{
blbsigma(fit)

blbsigma(fit)
}
\arguments{
\item{fit}{a fitted blblm model}
}
\value{
a numeric value, the value of the residual standard deviation sigma

a numeric value, the value of the residual standard deviation sigma
}
\description{
Residual Standard Deviation Sigma
Computes the Residual Standard Deviation Sigma

Residual Standard Deviation Sigma
Computes the Residual Standard Deviation Sigma
}
