% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blbglm_cluster.R, R/blblm.R,
%   R/blblm_cluster.R
\name{blbsigma}
\alias{blbsigma}
\title{compute sigma from fit}
\usage{
blbsigma(fit)

blbsigma(fit)

blbsigma(fit)

blbsigma(fit)
}
\description{
compute sigma from fit

compute sigma from fit

compute sigma from fit

compute sigma from fit
}
