% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Computing Sigma}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{a list containing information about the fitted model.}
}
\value{
the sigma of the model.
}
\description{
Compute sigma from fit.
}
