% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Blblm Standard Deviation
compute sigma from fit}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{Blblm object}
}
\description{
Blblm Standard Deviation
compute sigma from fit
}
