% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Sigma (C++)}
\usage{
blbsigma(object)
}
\arguments{
\item{object}{Fitted linear regression model}
}
\value{
Sigma value
}
\description{
Computes sigma value from fit with lmC
}
