% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Compute sigma from fit}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{The linear regression model from lm().}
}
\description{
Computed the standard deviation estimation from the fitted linear regression model.
}
