% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{Bag of Little Bootstraps Sigma}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{Fitted linear model.}
}
\value{
Sigma of the given fit.

compute sigma from fit
}
\description{
Computes sigma of a given linear model.
}
