% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{compute sigma from fit}
\usage{
blbsigma(fit)
}
\arguments{
\item{fit}{a model, generally linear, to compute the sigma for}
}
\value{
standard deviation
}
\description{
compute sigma from fit
}
