% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{blbsigma}
\alias{blbsigma}
\title{compute sigma from fit}
\usage{
blbsigma(fit)

blbsigma(fit)
}
\arguments{
\item{fit}{the result of lm()}
}
\value{
sigma from fit

sigma from fit
}
\description{
compute sigma from fit

compute sigma from fit
}
