% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Extract Model Coefficients}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is meaningfu}

\item{...}{other arguments}
}
\value{
coefficients extracted from the model object
}
\description{
coef.blbglm is a function which extracts model coefficients
}
