% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Calculate coefficients for a blbglm object}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{The object to calculate coefficients for.}

\item{...}{Additional parameters to pass.}
}
\value{
Bag of little bootstraps coefficients.
}
\description{
Calculate coefficients for a blbglm object
}
