% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{coef.blbglm}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{blbglm}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns blb coefficients of logistic regression model
}
