% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Coefficient of Model
Find the coefficients of the model object.}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{an object in which coefficient extraction will be performed on}

\item{...}{additional arguments}
}
\value{
the coefficient extract from the object
}
\description{
Coefficient of Model
Find the coefficients of the model object.
}
