% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{coefficeint of logistic regression model}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{estimated object}

\item{...}{further arguments passed from or to other methods}
}
\description{
coefficeint of logistic regression model
}
