% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{regression model coefficients}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{the model returned from the blbglm function}

\item{...}{additional parameters to be passed in}
}
\value{
matrix of the regression model coefficients
}
\description{
regression model coefficients
}
