% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Getting Coefficients of Fit}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{a blbglm model.}

\item{...}{additional arguments to be pased.}
}
\value{
coefficients of the model.
}
\description{
Get the coefficients for the blblm model.
}
