% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Extracts model coefficients}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{blbglm}

\item{...}{arg}
}
\description{
Extracts model coefficients
}
