% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Gets coefficients for blbglm}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{the blbglm model to use}

\item{...}{extra arguments}
}
\description{
overloads the standard \code{coef} functionality to get the coefficients for a \code{blbglm} model.
}
\examples{
blbglm_model <- blbglm(am ~ wt * hp, data = mtcars, m = 3, B = 100)
coef(blbglm_model)
}
