% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Computing the coeficients of blbglm model}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{the fit of blbglm}

\item{...}{Other arguments}
}
\description{
Computing the coeficients of blbglm model
}
