% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Get the model coefficients}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{A variable with the class "blblm"}

\item{...}{Included for method flexibility}
}
\value{
A vector of average values for the linear model's coefficients
}
\description{
Get an estimate for the coefficients
}
