% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Regression coefficients from blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm result}

\item{...}{other parameters}
}
\description{
Linear regression coefficients from blblm function is returned.
}
