% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Obtain the coefficients of the fitted model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{a blblm object, the fitted model}

\item{...}{other conditions}
}
\value{
return the of object (the fitted model)
}
\description{
Obtain the coefficients of the fitted model
}
\examples{
coef(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
> (Intercept)          wt          hp       wt:hp
> 48.88428523 -7.88702986 -0.11576659  0.02600976
}
