% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients for blblm model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{A fit blblm model}

\item{...}{Argument that is not predefined}
}
\description{
This function returns the coefficients when given your fit model produced from the blblm function
}
\examples{
fit = blblm(mpg ~ wt * hp, data = mtcars, m = 2, B = 10000)
coef(fit)
}
