% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Extract Model Coefficients}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is meaningfu}

\item{...}{other arguments}
}
\value{
coefficients extracted from the model object
}
\description{
coef.blblm is a function which extracts model coefficients
}
