% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Used to find the coef of blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm function}

\item{...}{}
}
\value{
object of class "blblm" (intercepts)
}
\description{
Used to find the coef of blblm
}
