% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coef}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{A blblm object}

\item{...}{...}
}
\description{
Coef
}
\details{
Returns the mean of all estimated coefficients for a given blblm fitted object.
}
