% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{compute coef from fit}
\usage{
\method{coef}{blbglm}(object, ...)
}
\arguments{
\item{object}{model}

\item{...}{est}
}
\value{
tibble
}
\description{
compute coef from fit
}
