% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{coef.blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns blb coefficients of linear model
}
