% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{BLB estimator of regression coefficients (\eqn{\beta} s)}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{a blblm object returned by function \code{blblm}}

\item{...}{additional arguments to be passed}
}
\value{
a dataframe of estimated regression coefficients.
}
\description{
Calculate the point estimation of regression coefficients (\eqn{\beta} s).
}
