% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{to be filled}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{to be filled}

\item{...}{to be filled}
}
\value{
to be filled
}
\description{
to be filled
}
