% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Obtain the coefficients}
\usage{
\method{coef}{blblm}(object, par = FALSE, ...)
}
\arguments{
\item{object}{object of class "blblm"}

\item{par}{logical FALSE by default. TRUE indicates desire to use parallelization.}
}
\value{
numeric
}
\description{
We use this function \code{coef.blblm} in order to find coefficients for all variables involved. This is comparable to using \code{coef}
}
\examples{
fit = blblm(mpg ~ hp * wt, data = mtcars, m = 3, B = 100)
coef.blblm(fit)
}
