% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{pulls the coefficients from the blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{a blblm result}

\item{...}{additional parameters}
}
\description{
pulls the coefficients from the blblm
}
