% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Get the coefficient of linear regression}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{estimated object}

\item{...}{further arguments passed from or to other methods}
}
\description{
Get the coefficient of linear regression
}
