% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{coefblblm}
\usage{
\method{coef}{blblm}(object, ncpu = 1, ...)
}
\arguments{
\item{object}{object}

\item{ncpu}{positive interger, number of CPU. Default to be 1 which represents one CPU.}
}
\value{

}
\description{
coefblblm
}
