% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Returns bootstrapped estimates of paramter coefficients. Parallel computing is optional.}
\usage{
\method{coef}{blblm}(object, para = FALSE, core = 2, ...)
}
\arguments{
\item{object}{linear model}

\item{para}{logical indicating parallel computing}

\item{core}{integer indiciating number of cores}

\item{...}{so on}
}
\description{
Returns bootstrapped estimates of paramter coefficients. Parallel computing is optional.
}
