% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{blblm coef calculation}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the variable or parameter}

\item{...}{extra arguments other than this function}
}
\description{
blblm coef calculation
}
