% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Gives coeficients of regression model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{regression model}

\item{...}{add whatever you want}
}
\description{
Gives coeficients of regression model
}
