% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients of Bag of Little Bootstraps}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the fitted object}

\item{...}{additional parameters related to coef.lm}
}
\description{
The mean coefficients for the BLB
}
