% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{regression model coefficients}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the model returned from the main blblm function}

\item{...}{additional parameters to be passed in}
}
\value{
matrix of the regression model coefficients
}
\description{
regression model coefficients
}
