% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Getting Coefficients of Fit}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{a blblm model.}

\item{...}{additional arguments to be pased.}
}
\value{
coefficients of the model.
}
\description{
Get the coefficients for the blblm model.
}
