% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Get the coefficient of fitted model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the fitted model}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get the coefficient of fitted model
}
