% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{computer coeficient}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm model}

\item{...}{...}
}
\value{
vector
}
\description{
computer coeficient from the estimate
}
