% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Get coefficient of objects from fitted model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{fit object}

\item{...}{extra conditions}
}
\value{
coefficient of fit objects
}
\description{
Get coefficient of objects from fitted model
}
\examples{
coef(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
}
