% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Creating the Coefficient}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the blblm interval}

\item{...}{continuation}
}
\description{
This function computes the boothstrap coefficient
}
