% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Computing the Coefficients for the Model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{additional arguments affecting the result}
}
\description{
Computing the Coefficients for the Model
}
