% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{Fitted linear regression model}

\item{...}{Additional parameters}
}
\description{
Extracts coefficients from the blblm
}
