% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients of blblm Model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{Fitted blblm model.}

\item{...}{Additional arguments.}
}
\value{
Estimated coefficients of the given fitted blblblm model.
}
\description{
Estimates the coefficients for a fitted blblm model.
}
