% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients for blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{model fit produced by blblm(formula, data, ...)}

\item{...}{other arguments.}
}
\description{
Outputs the coefficients for the fitted object (after blblm algorithm)
}
