% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Gets coefficients for BLBLM}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the blblm model to use}

\item{...}{extra arguments}
}
\description{
overloads the standard \code{coef} functionality to get the coefficients for a \code{blblm} model.
}
\examples{
blblm_model <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, num_workers = 4, parallelize = FALSE, fast = FALSE)
coef(blblm_model)
}
