% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{coef.blblm is a function to extract the model coefficients from the blblm model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{an object which has relevant coefficients to be extracted}

\item{...}{additional arguments to be passed}
}
\description{
coef.blblm is a function to extract the model coefficients from the blblm model
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
coef(fit)

}
