% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Computing the coeficients of blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{the fit of blblm}

\item{...}{Other arguments}
}
\description{
Computing the coeficients of blblm
}
