% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients from blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{The result from \code{blblm}}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The regression coefficient result from \code{blblm}
}
\description{
Coefficients from blblm
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
coef(fit)
}
