% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{calculate the blblm avg coeficient}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{a blblm object}
}
\value{
a coefficient
}
\description{
calculate the blblm avg coeficient
}
