% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{find the coefficient in blblm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{fitted model from blblm}

\item{...}{additional arguements/inputs}
}
\value{
the estimated coefficient
}
\description{
find the coefficient in blblm
}
