% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Calculate coefficient}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm model}

\item{...}{other parameters}
}
\description{
Calculate coefficient
}
