% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{coef.blblog}
\alias{coef.blblog}
\title{Find the coefficients}
\usage{
\method{coef}{blblog}(object, par = FALSE, ...)
}
\arguments{
\item{object}{an object of class "blblog"}

\item{par}{logical. Indicate whether or not user wants parallelization to be run. FALSE by default.}

\item{...}{optional arguments to be passed}
}
\value{
numeric. Returns the coefficient of the blblog model.
}
\description{
We use the bag of little bootstraps algorithm in order to calculate the statistics generated by weighted logistic regression.
}
\details{
Using \code{coef} produces the coefficients used in the bag of little bootsraps algorithm. These coefficients can be used to solve for the log odds raio.
}
\examples{
fit <- blblog(Species ~ Sepal.Length, data = iris, m = 3, B = 100)
coef(fit)
}
