% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{coef.blblog}
\alias{coef.blblog}
\title{find the coefficient in blblog}
\usage{
\method{coef}{blblog}(object, ...)
}
\arguments{
\item{object}{fitted model from blblog}

\item{...}{additional arguements/inputs}
}
\value{
the estimated coefficient
}
\description{
find the coefficient in blblog
}
