% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblogreg}
\alias{coef.blblogreg}
\title{Used to find the coef of blblogreg}
\usage{
\method{coef}{blblogreg}(object, ...)
}
\arguments{
\item{object}{blblogreg function}

\item{...}{}
}
\value{
object of class "blblogreg" (intercepts)
}
\description{
Used to find the coef of blblogreg
}
