% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{coef.blblogreg}
\alias{coef.blblogreg}
\title{coefblblogreg}
\usage{
\method{coef}{blblogreg}(object, ncpu = 1, ...)
}
\arguments{
\item{object}{object}

\item{ncpu}{positive interger, number of CPU. Default to be 1 which represents one CPU.}

\item{...}{unused variables}
}
\description{
coefblblogreg
}
