% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{coef.blblogreg}
\alias{coef.blblogreg}
\title{Coefficients of blblogreg Model}
\usage{
\method{coef}{blblogreg}(object, ...)
}
\arguments{
\item{object}{Fitted blblogreg model.}

\item{...}{Additional arguments.}
}
\value{
Estimated coefficients of the given fitted blblogreg model.
}
\description{
Estimates the coefficients for a fitted blblogreg model.
}
