% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{coef.blblogreg}
\alias{coef.blblogreg}
\title{Coefficients from blblogreg}
\usage{
\method{coef}{blblogreg}(object, ...)
}
\arguments{
\item{object}{The result from \code{blblogreg}}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The regression coefficient result from \code{blblogreg}
}
\description{
Coefficients from blblogreg
}
\examples{
fit <- blblogreg(Species ~ Sepal.Length * Petal.Length, binomial, iris[1:100,], 3, 100, FALSE)
coef(fit)
}
