% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Calculate confidence intervals for a blbglm object}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The object to calculate confidence intervals for.}

\item{parm}{The parameter to calculate confidence intervals for.}

\item{level}{The level to use for the confidence interval.}

\item{...}{Additional parameters to pass.}
}
\value{
Bag of little bootstraps confidence intervals.
}
\description{
Calculate confidence intervals for a blbglm object
}
