% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{conflint.blbglm}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blbglm}

\item{parm}{string specific fit variable}

\item{level}{double confidence interval level}

\item{...}{further arguments passed to or from other methods.}
}
\description{
blb confidence interval for logistic regression coefficients
}
