% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Model Parameter Confidence Intervals
Compute the confidence intervals for a parameter in the model.}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object for which a confidence interval is wanted}

\item{parm}{the parameters that you want to compute the confidence interval of}

\item{level}{numeric, the confidence level}

\item{...}{additonal arguments}
}
\value{
A matrix or vector that give the lower and upper confidence bounds for the parameter
}
\description{
Model Parameter Confidence Intervals
Compute the confidence intervals for a parameter in the model.
}
