% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{confidence interval of the selected independent variables}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the model returned from the main blbglm function}

\item{parm}{the independent variables you want the confidence interval for}

\item{level}{confidence level between 0 and 1}

\item{...}{additional parameters to be passed in}
}
\value{
matrix of the lower and upper quantiles of the selected independent variables.
}
\description{
confidence interval of the selected independent variables
}
