% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Getting the Confidence Interval}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a blbglm model.}

\item{parm}{a list of parameters used to construct confidence interval.}

\item{level}{the significance level for the confidence interval.}

\item{...}{additional arguments to be passed.}
}
\value{
a matrix or vector with columns giving lower and upper confidence limits for each parameter.
}
\description{
Get the confidence interval for the coeficients.
}
