% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Confidence Intervals For Model Parameters}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blbglm}

\item{parm}{list}

\item{level}{numeric}

\item{...}{arg}
}
\description{
Confidence Intervals For Model Parameters
}
