% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Gets confident interval for blbglm variables}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the blbglm model to use}

\item{parm}{the parameters we want to find the confidence intervals of}

\item{level}{the confidence level for each confidence interval}

\item{...}{extra arguments}
}
\description{
overloads the standard \code{confint} functionality to calculate the confidence intervals for a \code{blbglm} model's coefficients.
}
\examples{
blbglm_model <- blbglm(am ~ wt * hp, data = mtcars, m = 3, B = 100)
confint(blbglm_model, c("wt", "hp"))
}
