% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Computing the confidence interval of blbglm model}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level, ...)
}
\arguments{
\item{object}{the fit of blbglm}

\item{parm}{the users can choose whether to use parameters or not}

\item{level}{level of confidence, Eg.0.95}

\item{...}{Other arguments}
}
\description{
Computing the confidence interval of blbglm model
}
