% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Get a confidence interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A variable with the class "blblm"}

\item{parm}{A vector or list of the names of the variables for which an interval will be constructed.
Exact names (as output by stats::terms()) must be used. For intercepts, use "(Intercept)".}

\item{level}{The significance level for the confidence interval (1 - alpha)}

\item{...}{Included for method flexibility}
}
\value{
A matrix with confidence intervals for the specified parameters
}
\description{
Produce a confidence interval for some or all of the model c0efficients (the latter by default)
}
