% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence interval from blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm result}

\item{parm}{parameters for confidence interval}

\item{level}{confidence level in decimal}

\item{...}{other parameters}
}
\description{
Confidence interval from blblm function are returned for specified column names from dataset
}
