% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Obtain the confidence intervals for the parameters of the fitted model}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm object, the fitted model}

\item{parm}{a specification of which parameters, e.g a string vector, NULL by default}

\item{level}{double, the significance level 1-alpha, 0.95 by default}

\item{...}{extra conditions}
}
\value{
return the confidence intervals for the param of the object
}
\description{
Obtain the confidence intervals for the parameters of the fitted model
}
\examples{
confint(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100), c("wt", "hp"))
}
