% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence Interval for blblm model}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A fit blblm model}

\item{parm}{parameters set to compute your confidence interval}

\item{level}{The level of confidence to computer your confidence interval. Default level set to 0.96}

\item{...}{Argument that is not predefined}
}
\description{
This function is used to compute the confidence interval for a given fit blblm model
}
\examples{
fit = blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 10000)
confint(fit, c("wt", "hp"),level = 0.95)
}
