% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{The Confidence Level for the function blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required}

\item{...}{a model object for which prediction is desired}
}
\value{
A matrix or vector with columns giving lower and upper confidence limits for each parameter.
}
\description{
confint.blblm gets the confidence level for the model
}
