% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Used to find the confint of blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm function}

\item{parm}{logical. Specify estimated model parameters if parm is not NULL.}

\item{level}{confidence level. Set at 0.95.}

\item{...}{}
}
\value{
object of class "blblm"
}
\description{
Used to find the confint of blblm
}
