% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Construct a confidence interval for named coefficients in the model.
If coefficient names not specified, than confidence intervals constructed for all coeffienct names.}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm fitted model}

\item{parm}{character Names of the coeffiecients used for confidence intervals.}

\item{level}{numeric Confidence level.}

\item{...}{further arguments}
}
\value{
matrix Confidence Interval lower and upper bounds for each coefficient.
}
\description{
Construct a confidence interval for named coefficients in the model.
If coefficient names not specified, than confidence intervals constructed for all coeffienct names.
}
\examples{
confint(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100), c("wt", "hp"))
}
