% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confint blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A blblm object}

\item{parm}{The given parameters to be estimated}

\item{level}{The given level of confidence (default .95)}

\item{...}{...}
}
\description{
Confint blblm
}
\details{
Returns a confidence interval for the given parameters (if no parameters are given, gives a confidence interval for all parameters in the model formula)
}
