% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{The method for the blblm class}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{x}{Which is a blblm variable.}
}
\value{
sigma which is the confident interval of fit
}
\description{
For a blblm object input, we get the confident interval
}
\examples{
# fit<- blblm(mpg~., data = mtcars, m = 10, B = 5000)
# confint(fit)
}
