% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{compute confidence interval from fit}
\usage{
\method{confint}{blbglm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{model}

\item{parm}{boolean vector}

\item{level}{cofidential level}

\item{...}{est}
}
\value{
out tibble
}
\description{
compute confidence interval from fit
}
