% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{confidence interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Model}

\item{parm}{TRUE or FALSE}

\item{level}{confidence percentage}

\item{...}{extra stuff}
}
\description{
confidence interval
}
