% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{conflint.blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm}

\item{parm}{string specific fit variable}

\item{level}{double confidence interval level}

\item{...}{further arguments passed to or from other methods.}
}
\description{
blb confidence interval for Regression coefficients
}
