% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{to be filled}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{to be filled}

\item{parm}{to be filled}

\item{level}{to be filled}

\item{...}{to be filled}
}
\value{
to be filled
}
\description{
to be filled
}
