% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence interval of our blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{object / model that we want to test}

\item{parm}{Parameters of regression variables we want to see. this is set to null by default and will use all variables. Set variables by using c("variable1", "variable2")}

\item{level}{Confidence level we want to test}

\item{...}{Any other argument you might want to pass.}
}
\description{
Confidence interval of our blblm
}
