% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Get the confidence interval of parament of linear regression}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The data user defined.}

\item{parm}{The initial string}

\item{level}{The 95\% confidence interval}

\item{...}{further arguments passed from or to other methods}
}
\description{
Get the confidence interval of parament of linear regression
}
