% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{confintblblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ncpu = 1, ...)
}
\arguments{
\item{object}{object}

\item{parm}{parameter. default is NULL}

\item{level}{this is a confidence level, default at 0.96=5.}

\item{ncpu}{positive interger, number of CPU. Default to be 1 which represents one CPU.}
}
\value{

}
\description{
confintblblm
}
