% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Returns confidence intervals for estimated parameter coefficients. Parallel computing is optional.}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, para = FALSE, core = 2, ...)
}
\arguments{
\item{object}{linear model}

\item{parm}{parameter for confidence interval}

\item{level}{numeric indicating level of confidence}

\item{para}{logical indicating  parallel computing}

\item{core}{integer indicating number of cores in parallel computing}

\item{...}{so on}
}
\description{
Returns confidence intervals for estimated parameter coefficients. Parallel computing is optional.
}
