% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{confint interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the regression model}

\item{parm}{the input of parameter}

\item{level}{level of confidence}

\item{...}{extra arguments other than this function}
}
\description{
a function that computes the bootstrap confidence interval
}
