% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Obtaining confidence intervals}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)

\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the fit}

\item{parm}{indicate the target variables, leave as NULL for ones already specified in the fit}

\item{level}{significance level}

\item{...}{additional arguments to be passed to the low level regression fitting functions}
}
\value{
Returns the confidence interval of the fit.

Returns the confidence interval of the fit.
}
\description{
Obtaining confidence intervals

Obtaining confidence intervals
}
\examples{
fit <- blblm_mod(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
confint(fit, c("wt", "hp"))
}
