% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{parallelized confidence intervals}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{linear regression model}

\item{parm}{parameters for confidence interval}

\item{level}{confidence level}

\item{...}{add whatever parameters you want}
}
\description{
parallelized confidence intervals
}
