% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence interval of parameters}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model}

\item{parm}{parameters}

\item{level}{level of confidence}

\item{...}{further arguments passed to or from other methods}
}
\description{
compute all parameters' CIs if parm is not specified,
otherwise compute specific parm's CI
}
