% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence intervals of blblm
Returns regression coefficients of blblm object calculated from averaging out all coefficient estimates}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Blblm object you want to find coefficients of}

\item{parm}{Attributes that you want to find significance level of. By default is NULL}

\item{level}{Significance level of your confidence intervals}
}
\description{
Confidence intervals of blblm
Returns regression coefficients of blblm object calculated from averaging out all coefficient estimates
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
confint(fit)
}
