% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{computer coeficient}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm model}

\item{parm}{parameter}

\item{level}{confidence level}

\item{...}{...}
}
\value{
vector of interval
}
\description{
computer coeficient from the estimate
}
