% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blblm}
\alias{confint.blblm}
\title{confidence interval for the estimate of blblm objects}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{parm}{label of blblm object formula}

\item{level}{significance level}

\item{...}{extra conditions}
}
\value{
confidence interval of the estimate of blblm object
}
\description{
confidence interval for the estimate of blblm objects
}
\examples{
confint(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100), c("wt", "hp"))
}
