% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Creating Confint Interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the blblm interval}

\item{parm}{null and factor-variables}

\item{level}{for confidence interval}

\item{...}{continuation}
}
\description{
This function computes the boothstrap confidence Interval
}
