% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence Interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{parm}{numeric value indicating the parameters you want to construct confidence interval for}

\item{level}{integer between 0 and 1 indicating the confidence level}

\item{...}{additional arguments affecting the result}
}
\value{
numeric confidence interval
}
\description{
Confidence Interval
}
