% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence Interval blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted linear regression model}

\item{parm}{Model parameters}

\item{level}{Confidence level}

\item{...}{Additional parameters}
}
\description{
Computes confidence interval for a given confidence level
}
