% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence Interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The object from which the coefficients are estimated.}

\item{parm}{The paramters for which to find a confidence interval.}

\item{level}{The \% confidence level of the estimate.}

\item{...}{Other arguments}
}
\description{
Gives a 95\% confidence interval for the coefficients used in the formula specified in the linear regression.
}
