% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Gets confident interval for BLBLM variables}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{the blblm model to use}

\item{parm}{the parameters we want to find the confidence intervals of}

\item{level}{the confidence level for each confidence interval}

\item{...}{extra arguments}
}
\description{
Gets confident interval for BLBLM variables
}
\examples{
blblm_model <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, num_workers = 4, parallelize = FALSE, fast = FALSE)
confint(blblm_model, c("wt", "hp"))
}
