% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{confint.blblm is a function to compute confidence interval for one or multiple parameters in the blblm  model}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm model}

\item{parm}{the parameters given to calculate the confidence intervals}

\item{level}{an alpha level for the confidence interval}

\item{...}{additional arguments to be passed}
}
\description{
confint.blblm is a function to compute confidence interval for one or multiple parameters in the blblm  model
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
confint(fit, c("wt", "hp"))

}
