% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Computing the confidence interval of blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level, ...)
}
\arguments{
\item{object}{the fit of blblm}

\item{parm}{the users can choose whether to use parameters or not}

\item{level}{level of confidence, Eg.0.95}

\item{...}{Other arguments}
}
\description{
Computing the confidence interval of blblm
}
