% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence Interval from blblm}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The result from \code{blblm}}

\item{parm}{The Object (or objects) from \code{blblm} model}

\item{level}{The significant value. (default is 0.95)}

\item{...}{Additional arguments to be passed to the function}
}
\description{
Confidence Interval from blblm
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
confint(fit)
confint(fit, c("wt", "hp"))
}
