% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{calculate the blblm avg confidence interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm object}

\item{parm}{variable in the model}

\item{level}{confidence level}
}
\description{
calculate the blblm avg confidence interval
}
