% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{obtain confidence interval of coefficients}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, nthreads = 1)
}
\arguments{
\item{object}{object of "blblm"}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level, the default value is 0.95}

\item{nthreads}{the number of cores you want to use when processing}
}
\value{
the confidence intervals of coefficients
}
\description{
obtain confidence interval of coefficients
}
