% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Calculate the confidence interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm model}

\item{parm}{parameter wanted}

\item{level}{confidence  level}

\item{...}{other parameters}
}
\description{
Calculate the confidence interval
}
