% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{confint.blblog}
\alias{confint.blblog}
\title{find the confidence interval in blblog}
\usage{
\method{confint}{blblog}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model from blbog}

\item{parm}{numerical value indicating the parameters you want to construct the ci for}

\item{level}{numeric, the confidence interval level}

\item{...}{additional arguements/inputs}
}
\value{
the confidence interval of the estimates
}
\description{
find the confidence interval in blblog
}
