% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{Obtain the confidence intervals for the parameters of the logistic model}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a blblogreg object, the logistic model}

\item{parm}{a specification of which parameters, e.g a string vector, NULL by default}

\item{level}{double, the significance level 1-alpha, 0.95 by default}

\item{...}{extra conditions}
}
\value{
return the confidence intervals for the param of the object
}
\description{
Obtain the confidence intervals for the parameters of the logistic model
}
\examples{
confint(blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = ISLR::Smarket, m = 3, B = 100, parallel = TRUE))
}
