% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{Obtain the confidence intervals for the parameters of the logistic model}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a blblogreg object, the logistic model}

\item{parm}{a specification of which parameters, e.g a string vector, NULL by default}

\item{level}{double, the significance level 1-alpha, 0.95 by default}

\item{...}{extra conditions}
}
\value{
return the confidence intervals for the param of the object
}
\description{
Obtain the confidence intervals for the parameters of the logistic model
}
\examples{
> confint(blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = Smarket, m = 3, B = 100, parallel = TRUE))
2.5\%      97.5\%
Lag1   -0.2499896 0.06857932
Lag2   -0.2276705 0.10999090
Volume -0.4333947 0.59521277
}
