% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{Used to find the confint of blblogreg}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblogreg function}

\item{parm}{logical. Specify estimated model parameters if parm is not NULL.}

\item{level}{confidence level. Set at 0.95.}

\item{...}{}
}
\value{
object of class "blblogreg"
}
\description{
Used to find the confint of blblogreg
}
