% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{confintblblogreg}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ncpu = 1, ...)
}
\arguments{
\item{object}{object}

\item{parm}{parameter. default is NULL}

\item{level}{this is a confidence level, default at 0.96=5.}

\item{ncpu}{positive interger, number of CPU. Default to be 1 which represents one CPU.}

\item{...}{unused variables}
}
\description{
confintblblogreg
}
