% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{Confidence Interval for blblogreg Model Estimates}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted blblogreg model.}

\item{parm}{Parameters to calculate confidence intervals for.}

\item{level}{Confidence level.}

\item{...}{Additional arguments.}
}
\value{
Confidence interval(s) for model estimate(s) or given parameter(s) at a specified confidence level.
}
\description{
Computes confidence interval(s) for model estimate(s) or given parameter(s) at a specified confidence level.
}
