% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{confint.blblogreg}
\alias{confint.blblogreg}
\title{Confidence Interval from blblogreg}
\usage{
\method{confint}{blblogreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{The result from \code{blblogreg}}

\item{parm}{The Object (or objects) from \code{blblogreg} model}

\item{level}{The significant value. (default is 0.95)}

\item{...}{Additional arguments to be passed to the function}
}
\description{
Confidence Interval from blblogreg
}
\examples{
fit <- blblogreg(Species ~ Sepal.Length * Petal.Length, binomial, iris[1:100,], 3, 100, FALSE)
confint(fit)
confint(fit, c("Sepal.Length", "Petal.Length"))
}
