% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_lm}
\alias{fast_lm}
\title{For more on using Rcpp click the Help button on the editor toolbar}
\arguments{
\item{x}{A single integer.}
}
\description{
For more on using Rcpp click the Help button on the editor toolbar
}
