% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{glm1}
\alias{glm1}
\title{#' Fitting Linear Model with A Bag of Little Bootstraps
estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs, n)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame, list or environment}

\item{freqs}{numeric}

\item{n}{numeric}
}
\value{
an object of class "lm" or for multiple responses of class
}
\description{
#' Fitting Linear Model with A Bag of Little Bootstraps
estimate the regression estimates based on given the number of repetitions
}
