% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{compute the logistic regression estimates for a blb dataset}
\usage{
glm1(formula, data, n, family)
}
\arguments{
\item{formula}{formula to use in blblm}

\item{data}{data}

\item{n}{number of rows}

\item{family}{glm family to specify}
}
\value{
list of coefficients and sigma
}
\description{
compute the logistic regression estimates for a blb dataset
}
