% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{Compute regression esimates
estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{variale formula for model}

\item{data}{data frame we are working on}

\item{freqs}{number of repetitions}
}
\description{
Compute regression esimates
estimate the regression estimates based on given the number of repetitions
}
