% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{same but different}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data, just data}

\item{freqs}{weights}

\item{family}{family in glm}
}
\description{
same but different
}
