% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{compute the regression estimates for a blb dataset}
\usage{
glm1(X, y, n)
}
\arguments{
\item{X}{Least Squares Matrix}

\item{y}{Least Squares target Matrix}

\item{n}{length of dataset}
}
\description{
compute the regression estimates for a blb dataset
}
