% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{glm1}
\alias{glm1}
\title{Fitting a model
Fitting a model using bag of little bootstraps}
\usage{
glm1(formula, data, freqs, n)
}
\arguments{
\item{formula}{an object of class called formula}

\item{data}{a data frame, list, or character vector}

\item{freqs}{a numeric value}

\item{n}{a numeric value}
}
\value{
a list of the coefficients and the sigma values
}
\description{
Fitting a model
Fitting a model using bag of little bootstraps
}
