% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blbglm_cluster.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs)

glm1(formula, data, freqs)
}
\description{
estimate the regression estimates based on given the number of repetitions

estimate the regression estimates based on given the number of repetitions
}
