% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{freqs}{number of frequencies}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
