% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{creates the glm model for each repetition and determines the coefficients and sigmas of each BLB.}
\usage{
glm1(formula, data, freqs)
}
\description{
creates the glm model for each repetition and determines the coefficients and sigmas of each BLB.
}
