% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{creates the glm model for each repetition and determines the coefficients and sigmas of each BLB.}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{the regression model you would like to create.}

\item{data}{the data you would like to analyze.}

\item{freqs}{frequency you want to create glm model for}
}
\description{
creates the glm model for each repetition and determines the coefficients and sigmas of each BLB.
}
