% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{Computing Regression Estimates}
\usage{
glm1(formula, data, n, family)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a set of data.}

\item{n}{number of rows of the data.}

\item{family}{a description of the error distribution and link function to be used in the model.}
}
\value{
a list of coefficients and sigma for the blb dataset.
}
\description{
Compute the regression estimates for a blb dataset.
}
