% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, family, freqs)
}
\arguments{
\item{formula}{linear model}

\item{data}{data.frame}

\item{family}{string type of glm}

\item{freqs}{weights of each row in sub-sample}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
