% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{freqs}{frequency}
}
\value{
estimate the regression estimates based on given the number of repetitions
estimate the regression estimates based on given the number of repetitions
}
\description{
estimate the regression estimates based on given the number of repetitions
}
