% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data we are dealing with}

\item{freqs}{frequency}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
