% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{glm1}
\alias{glm1}
\title{Fitting Linear Model using Bag of Little Bootstraps}
\usage{
glm1(formula, data, freqs, n)
}
\arguments{
\item{formula}{logistic regression formula}

\item{data}{dataframe}

\item{freqs}{numeric value}

\item{n}{numeric value}
}
\value{
object of class glm
}
\description{
Estimate the regression estimates based on given the number of repetitions
}
