% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe, list}

\item{freqs}{numeric}

\item{family}{family}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
