% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{Logistic Regression for BLB data set}
\usage{
glm1(X, y, n)
}
\arguments{
\item{X}{data model frame of predictor variables.}

\item{y}{data model frame of response variable.}

\item{n}{an integer, the number of rows in the data model frames.}
}
\description{
Compute the regression estimates for a blb data set
}
