% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{the generalized linear regression model with Normal distribution}

\item{data}{data frame}

\item{freqs}{weight}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
