% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{the formula for linear regression}

\item{data}{the subsample data}

\item{freqs}{the weights for each case in subsample data}

\item{family}{a description of the error distribution and link function to be used in the model}
}
\value{
the estimate of coefficients and sigma in one weighted linear regression
}
\description{
estimate the regression estimates based on given the number of repetitions
}
