% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{glm1}
\alias{glm1}
\title{compute the regression estimates for a blb dataset}
\usage{
glm1(formula, family, data, n)
}
\arguments{
\item{formula}{Model formula for regression}

\item{family}{A family (function) used for \code{glm}}

\item{data}{Data used to do regression}

\item{n}{Number of rows in \code{data}}
}
\description{
compute the regression estimates for a blb dataset
}
