% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs, n)
}
\arguments{
\item{formula}{logistic regression formula}

\item{data}{data frame}

\item{freqs}{number of repetitions}

\item{n}{integer}
}
\value{
the regression estimates
}
\description{
estimate the regression estimates based on given the number of repetitions
}
