% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, family, data, freqs)
}
\arguments{
\item{formula}{fomula: V1~V2}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{data}{dataframe}

\item{freqs}{multinomially distributed random number vectors}
}
\value{
list
}
\description{
estimate the regression estimates based on given the number of repetitions
}
