% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{a formula fitting a linear regression model.}

\item{data}{dataset for the model}

\item{freqs}{frequency for each bootstrapping}

\item{family}{family type for models, such as binomial, gaussian, gamma etc.}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
