% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{a formula}

\item{data}{a dataframe}

\item{n}{an integer}
}
\description{
compute the regression estimates for a blb dataset
}
