% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, family, data, n)
}
\description{
compute the regression estimates for a blb dataset
}
