% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Compute blb estimates
compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{variale formula for model}

\item{data}{data frame we are working on}

\item{n}{numeric}
}
\description{
Compute blb estimates
compute the regression estimates for a blb dataset
}
