% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Regression estimates for blbglm data set
Compute the regression estimates for blbglm data set}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame}

\item{n}{an integer}
}
\description{
Regression estimates for blbglm data set
Compute the regression estimates for blbglm data set
}
