% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Under BLB, this computes regression estimates.}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{the regression model you would like to create.}

\item{data}{the data you would like to analyze.}

\item{n}{number of rows of the original data.}
}
\description{
Under BLB, this computes regression estimates.
}
