% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Build glm for each bootstrap dataset}
\usage{
glm_each_boot(formula, data, family, n)
}
\arguments{
\item{formula}{linear model}

\item{data}{data.frame}

\item{family}{string type of glm}

\item{n}{integer each boot's size}
}
\description{
Build glm for each bootstrap dataset
}
