% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Generate logistic regression with boostrap}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{regression formula}

\item{data}{the data you want to regression}

\item{n}{number of row of the data}
}
\value{
regression estimates
compute the regression estimates for a blb dataset
}
\description{
compute the regression estimates for a blb dataset
}
