% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{Compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{a formula}

\item{data}{dataframe}

\item{n}{integer}
}
\description{
Compute the regression estimates for a blb dataset
}
