% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n, family)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe, list}

\item{n}{numeric}

\item{family}{family}
}
\description{
compute the regression estimates for a blb dataset
}
