% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{the generalized linear regression model}

\item{data}{data frame}

\item{n}{numeric, length of the data frame}
}
\description{
compute the regression estimates for a blb dataset
}
