% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset given a generated weight}
\usage{
glm_each_boot(formula, data, n, family)
}
\arguments{
\item{formula}{the formula for linear regression}

\item{data}{the subsample data}

\item{n}{the row number of original data}

\item{family}{a description of the error distribution and link function to be used in the model}
}
\value{
the estimate of coefficients and sigma in one weighted linear regression
}
\description{
compute the regression estimates for a blb dataset given a generated weight
}
