% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the estimates for each boot}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{logistic regression formula}

\item{data}{data frame}

\item{n}{integer}
}
\value{
the regression estimate for each blb dataset
}
\description{
Compute the regression estimate in each boot
}
