% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, family, data, n)
}
\arguments{
\item{formula}{fomula: V1~V2}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{data}{data list}

\item{n}{total number of rows in data}
}
\value{
model
}
\description{
compute the regression estimates for a blb dataset
}
