% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n, family)
}
\arguments{
\item{formula}{a formula fitting a linear regression model.}

\item{data}{dataset for the model}

\item{n}{number of rows of dataset}

\item{family}{family type for models, such as binomial, gaussian, gamma etc.}
}
\description{
compute the regression estimates for a blb dataset
}
