% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a formula}

\item{data}{a dataframe}

\item{n}{numeric}

\item{B}{numeric}
}
\description{
Compute the estimates
}
