% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute glm for each bootstrap}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{formula being passed from blbglm}

\item{data}{data}

\item{n}{number of rows for data}

\item{B}{number of bootstraps}

\item{family}{glm family to specify}
}
\value{
a list of glm objects
}
\description{
compute glm for each bootstrap
}
