% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Compute the estimates
Compute the estimates using bootstrap technique}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{variale formula for model}

\item{data}{data frame we are working on}

\item{n}{numeric}

\item{B}{number of bootstrap samples}
}
\description{
Compute the estimates
Compute the estimates using bootstrap technique
}
