% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{same same but different}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data, just data}

\item{n}{number of random vectors}

\item{B}{number of bootstraps}

\item{family}{family in glm}
}
\description{
same same but different
}
