% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{formula}

\item{data}{split dataset}

\item{n}{length of dataset}

\item{B}{number of repetitions}
}
\description{
compute the estimates
}
