% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Compute the estimates
Compute the estimates for the given formula}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame}

\item{n}{numeric}

\item{B}{numeric}
}
\description{
Compute the estimates
Compute the estimates for the given formula
}
