% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blbglm_cluster.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)

glm_each_subsample(formula, data, n, B)
}
\description{
compute the estimates

compute the estimates
}
