% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted}

\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{n}{number of subsets for the data}

\item{B}{number of bootstraps}
}
\description{
compute the estimates
}
