% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Computing Estimates}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a set of data.}

\item{n}{number of rows of the data.}

\item{B}{number of bootstraps.}

\item{family}{a description of the error distribution and link function to be used in the model.}
}
\value{
a list of estimates for the subsample.
}
\description{
Compute the estimates for each subsample.
}
