% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Set up bootstraps for each sub-samples}
\usage{
glm_each_subsample(formula, data, family, n, B)
}
\arguments{
\item{formula}{linear model}

\item{data}{data.frame}

\item{family}{string type of glm}

\item{n}{integer each boot's size}

\item{B}{integer number of bootstraps}
}
\description{
Set up bootstraps for each sub-samples
}
