% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data we are dealing with}

\item{n}{integer}

\item{B}{number of bootstrips}
}
\description{
compute the estimates
}
