% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe, list}

\item{n}{numeric}

\item{B}{numeric}

\item{family}{family}
}
\description{
compute the estimates
}
