% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Logistic Regression Model for Each Sub-sample}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{A symbolic description of the logistic regression model to be fitted.}

\item{data}{A data set, such as a data frame or list, containing the variables in the model.}

\item{n}{Number of rows of the given data set.}

\item{B}{Number of bootstrap samples.}
}
\value{
Model coefficients.

compute the estimates
}
\description{
Fits a logistic regression model on each sub-sample and returns the model coefficients
}
