% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Applying Each Sub-Sample to Function glm1}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame with the response and predictor variables.}

\item{n}{an integer, the number of rows in the data frame.}

\item{B}{an integer, the number of iterations over bag of little bootstraps algorithm.}
}
\description{
Compute the BLB estimates in conjunction with glm1
}
