% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{the generalized linear regression model with Normal distribution}

\item{data}{data frame}

\item{n}{numeric, length of the data frame}

\item{B}{numeric, repeat the process B times}
}
\description{
compute the estimates
}
