% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, family, data, n, B)
}
\arguments{
\item{formula}{Model formula for regression}

\item{family}{A family (function) used for \code{glm}}

\item{data}{Data used to do regression}

\item{n}{Number of rows in \code{data}}

\item{B}{Number of bootstrap}
}
\description{
compute the estimates
}
