% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates in each subsample}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{logistic regression formula}

\item{data}{data frame}

\item{n}{numeric}

\item{B}{numeric}
}
\value{
the regression estimate in each subsample
}
\description{
Compute the regression estimate in each subsample
}
