% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, family, data, n, B)
}
\arguments{
\item{formula}{fomula: V1~V2}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{data}{data list}

\item{n}{total number of rows in data}

\item{B}{number of bootstraps}
}
\value{
list
}
\description{
compute the estimates
}
