% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{a formula fitting a linear regression model.}

\item{data}{dataset for the model}

\item{n}{number of rows of dataset}

\item{B}{number of simulations}

\item{family}{family type for models, such as binomial, gaussian, gamma etc.}
}
\description{
compute the estimates
}
