% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on the given number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{dataset}

\item{freqs}{row repetitions}
}
\description{
Linear regression is made based on the formula given and a list of regression estimates are returned
}
