% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{List of coefficients and sigma}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{data split in m ways as a matrix}

\item{y}{response values split in m ways}

\item{n}{size of data}
}
\description{
calls blbcoef and blbsigma and returns a list of both coefficients and sigma values
}
\details{
compute the regression estimates for a blb dataset
}
