% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{Estimate the regression estimates based on given number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}

\item{freqs}{matrix}
}
\value{
list of coefficents and sigma
}
\description{
Estimate the regression estimates based on given number of repetitions
}
