% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{compute the regression estimates for a blb dataset}
\usage{
lm1(X, y, n, seed = NaN)
}
\arguments{
\item{X}{Model Matrix for Regression}

\item{y}{Model Response for Regression}

\item{n}{Number of Rows in Data}

\item{seed}{Optional seed for testing}
}
\description{
compute the regression estimates for a blb dataset
}
