% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data, just data, probably a vector}

\item{freqs}{weights of model fit}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
