% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{compute the regression estimates for a blb dataset}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{to be filled}

\item{y}{to be filled}

\item{n}{to be filled}
}
\value{
to be filled
}
\description{
to be filled
}
