% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs = NULL)
}
\arguments{
\item{formula}{a formula}

\item{data}{dateframe}

\item{freqs}{an optional vector}
}
\value{
linear regression model
}
\description{
estimate the regression estimates based on given the number of repetitions
}
\examples{
lm1(mpg ~ wt * hp , mtcars, NULL)
}
