% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{lm1}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{x values}

\item{y}{y values}

\item{n}{number of samples}
}
\value{
linear regression with weight, more specifically coefficiemt and sigma.
compute the regression estimates for a blb dataset
}
\description{
lm1
}
