% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{formula to fit the bootstraped model}

\item{data}{subset data}

\item{freqs}{frequency}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
