% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{lm1}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset}

\item{freqs}{weights}
}
\value{
returns list of coef and sigma
estimate the regression estimates based on given the number of repetitions
}
\description{
lm1
}
