% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{compute the regression estimates for a blb dataset}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{independent variables}

\item{y}{dependent varibales}

\item{n}{number of rows of the data}
}
\description{
compute the regression estimates for a blb dataset
}
