% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{Computing Regression Estimates}
\usage{
lm1(X, y, n, fast_option = FALSE)
}
\arguments{
\item{X}{independent variables of the data.}

\item{y}{dependent variables of the data.}

\item{n}{number of rows of the data.}

\item{fast_option}{optional to use Rcpp for faster computation. By default, \code{fast_option = FALSE}.}
}
\value{
a list of coefficients and sigma for the blb dataset.
}
\description{
Compute the regression estimates for a blb dataset.
}
