% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm1.R
\name{lm1}
\alias{lm1}
\title{compute the linear regression estimates for a blb dataset}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{explanatory variables}

\item{y}{response variable}

\item{n}{number of rows}
}
\value{
list consists of coefficient and sigma
}
\description{
compute the linear regression estimates for a blb dataset
}
