% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{compute the regression estimates for a blb dataset}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{Terms}

\item{y}{Response}

\item{n}{Number of rows in data}
}
\description{
compute the regression estimates for a blb dataset
}
