% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{compute the regression estimates for a blb dataset}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{A symbolic description of the model}

\item{y}{dataframe}

\item{n}{freqency of each observations}
}
\value{
list, coefficients and sigma
}
\description{
compute the regression estimates for a blb dataset
}
