% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm2}
\alias{lm2}
\title{Rcpp version of lm1}
\usage{
lm2(X, y, n)
}
\arguments{
\item{X}{Least Squares Matrix}

\item{y}{Least Squares target Matrix}

\item{n}{length of dataset}
}
\description{
Rcpp version of lm1
}
