% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{dataset}

\item{n}{size of original dataset}
}
\description{
Freqency of data row repetitions are generated using multinomial distribution, which is then used
in linear regression
}
