% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the linear regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data, just data}

\item{n}{number of random vectors}
}
\description{
compute the linear regression estimates for a blb dataset
}
