% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{Linear Regression on each bootstrap
compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{model that was specified}

\item{data}{dataset to perform lm on}

\item{n}{number of observations}
}
\description{
Linear Regression on each bootstrap
compute the regression estimates for a blb dataset
}
