% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{a formula}

\item{data}{dateframe}

\item{n}{integer}
}
\value{
dataframes
}
\description{
compute the regression estimates for a blb dataset
}
\examples{
lm_each_boot(mpg ~ wt * hp , mtcars, 3, 100, 3)
}
