% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{To fit the linear regression model}

\item{data}{The dataset to perform formula}

\item{n}{number of observations}
}
\description{
compute the regression estimates for a blb dataset
}
