% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{lm_each_boot}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{freqs}{list of numbers
compute the regression estimates for a blb dataset}
}
\description{
lm_each_boot
}
