% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{linear regression model}

\item{data}{input data}

\item{n}{number of rows in input data}
}
\description{
compute the regression estimates for a blb dataset
}
