% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset given a generated weight}
\usage{
lm_each_boot(formula, data, n, method)
}
\arguments{
\item{formula}{the formula for linear regression}

\item{data}{the subsample data}

\item{n}{the row number of original data}

\item{method}{the method to compute coef and sigma, choice: "lm","lmR","lmC", default as "lm"}
}
\value{
the estimate of coefficients and sigma in one weighted linear regression
}
\description{
compute the regression estimates for a blb dataset given a generated weight
}
