% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{fomula: V1~V2}

\item{data}{dataframe}

\item{n}{total number of rows in data}
}
\value{
model
}
\description{
compute the regression estimates for a blb dataset
}
