% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{for linear regression}

\item{data}{dataset}

\item{n}{size of original dataset}

\item{B}{bootstrap sample size}
}
\description{
Subsamples are resampled to size of original data B times, and linear regression is applied
to each of them
}
