% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{BootStrap estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{formula of given fit}

\item{data}{data as a dataframe}

\item{n}{size of data}

\item{B}{number of times to be bootstrapped}
}
\description{
Computes the estimates B number of times and returns the results
}
\details{
compute the estimates
}
