% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Compute the Estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a formula}

\item{data}{a dataframe}

\item{n}{numeric}

\item{B}{numeric}
}
\description{
Compute the Estimates
}
