% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Compute the estimates by repeating lm_each_boot B times for a given subsample.}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}

\item{n}{numeric}

\item{B}{numeric}
}
\value{
list of coefficents and sigma
}
\description{
Compute the estimates by repeating lm_each_boot B times for a given subsample.
}
