% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute bootstrap estimates in sub-samples.}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a data frame containing the variables in the model.}

\item{data}{a data frame (after from sub-sampled from original data) containing the variables in the model.}

\item{n}{an integer specifying the number of rows from original data.}

\item{B}{an integer specifying the number of bootstraps done within each sub-sample.}
}
\value{
a list of length m of sub-sampled data
}
\description{
compute the bootstrapped estimates of linear model parameters (regression coefficients \eqn{\beta} and common standard deviation \eqn{\sigma} in error terms).
}
