% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Linear regression on each sub sample
compute the estimates for each subsample}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{linear regression model you want}

\item{data}{dataset you want to perform formula on}

\item{n}{number of observations}

\item{B}{times to run bootstrap}
}
\description{
Linear regression on each sub sample
compute the estimates for each subsample
}
