% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a lm formula}

\item{data}{a data frame containing the features mentioned in formula}

\item{n}{the number of rows in the dataset}

\item{B}{the number of bootstrap samples to perform}
}
\description{
compute the estimates
}
