% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{To fit the linear regression model}

\item{data}{The dataset to perform formula}

\item{n}{number of observations}

\item{B}{number of bootstraps}
}
\description{
compute the estimates
}
