% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{the model you would like to create}

\item{data}{the data you will use to pass into the function}

\item{n}{number of rows of the data}

\item{B}{number of bootstraps on each m split}
}
\description{
compute the estimates
}
