% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Computing Estimates}
\usage{
lm_each_subsample(formula, data, n, B, fast_option)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a set of data.}

\item{n}{number of rows of the data.}

\item{B}{number of bootstraps.}

\item{fast_option}{logical. \code{TRUE} if using Rcpp code in \code{lm1()}. By default, \code{fast_option = FALSE}.}
}
\value{
a list of estimates for the subsample.
}
\description{
Compute the estimates for each subsample.
}
