% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{linear regression model}

\item{data}{input data}

\item{n}{size of each blb dataset}

\item{B}{number of bootstraps}
}
\description{
Use bootstrap on each subsample to compute its coefficients
}
