% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Linear Model Subsample (C++)}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{Linear regression formula}

\item{data}{Chosen dataset}

\item{n}{Number of rows in the dataset}

\item{B}{Number of bootstrap samples}
}
\description{
Computes estimates
}
