% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{Regression formula}

\item{data}{Input data frames.}

\item{n}{Total number of data points}

\item{B}{Number of bootstraps to be performed.}
}
\description{
Calls the function lm_each_boot B times, with B corresponding to the number of bootstraps.
}
