% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{the linear regression model}

\item{data}{data frame}

\item{n}{numeric, length of the data frame}

\item{B}{numeric, repeat the process B times}
}
\description{
compute the estimates
}
