% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{linear regression formula}

\item{data}{data frame}

\item{n}{numeric}

\item{B}{numeric}
}
\value{
the regression estimate in each subsample
}
\description{
compute the estimates
}
