% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Linear regression modeling based on one subsample}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{A symbolic description of the model}

\item{data}{dataframe}

\item{n}{sample size}

\item{B}{integer, bootstrap iteration times}
}
\description{
Linear regression modeling based on one subsample
}
