% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{formula for regression}

\item{data}{dataset of interest}

\item{n}{number of rows in total}

\item{B}{number of bootstrap samples}
}
\description{
compute the estimates
}
