% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{fomula: V1~V2}

\item{data}{dataframe}

\item{n}{total number of rows in data}

\item{B}{number of bootstraps}
}
\value{
list
}
\description{
compute the estimates
}
