% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lmC}
\alias{lmC}
\title{Linear Model (in C++)}
\usage{
lmC(formula, data, freqs)
}
\arguments{
\item{formula}{Linear regression formula}

\item{data}{Chosen dataset}

\item{freqs}{Weights computed from lm_each_boot}
}
\description{
Estimate the regression estimates using a lm function written in C++
}
