% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{logreg1}
\alias{logreg1}
\title{logistic regression coefficient and sigma}
\usage{
logreg1(data, formula, freqs)
}
\arguments{
\item{data}{data}

\item{formula}{formula}

\item{freqs}{frequency}
}
\value{
logistic regression
compute the regression estimates for a blb dataset
}
\description{
logistic regression coefficient and sigma
}
