% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{logreg_each_subsample}
\alias{logreg_each_subsample}
\title{logistic regression of each subsample}
\usage{
logreg_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{the formula of choice - logistic regression with weight in this case.}

\item{data}{data of choice}

\item{n}{a positive number, the number of samples.}

\item{B}{B a posive integer. It represents a number of bootstrap samples.}
}
\value{
linear regression of each subsample with bootstrap
compute the estimates
}
\description{
logistic regression of each subsample
}
