% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_cbind}
\alias{map_cbind}
\title{make a list as a form of column bind}
\usage{
map_cbind(.x, .f, ...)
}
\arguments{
\item{.x}{list of data}

\item{.f}{function}

\item{...}{other arguments}
}
\value{
the column bind of the list
}
\description{
make a list as a form of column bind
}
