% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_mean}
\alias{map_mean}
\title{Get the mean}
\usage{
map_mean(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function,formula or vector}

\item{...}{further arguments passed from or to other methods}
}
\description{
Get the mean
}
