% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_mean}
\alias{map_mean}
\title{get the mean of the estimates in subsamples by column}
\usage{
map_mean(.x, .f, ...)
}
\arguments{
\item{.x}{lits of data}

\item{.f}{function}

\item{...}{other arguments}
}
\value{
the mean of the list for each element
}
\description{
get the mean of the estimates in subsamples by column
}
