% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_rbind}
\alias{map_rbind}
\title{Rowbind}
\usage{
map_rbind(.x, .f, ...)
}
\arguments{
\item{.x}{the dataset}

\item{.f}{A function,formula or vector}

\item{...}{further arguments passed from or to other methods}
}
\description{
Rowbind
}
