% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{map_rbind}
\alias{map_rbind}
\title{make a list as a form of row bind}
\usage{
map_rbind(.x, .f, ...)
}
\arguments{
\item{.x}{list of data}

\item{.f}{function}

\item{...}{other arguments}
}
\value{
the row bind of the list
}
\description{
make a list as a form of row bind
}
