% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{mean_lwr_upr}
\alias{mean_lwr_upr}
\title{Calculates the mean of an object and the lower and upper bounds for it}
\usage{
mean_lwr_upr(x, level = 0.95)
}
\arguments{
\item{x}{numeric Vector of estimates.}

\item{level}{numeric Confidence level.}
}
\value{
numeric estimate with bounds
}
\description{
Calculates the mean of an object and the lower and upper bounds for it
}
