% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{mean_lwr_upr}
\alias{mean_lwr_upr}
\title{Get the confidence interval of mean}
\usage{
mean_lwr_upr(x, level = 0.95)
}
\arguments{
\item{x}{The dataset}

\item{level}{The alpha value}
}
\description{
Get the confidence interval of mean
}
