% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{mean_lwr_upr}
\alias{mean_lwr_upr}
\title{get the mean as well as quantile from a data}
\usage{
mean_lwr_upr(x, level = 0.95)
}
\arguments{
\item{x}{list of data}

\item{level}{the confidence level of confidence interval}
}
\description{
get the mean as well as quantile from a data
}
