% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blblm}
\alias{par_blblm}
\title{Computes the blblm results using parallization with user-specified number of cores}
\usage{
par_blblm(formula, data, m = 10, B = 5000, cl)
}
\arguments{
\item{formula}{an lm formula}

\item{data}{the dataset containing the features mentioned in formula}

\item{m}{the number of subsamples the data is to be split into}

\item{B}{the number of bootstrap samples to take}

\item{cl}{this is equvalent to makeCluster(). This is created by the user outside the funtion.}
}
\value{
list
}
\description{
Computes the blblm results using parallization with user-specified number of cores
}
\examples{
\dontrun{
cl<-makeCluster(4)
par_blblm(mpg~hp,mtcars,10,100,cl)
stopCluster(cl)
}
}
