% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Predictions}
\usage{
\method{predict}{blbglm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired}

\item{new_data}{a dataframe or character vector}

\item{confidence}{logical}

\item{level}{double the confident level}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
a vector or matrix or list of your predictions from your model
}
\description{
This function gets the predicitions for the model
}
