% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Predict values from a blbglm object}
\usage{
\method{predict}{blbglm}(
  object,
  new_data,
  confidence = FALSE,
  level = 0.95,
  inv_link = NULL,
  ...
)
}
\arguments{
\item{object}{The object to calculate prediction values for.}

\item{new_data}{The values to predict from.}

\item{confidence}{Whether to provide a confidence interval.}

\item{level}{The level to use for the confidence interval.}

\item{inv_link}{The inverse of the link function, where needed. Logit link functions are detected automatically.}

\item{...}{Additional parameters to pass.}
}
\description{
Predict values from a blbglm object
}
