% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blglml.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{predict bootstrap from fit}
\usage{
\method{predict}{blbglm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{new_data}{matrics/vector}

\item{confidence}{boolean vector}

\item{level}{confidential level}

\item{...}{est}

\item{fit}{model}
}
\value{
tibble
}
\description{
predict bootstrap from fit
}
