% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{predict.blbglm}
\usage{
\method{predict}{blbglm}(object, new_data, confidence = FALSE, level = 0.95, nthreads = 1, ...)
}
\arguments{
\item{object}{blbglm}

\item{new_data}{dataframe of new data entries}

\item{confidence}{boolean return confidence interval}

\item{level}{double level of confidence interval}

\item{nthreads}{an integer denoting number of workers}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predict with new observation. Can return prediction or confidence interval.
}
