% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Obtains predictions and optionally estimates standard errors of those predictions from a fitted generalized linear model object.}
\usage{
\method{predict}{blbglm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blbglm}

\item{new_data}{dataframe}

\item{confidence}{logical}

\item{level}{numeric}

\item{...}{arg}
}
\description{
Obtains predictions and optionally estimates standard errors of those predictions from a fitted generalized linear model object.
}
