% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Predicts new dataset for blbglm}
\usage{
\method{predict}{blbglm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the blbglm model to use}

\item{new_data}{the dataset representing \code{X} to predict \code{y} for}

\item{confidence}{whether or not to create a confident interval, defaults to \code{FALSE}}

\item{level}{the confidence level to run the confidence interval on, defaults to \code{0.95}}

\item{...}{extra arguments}
}
\description{
overloads the standard \code{predict} functionality to get predictions for a \code{blbglm} model given an X value.
}
\examples{
blbglm_model <- blbglm(am ~ wt * hp, data = mtcars, m = 3, B = 100)
predict(blbglm_model, data.frame(wt = c(2.5, 3), hp = c(150, 170)))
predict(blbglm_model, data.frame(wt = c(2.5, 3), hp = c(150, 170)), confidence = TRUE)
}
