% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Computing the predictions of new data of blbglm model}
\usage{
\method{predict}{blbglm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the fit of blbglm}

\item{new_data}{data frame}

\item{confidence}{users can choose whether to have a confidence interval or not}

\item{level}{the level of confidence}

\item{...}{Other arguments}
}
\description{
Computing the predictions of new data of blbglm model
}
