% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Make a prediction}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A variable with the class "blblm"}

\item{new_data}{New values for the predictor variables to use in estimating the response}

\item{confidence}{Boolean for whether a confidence interval should be returned}

\item{level}{The significance level for the confidence itnerval (1 - alpha)}

\item{...}{Included for method flexibility}
}
\value{
A single number or vector of estimates for the predictor variable (potentially as a confidence interval)
}
\description{
Use the model and new data to produce estimates for the response variable
}
