% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Prediction from blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm result}

\item{new_data}{data frame of predictor values}

\item{confidence}{bool confidence interval}

\item{level}{confidence level in decimal}

\item{...}{other parameters}
}
\description{
Linear regression prediction from blblm function given predictor values. Confidence interval for
prediction can be specified.
}
