% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Model predictions for fitted models generated by blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm object, the fitted model}

\item{new_data}{a data.frame, environment, or list}

\item{confidence}{boolean, specify whether to compute the confidence interval}

\item{level}{double, the significance level 1-alpha, 0.95 by default}

\item{...}{other conditions}
}
\value{
return the predicted value of the object (the fitted model)
}
\description{
Acquire the predicted value by providing new data and the fitted model that generated by blblm.
}
\examples{
Can use either predict.blblm or predict

predict.blblm(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100), data.frame(wt = c(2.5, 3), hp = c(150, 170)))
>        1        2
> 21.55538 18.80785
predict.blblm(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100), data.frame(wt = c(2.5, 3), hp = c(150, 170)), confidence = TRUE)
>        fit      lwr      upr
> 1 21.55538 20.02457 22.48764
> 2 18.80785 17.50654 19.71772
}
