% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predicting on blblm model}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A fit blblm model}

\item{new_data}{New data set to be predicted on using model obtained from object}

\item{confidence}{If set to TRUE, confidence interval of prediction is returned}

\item{level}{If confidence is set to TRUE, level can be adjusted for wanted confidence interval}

\item{...}{Argument that is not predefined}
}
\description{
Given a fit blblm model, this function predicts on new data using the already fit model.
}
\examples{
fit = blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 10000)
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)), confidence = TRUE)
}
