% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict a response for new data points using model formula.
Also calculate confidence intervale for these response estimates if specified.}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm fitted model}

\item{new_data}{numeric New data points to predict.}

\item{confidence}{logical Specify whether or not you would like a confidence interval.}

\item{level}{numeric Confidence level.}

\item{...}{further arguments}
}
\value{
numeric Predicted estimates with or without a confidence interval.
}
\description{
Predict a response for new data points using model formula.
Also calculate confidence intervale for these response estimates if specified.
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
predict(fit, data.frame(wt = c(7, 5.5), hp = c(130, 190)), confidence = TRUE, level = 0.95)
}
