% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A blblm object}

\item{new_data}{A set of new data corresponding to the predictor variables of the formula (if formula is y ~ x, data is x = ...)}

\item{confidence}{Boolean, If TRUE will generate a confidence interval at the specified level, default FALSE}

\item{level}{The desired level of confidence for the confidence interval. Default .95}

\item{...}{...}
}
\description{
Predict
}
\details{
Returns the predicted outputs for the values of the new data given. If confidence = TRUE, returns a confidence
interval for the predicted output value at the specified level of confidence.
}
