% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{The method for the blblm class}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Which is a blblm variable.}

\item{new_data}{which is a data set, it is the value we want to predict}

\item{confidence}{Whcih is the alpha level we want to choose}

\item{level}{The level of alpha}

\item{...}{Nothing}
}
\value{
which returns the predictive value or the confidence interval.
}
\description{
For a blblm object input, we get the predictive value and the confident interval for the variable.
}
\examples{
# fit<- blblm(mpg~disp, data = mtcars, m = 10, B = 5000)
newdata<- data.frame(disp = 10)
# predict(fit,newdata)
}
