% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{compute confidence interval from fit}
\usage{
\method{confint}{blbglm}(fit, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{fit}{model}
}
\description{
compute confidence interval from fit
}
