% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{predict}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Model}

\item{new_data}{new studff}

\item{confidence}{TRUE or FALSE}

\item{level}{confidence percentage}

\item{...}{extra conditions}
}
\description{
predict
}
