% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{predict.blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm}

\item{new_data}{dataframe of new data entries}

\item{confidence}{boolean return confidence interval}

\item{level}{double level of confidence interval}
}
\description{
predict.blblm
}
