% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Obtain the predictions/prediction intervals}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{object of class "blblm"}

\item{new_data}{object}

\item{confidence}{a logical value, FALSE by default, indicating whether or not user wants the prediction interval in addition to the predicted value}

\item{level}{numeric specify the level of the prediction interval}
}
\value{
numeric
}
\description{
We use this function \code{predict.blblm} in order to find coefficients for all variables involved. This is comparable to using \code{coef}
}
\examples{
fit = blblm(mpg ~ hp * wt, data = mtcars, m = 3, B = 100)
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)))
}
