% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model from blblm}

\item{new_data}{a data.frame with the values for the new data}

\item{confidence}{TRUE for a confidence interval of predicted value}

\item{level}{the level of confidence to use, default is .95}

\item{...}{additional parameters}
}
\value{
vector of predicted values and upper,lower bounds if confidence is TRUE
}
\description{
calculates a predicted value based on new data
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)))
}
