% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Prediction Interval for Model Parameters
Make prediction intervals for model paramters}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a model object for which prediction is wanted}

\item{new_data}{a data frame or character vector}

\item{confidence}{logical}

\item{level}{numeric, the prediction level}

\item{...}{additonal arguments that impact the predictions}
}
\value{
A matrix or vector with columns, depends on the class of the argument
}
\description{
Prediction Interval for Model Parameters
Make prediction intervals for model paramters
}
