% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Based on the previous dataset to predict the new data}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{estimated object}

\item{new_data}{the new data wanted to predict}

\item{confidence}{confidence interval}

\item{level}{alpha value}

\item{...}{further arguments passed from or to other methods}
}
\description{
Based on the previous dataset to predict the new data
}
