% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{prediction using blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the variable or parameter}

\item{new_data}{data of interest}

\item{confidence}{pick if they want the confidence level}

\item{level}{the percentage of confidence level}

\item{...}{extra arguments that is not related to this function}
}
\description{
it predicts new data with the use of blblm
}
