% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{gives estimated predictor value of linear regression model}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{linear regression model}

\item{new_data}{new dataset}

\item{confidence}{choice to apply confidence interval calculations}

\item{level}{confidence level}

\item{...}{add whatever parameters you want}
}
\description{
gives estimated predictor value of linear regression model
}
