% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Model Prediction}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm model.}

\item{new_data}{a set of new data used for prediction}

\item{confidence}{logicals. If \code{TRUE} the confidence interval will be calculated. By default, \code{confidence = FALSE}.}

\item{level}{the significance level for the confidence interval.}

\item{...}{additional arguments to be passed.}
}
\value{
a numeric of vector for predicted value (and its confidence interval if applicable).
}
\description{
Predict from the results of various blblm model fitting functions for new data.
}
