% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict outcome with new data}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model}

\item{new_data}{data matrix}

\item{confidence}{boolean}

\item{level}{level of confidence}

\item{...}{further arguments passed to or from other methods}
}
\description{
Compute response variable and its predicted confidence interval given a set of new data
}
