% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predicting blblm
Outputs predictions based on coefficient estimates of blblm object}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{new_data}{A data frame object that contains the new data you would like to predict on}

\item{confidence}{Boolean indicating whether you want confidence intervals or not. By default it is False}

\item{level}{Significance level of your confidence intervals (if applicable)}

\item{fit}{A blblm object}
}
\description{
Predicting blblm
Outputs predictions based on coefficient estimates of blblm object
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)))
}
