% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Get the predicted value of new observation using blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{new_data}{new observation}

\item{confidence}{boolean value to specify whether to use confidence interval}

\item{level}{significance level}

\item{...}{left blank}
}
\value{
predicted value of new observation using blblm
}
\description{
Get the predicted value of new observation using blblm
}
