% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Creating Predict Interval}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the blblm interval}

\item{new_data}{part of the data}

\item{confidence}{of the interval}

\item{level}{level of confidence}

\item{...}{continuation}
}
\description{
This function computes the boothstrap Prediction for a linear model
}
