% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Prediction Function}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{new_data}{data frame}

\item{confidence}{logical value indicating whether we construct the confidence interval}

\item{level}{integer between 0 and 1 indicating the confidence level}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
A matrix, vector, or list with the fit, upper and lower value
}
\description{
Computes the predictions and the confidence interval for the predictions
}
