% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{LM Prediction}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The final estimated fitted linear regression model.}

\item{new_data}{New data to predict on.}

\item{confidence}{Unknown.}

\item{level}{The \% confidence interval for this prediction,}

\item{...}{Other arguments.}
}
\description{
Predicts the value of new data based on the coefficients estimated by blblm.
}
