% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict New Values with blblm Model}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted blblm model.}

\item{new_data}{New data set to be used in prediction.}

\item{confidence}{Single logical indicating whether the output should contain the confidence interval.}

\item{level}{Confidence level.}

\item{...}{Additional arguments.}
}
\value{
New, predicted values based on given new data.
}
\description{
Predicts new values given new data.
}
