% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{blblm model prediction}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{model fit produced by blblm(formula, data, ...)}

\item{new_data}{data frame with same column dimensions as in object.}

\item{confidence}{logical, TRUE gives a confidence interval, FALSE gives a single estimate.}

\item{level}{numeric, confidence interval level (only if confidence = TRUE).}

\item{...}{}
}
\description{
Makes a prediction using the fitted model in object on new data with similar dimensions
}
