% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{predict.blblm is a function for predictions from blblm model for new data}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm model}

\item{new_data}{an data frame, list or env containing new data to be used for the prediction}

\item{confidence}{logical. If TRUE gives confidence interval of new prediction on blblm linear model/}

\item{level}{an alpha level for the confidence interval}

\item{...}{additional arguments to be passed}
}
\description{
predict.blblm is a function for predictions from blblm model for new data
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)))
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)), confidence = TRUE)

}
