% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Computing the predictions of new data of blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the fit of blblm}

\item{new_data}{data frame}

\item{confidence}{users can choose whether to have a confidence interval or not}

\item{level}{the level of confidence}

\item{...}{Other arguments}
}
\description{
Computing the predictions of new data of blblm
}
