% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Prediction for new data with blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The result from \code{blblm}}

\item{new_data}{New observation that want to predict}

\item{confidence}{Determine if the confience interval is needed. (logical, default is F)}

\item{level}{The significant value. (default is 0.95)}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The prediction of new data using \code{blblm}
}
\description{
Prediction for new data with blblm
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)))
predict(fit, data.frame(wt = c(2.5, 3), hp = c(150, 170)), confidence = TRUE)
}
