% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{calculate the prediction from blblm}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm object}

\item{new_data}{a new dataframe for prediction}

\item{confidence}{whether do confidence interval}

\item{level}{confidence level}
}
\description{
calculate the prediction from blblm
}
