% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{prediciton}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, nthreads = 1)
}
\arguments{
\item{object}{an object of "blblm"}

\item{new_data}{new data frame}

\item{confidence}{need confidence interval or not}

\item{level}{confidence level, default is 0.95}

\item{nthreads}{the number of cores you want to use when processing#' @param nthreads the number of cores you want to use when processing}
}
\value{
prediction result
}
\description{
prediciton
}
