% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Prediction using the blblm model}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm model}

\item{new_data}{data used to predict}

\item{confidence}{give a confidene interval or not}

\item{level}{confidence level}

\item{...}{other parameters}
}
\description{
Prediction using the blblm model
}
