% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{predict.blblogreg}
\alias{predict.blblogreg}
\title{Model predictions for logistic models generated by blblogreg}
\usage{
\method{predict}{blblogreg}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblogreg object, the logistic model}

\item{new_data}{a data.frame, environment, or list}

\item{confidence}{boolean, specify whether to compute the confidence interval}

\item{level}{double, the significance level 1-alpha, 0.95 by default}

\item{...}{other conditions}
}
\value{
return the predicted value of the object (the logistic model)
}
\description{
Acquire the predicted value by providing new data and the logistic model that generated by blblogreg
}
\examples{
> predict(blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = Smarket, m = 3, B = 100, parallel = TRUE), new_data = data.frame("Lag1" = c(0.382, 0.4, 0.5),"Lag2" = c(1.03, 1.4, 0.5), "Volume" = c(1.2, 1.34, 1.1)))
1           2           3
-0.02709720 -0.03393728 -0.02105841
}
