% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblogreg}
\alias{predict.blblogreg}
\title{Predict blblogreg}
\usage{
\method{predict}{blblogreg}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblogreg object.}

\item{new_data}{data frame.}

\item{confidence}{logical. Specifies whether to use confidence.}

\item{level}{factor. level of the confidence. Set at 0.95.}

\item{...}{R objects.}
}
\value{
object of class "blblogreg"
}
\description{
Predict blblogreg
}
