% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{predict.blblogreg}
\alias{predict.blblogreg}
\title{Prediction for new data with blblogreg}
\usage{
\method{predict}{blblogreg}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The result from \code{blblogreg}}

\item{new_data}{New observation that want to predict}

\item{confidence}{Determine if the confience interval is needed. (logical, default is F)}

\item{level}{The significant value. (default is 0.95)}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The prediction of new data using \code{blblogreg}
}
\description{
Prediction for new data with blblogreg
}
\examples{
fit <- blblogreg(Species ~ Sepal.Length * Petal.Length, binomial, iris[1:100,], 3, 100, FALSE)
predict(fit, data.frame(Sepal.Length = c(5.0, 5.1), Petal.Length = c(1.3, 1.4)))
predict(fit, data.frame(Sepal.Length = c(5.0, 5.1), Petal.Length = c(1.3, 1.4)), confidence = TRUE)
}
