% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm_logistic.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Print Values}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
a value
}
\description{
print.blbglm prints an argument
}
