% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Print out a blbglm object}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Additional parameters to pass.}
}
\description{
Print out a blbglm object
}
