% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{print.blbglm}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{blbglm}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints formula of blbglm model
}
