% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Print Blbglm Values
Print values returned by blbglm model.}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional arguments}
}
\description{
Print Blbglm Values
Print values returned by blbglm model.
}
