% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{prints blbglm model}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{the item to be printed}

\item{...}{additional parameters to be passed in}
}
\description{
prints blbglm model
}
