% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Printing Regression Model}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{a blbglm model.}

\item{...}{additional arguments to be passed.}
}
\value{
formula.
}
\description{
Print the formula for the blbglm model.
}
