% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Displays blbglm model formula}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{blbglm}

\item{...}{arg  arg}
}
\description{
Displays blbglm model formula
}
