% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{print.blbglm}
\alias{print.blbglm}
\title{Prints blbglm}
\usage{
\method{print}{blbglm}(x, ...)
}
\arguments{
\item{x}{the model to be printed}

\item{...}{extra arguments}
}
\description{
overloads the \code{print} functionality to show the formula used when printing a \code{blbglm} model.
}
\examples{
blbglm_model <- blbglm(am ~ wt * hp, data = mtcars, m = 3, B = 100)
print(blbglm_model)
}
