% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the model formula}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{An object with class "blblm"}

\item{...}{Included for method flexibility}
}
\description{
The formula will be printed for a "blblm" variable
}
