% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the blblm model (formula) of the fitting model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{a blblm object}

\item{...}{other conditions}
}
\value{
return the blblm model (formula) of x
}
\description{
Print the blblm model (formula) of the fitting model
}
\examples{
print(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
}
