% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Blblm Model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{A fit blblm Model}

\item{...}{Argument that is not predefined}
}
\description{
Prints the Model formula used to fit your data using the blblm function .
}
\examples{
fit = blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
print(fit)
}
