% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Used to print blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{blblm function}

\item{...}{}
}
\value{
Outputs the concatenated object of class "blblm"
}
\description{
Used to print blblm
}
