% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the fitted blblm formula}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{blblm fitted model}

\item{...}{further arguments}
}
\value{
character The Model Formula.
}
\description{
Print the fitted blblm formula
}
\examples{
print(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
}
