% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print Blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{blblm object}

\item{...}{...}
}
\description{
Print Blblm
}
\details{
Called via the standard print() in R, it prints the formula used by the blblm object.
}
