% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{The method for the blblm class}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{Which is a blblm variable.}

\item{...}{Nothing}
}
\value{
char return the formula from the blblm variable
}
\description{
For a blblm object input, we get the print
}
\examples{
# fit<- blblm(mpg~., data = mtcars, m = 10, B = 5000)
# print(fit)
}
