% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{print.blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{blblm}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints formula of blblm model
}
