% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the BLB linear regression model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{a blblm object}

\item{...}{additional arguments to be passed}
}
\description{
print out the model formula and regression coefficients of a blblm object to the console
}
