% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{to be filled}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{to be filled}

\item{...}{to be filled}
}
\value{
to be filled
}
\description{
to be filled
}
