% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print out model and the respective formula}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{object of class "blblm"}
}
\value{
This returns the model and the formula used to fit it.
}
\description{
Use this to print out the formula.
}
