% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Used to return coefficient estimates of our blb linear regression.}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{number of observations}

\item{...}{any other argument you might want to pass through here.}
}
\description{
Used to return coefficient estimates of our blb linear regression.
}
