% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{prints the model used in blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{a blblm results}

\item{...}{additional parameters}
}
\description{
prints the model used in blblm
}
