% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print Blblm Values
Print values returned by blblm model.}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional arguments}
}
\description{
Print Blblm Values
Print values returned by blblm model.
}
