% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{print blblm model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{blblm model}

\item{...}{extra arguments other than this function}
}
\description{
print blblm model
}
