% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Printing the model}
\usage{
\method{print}{blblm}(x, ...)

\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{the fit}

\item{...}{additional arguments to be passed to the low level regression fitting functions}
}
\value{
A string of the blbglm model with regression formula.
print(fit)blbglm model: mpg ~ wt * hp

A string of the blblm model with regression formula.
}
\description{
\code{print.blblm} is used to print out 'blbglm model:' followed by the logistic regression formula

\code{print.blblm} is used to print out 'blblm model:' followed by the regression formula
}
\examples{
fit <- blblm_mod(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
print(fit)
# blblm model: mpg ~ wt * hp
}
