% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{print results of blblm function}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{number}

\item{...}{add whatever you want}
}
\description{
print results of blblm function
}
