% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{return coefficient estimates from blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{number of observations}

\item{...}{additional arguments}
}
\description{
return coefficient estimates from blblm
}
