% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{printblblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{lm bootstrap model taken as a parameter}

\item{...}{more parameters}
}
\description{
printblblm
}
