% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{prints blblm model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{the item to be printed}

\item{...}{additional parameters to be passed in}
}
\description{
prints blblm model
}
