% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Printing Regression Model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{a blblm model.}

\item{...}{additional arguments to be passed.}
}
\value{
formula.
}
\description{
Print the formula for the blblm model.
}
