% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the linear model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{the fitted model}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the linear model
}
