% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Printing blblm Object
Outputs the formula associated with a blblm object}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{Blblm object}
}
\description{
Printing blblm Object
Outputs the formula associated with a blblm object
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
print(fit) # mpg ~ wt * hp
}
