% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the formula of blblm object}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{blblm object}

\item{...}{extra conditions}
}
\value{
formula of blblm object x
}
\description{
Print the formula of blblm object
}
\examples{
print(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
}
