% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the Linear Regression Model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments affecting the result}
}
\value{
formula of blblm
}
\description{
Print the Linear Regression Model
}
