% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{Fitted linear regression model}

\item{...}{Additional parameters}
}
\description{
Prints formula used in blblm
}
