% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Prints Estimated model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{The estimated coefficients, sigma, etc.}

\item{...}{Other arguments}
}
\description{
Prints the estimated model from the average of the B bootstraps.
}
