% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print Bag of Little Bootstraps Linear Model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{Fitted blblm model.}

\item{...}{Additional arguments.}
}
\value{
Formula used to create the given fitted blblm model.
}
\description{
Prints the formula used to create the given fitted blblm model.
}
