% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{model fit produced by blblm(formula, data, ...)}

\item{...}{other arguments.}
}
\description{
Prints the formula from the blblm function
}
