% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{print.blblm is used to print the blblm model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{represents the blblm model or object to be printed}

\item{...}{additional arguments to be passed}
}
\description{
print.blblm is used to print the blblm model
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
print.blblm(fit)

}
