% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the blblm Formula}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{the fit of the linear regression model of blblm}

\item{...}{Other arguments}
}
\description{
Print the blblm Formula
}
