% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the model of blblm}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{The result from \code{blblm}}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The formula for the linear regression in \code{blblm}
}
\description{
Print the model of blblm
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
print(fit)
}
