% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print out the blblm linear model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{a blblm object}
}
\value{
a string: blblm model: formula
}
\description{
Print out the blblm linear model
}
