% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{print.blblog}
\alias{print.blblog}
\title{Print out model and the respective formula}
\usage{
\method{print}{blblog}(x, ...)
}
\arguments{
\item{x}{object of class "blblog"}

\item{...}{optional arguments to be passed}
}
\value{
This returns the model and the formula used to fit it.
}
\description{
Use this to print out the formula.
}
\details{
If the user wants to print out the model and the respective model, using \code{print} will do just that.
}
\examples{
fit <- blblog(Species ~ Sepal.Length, data = iris, m = 3, B = 100)
print(fit)
}
