% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{print.blblog}
\alias{print.blblog}
\title{Print the result}
\usage{
\method{print}{blblog}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{additional arguements/inputs}
}
\value{
the printed result
}
\description{
Print the result
}
