% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{print.blblogreg}
\alias{print.blblogreg}
\title{Print the blblogreg model (formula) of the logistic model}
\usage{
\method{print}{blblogreg}(x, ...)
}
\arguments{
\item{x}{a blblogreg object}

\item{...}{other conditions}
}
\value{
return the blblogreg model (formula) of x
}
\description{
Print the blblogreg model (formula) of the logistic model
}
\examples{
print(blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = ISLR::Smarket, m = 3, B = 100))
}
