% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblogreg}
\alias{print.blblogreg}
\title{Used to print blblogreg}
\usage{
\method{print}{blblogreg}(x, ...)
}
\arguments{
\item{x}{blblogreg function}

\item{...}{}
}
\value{
Outputs the concatenated object of class "blblogreg"
}
\description{
Used to print blblogreg
}
