% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{print.blblogreg}
\alias{print.blblogreg}
\title{Print Bag of Little Bootstraps Logistic Regression Model}
\usage{
\method{print}{blblogreg}(x, ...)
}
\arguments{
\item{x}{Fitted blblogreg model.}

\item{...}{Additional arguments.}
}
\value{
Formula used to create the given fitted blblogreg model.
}
\description{
Prints the formula used to create the given fitted blblogreg model.
}
