% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{print.blblogreg}
\alias{print.blblogreg}
\title{Print the model of blblogreg}
\usage{
\method{print}{blblogreg}(x, ...)
}
\arguments{
\item{x}{The result from \code{blblogreg}}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The formula for the linear regression in \code{blblogreg}
}
\description{
Print the model of blblogreg
}
\examples{
fit <- blblogreg(Species ~ Sepal.Length * Petal.Length, binomial, iris[1:100,], 3, 100, FALSE)
print(fit)
}
