% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{Calculate sigma for a blbglm object}
\usage{
\method{sigma}{blbglm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The object to calculate sigma for.}

\item{confidence}{Whether to provide a confidence interval.}

\item{level}{The level to use for the confidence interval.}

\item{...}{Additional parameters to pass.}
}
\value{
Bag of little bootstraps sigma.
}
\description{
Calculate sigma for a blbglm object
}
