% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blglml.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{compute sigma from fit}
\usage{
\method{sigma}{blbglm}(fit, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{fit}{model}
}
\description{
compute sigma from fit
}
