% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blglml.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{compute sigma from fit}
\usage{
\method{sigma}{blbglm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{model}

\item{confidence}{boolean vector}

\item{level}{confidencial level}

\item{...}{est}
}
\value{
double
}
\description{
compute sigma from fit
}
