% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{Getting Sigma of Fit}
\usage{
\method{sigma}{blbglm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm model.}

\item{confidence}{logicals. If \code{TRUE} the confidence interval will be calculated. By default, \code{confidence = FALSE}.}

\item{level}{the significance level for the confidence interval.}

\item{...}{additional arguments to be passed.}
}
\value{
estimated sigma or a confidence interval for it.
}
\description{
Get the estimates of the standard deviation for the model. Can be used to generate a confidence interval for the sigma.
}
