% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{Computes variance of coefficients with confidence interval at alpha = 1-level}
\usage{
\method{sigma}{blbglm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blbglm}

\item{confidence}{logical}

\item{level}{numeric}

\item{...}{arg}
}
\description{
Computes variance of coefficients with confidence interval at alpha = 1-level
}
