% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Get sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A variable with the class "blblm"}

\item{confidence}{Boolean for whether a confidence interval should be reported}

\item{level}{The significance level for a confidence interval (1 - alpha)}

\item{...}{Included for method flexibility}
}
\value{
A numerical estimate for sigma (or a vector with a confidence interval for sigma)
}
\description{
Get an estimate for the Residual Standard Deviation, with the option of a confidence interval
}
