% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Residual standard deviation from blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm result}

\item{confidence}{bool for confidence interval}

\item{level}{confidence level in decimal}

\item{...}{other parameters}
}
\description{
The residual standard deviation from blblm function is returned. Confidence intervals can also
be specified
}
