% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Sigma for blb fit model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A fit blb Model}

\item{confidence}{If this parameter is set to TRUE, the confidence interval of the sigma value will be returned. Else is set the FALSE, a
single value for the sigma will be returned.}

\item{level}{This is the level of confidence that is used to calculate your confidence interval.When confidence is set to TRUE,
this value can be changed to alter your confidence interval.}

\item{...}{Argument that is not predefined}
}
\description{
This function returns the sigma value of a fit model using the blblm function.
}
\examples{
fit = blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
sigma(fit)
sigma(fit, confidence = TRUE, level = 0.95)
}
