% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Extracts the Value of Sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{an R object}

\item{confidence}{logical}

\item{level}{the confidence level}

\item{...}{additional arguments}
}
\value{
a number
}
\description{
sigma.blblm extracts the sigma from the model
}
