% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Used to find the sigma of blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm function}

\item{confidence}{logical. If confidence = TRUE, we will use alpha in the calculations.}

\item{level}{confidence level. Set at 0.95.}

\item{...}{}
}
\value{
object of class "blblm"
}
\description{
Used to find the sigma of blblm
}
