% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A blblm object}

\item{confidence}{Boolean, TRUE if confidence interval is desired, FALSE by default}

\item{level}{The desired level of confidence for the confidence interval (default .95)}

\item{...}{...}
}
\description{
Sigma
}
\details{
Returns the calculated mean of all estimated sigmas from a given blblm fitted object. If confidence = TRUE,
returns a confidence interval at the specified level for the sigma value
}
