% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{The method for the blblm class}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{x}{Which is a blblm variable.}
}
\value{
sigma which is the sigema of fit
}
\description{
For a blblm object input, we get the sigma
}
\examples{
# fit<- blblm(mpg~., data = mtcars, m = 10, B = 5000)
# sigma(fit)
}
