% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{compute sigma from fit}
\usage{
\method{sigma}{blbglm}(fit, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{fit}{model}
}
\description{
compute sigma from fit
}
