% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Regression Model Sigma: BLB}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Model}

\item{confidence}{True or False}

\item{level}{confidence percentage}

\item{...}{extra stuff}
}
\description{
Regression Model Sigma: BLB
}
