% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{sigma.blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm}

\item{confidence}{boolean return confidence interval or not}

\item{level}{if confidence is TRUE level of confidence interval}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Returns blb prediction of sigma of linear model.
Can return a confidence interval instead.
}
