% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{to be filled}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{to be filled}

\item{confidence}{to be filled}

\item{level}{to be filled}

\item{...}{to be filled}
}
\value{
to be filled
}
\description{
to be filled
}
