% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{BLB estimator of error term standard deviation (\eqn{\sigma})}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm object returned by function \code{blblm}}

\item{confidence}{logical, whether to show confidence interval around the point estimate (Default: FALSE)}

\item{level}{numeric, a number range from (0,1) as the confidence level (Default: 0.95)}

\item{...}{additional arguments to be passed}
}
\value{
A single point estimate of error term standard deviation (\eqn{\sigma}). If \code{confidence=TRUE}, also shows the confidence interval around the point estimation.
}
\description{
Calculate the point estimation of error term standard deviation (\eqn{\sigma}), with the option to show the confidence internal by user specified confidence level.
}
