% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Obtain sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, par = FALSE, ...)
}
\arguments{
\item{object}{object of class "blblm"}

\item{par}{logical FALSE by default. TRUE indicates desire to use parallelization.}
}
\value{
numeric indicating the value for sigma for this model.
}
\description{
We use this function \code{sigma.blblm} in order to find coefficients for all variables involved. This is comparable to using \code{sigma}
}
\examples{
fit = blblm(mpg ~ hp * wt, data = mtcars, m = 3, B = 100)
sigma(fit)
}
