% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{calculates sigma from the blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm result}

\item{confidence}{creates a confidence interval}

\item{level}{the level of confidence}

\item{...}{additional parameters}
}
\description{
calculates sigma from the blblm
}
