% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Obtain the value of the Standard Deviation Sigma
Compute the value of the Standard Deviation Sigma for the Model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{an object model}

\item{confidence}{logical, if TRUE you construct the confidence interval of sigma}

\item{level}{the confidence level}

\item{...}{additional arguments}
}
\value{
A numeric value
}
\description{
Obtain the value of the Standard Deviation Sigma
Compute the value of the Standard Deviation Sigma for the Model
}
