% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{sigmablblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ncpu = 1, ...)
}
\arguments{
\item{object}{object}

\item{confidence}{whether or not the result shows upper and lower limits at 0.95 confidence level. Default is FALSE.}

\item{level}{this is a confidence level, default at 0.96=5.}

\item{ncpu}{positive interger, number of CPU. Default to be 1 which represents one CPU.}
}
\value{

}
\description{
sigmablblm
}
