% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{returns bootstrapped estimates of sigma (variance) with optional confidence interval. Parallel computing is also optional}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, para = FALSE, core = 4, ...)
}
\arguments{
\item{object}{linear model}

\item{confidence}{logical indicating confidence interval}

\item{level}{numeric indicating level of confidence}

\item{para}{logical indicating parallel computing}

\item{core}{integer indicating number of cores for parallel computing}

\item{...}{so on}
}
\description{
returns bootstrapped estimates of sigma (variance) with optional confidence interval. Parallel computing is also optional
}
