% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{compute the condifence interval for sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the variable or parameter}

\item{confidence}{if user wants confidence interval}

\item{level}{level of confidence interval}

\item{...}{extra arguments other than this function}
}
\description{
compute the condifence interval for sigma
}
