% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Obtaining sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)

\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the fit}

\item{confidence}{TRUE/FALSE for confidence interval for sigma}

\item{level}{significance level for confidence interval}

\item{...}{additional arguments to be passed to the low level regression fitting functions}
}
\value{
Returns the sigma value and/or confidence interval of the fit.

Returns the sigma value and/or confidence interval of the fit.
}
\description{
Obtaining sigma

Obtaining sigma
}
\examples{
fit <- blblm_mod(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
sigma(fit)
sigma(fit, confidence = TRUE, level = .95)
}
