% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Finds standard deviation or regression model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{regression model}

\item{confidence}{apply confidence interval}

\item{level}{confidence level}

\item{...}{Add whatever parameters you want}
}
\description{
Finds standard deviation or regression model
}
