% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{return sigma estimate from blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the object we want to estimate}

\item{confidence}{set to true for confidence intervals, set to just estimate by default.}

\item{level}{significance level for confidence interval}

\item{...}{additional arguments}
}
\description{
return sigma estimate from blblm
}
