% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{sigmablblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{lm bootstrap model taken as a parameter}

\item{confidence}{specify weather CI should be reurned if so confidence = TRUE}

\item{level}{insert level to compute CI}

\item{...}{more parameters}
}
\description{
sigmablblm
}
