% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{return sigma value of the models}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the model returned from the main blblm function}

\item{confidence}{boolean to include confidence interval}

\item{level}{confidence level between 0 and 1}

\item{...}{additional parameters to be passed in}
}
\value{
a value of sigma after putting everything back from the bootstraps
}
\description{
return sigma value of the models
}
