% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Print the sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the fitted model}

\item{confidence}{boolean}

\item{level}{level of confidence}

\item{...}{further arguments passed to or from other methods}
}
\description{
Compute the sigma and its confidence interval if confidence = TRUE
}
