% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Standard deviation/confidence intervals of blblm model
Returns standard deviation estimate and confidence interval estimate based on given significance level (if applicable)}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Blblm object you would like to read (linear regression)}

\item{confidence}{Boolean indicating whether you want just standard deviation (False) or also the lower + upper confidence intervals}

\item{level}{If confidence is True, this level indicates the significance level of your confidence intervals}
}
\description{
Standard deviation/confidence intervals of blblm model
Returns standard deviation estimate and confidence interval estimate based on given significance level (if applicable)
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
sigma(fit)
}
