% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{compute sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm model}

\item{confidence}{logic value}

\item{level}{confidence level}

\item{...}{...}
}
\value{
double or a vector of interval
}
\description{
computer sigma from  estimated values
}
