% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Calculate the sigma value from fit}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{fit object}

\item{confidence}{boolean value to specify whether to have confidence interval}

\item{level}{significance level}

\item{...}{extra conditions}
}
\value{
sigma value of object
}
\description{
Calculate the sigma value from fit
}
\examples{
sigma(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100))
sigma(blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100), confidence = TRUE)
}
