% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Creating the Signa}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the blblm interval}

\item{confidence}{false}

\item{level}{for confidence}

\item{...}{continuation}
}
\description{
This function computes the boothstrap coefficient
}
