% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Computing Sigma for Blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{confidence}{logical value indicating whether we construct the confidence interval}

\item{level}{integer between 0 and 1 indicating the confidence level}

\item{...}{additional arguments affecting the result}
}
\value{
numeric
}
\description{
Computes the standard deviation/errors using Polymorphism
}
