% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Sigma blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted linear regression model}

\item{confidence}{Enables confidence interval}

\item{level}{Confidence level}

\item{...}{Additional parameters}
}
\description{
Computes sigma for blblm
}
