% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Caculates Sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The lm models, of which to estimate the standard devation.}

\item{confidence}{Unknown,}

\item{level}{The confidence level for which to calculate the standard deviation.}

\item{...}{Other arguments}
}
\description{
Calculates the estimated standard deviation by taking the mean of each of the standard devations
estimated during each boostrap.
}
