% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Sigma of blblm Model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted blblm model.}

\item{confidence}{Single logical indicating whether the output should contain the confidence interval.}

\item{level}{Confidence level.}

\item{...}{Additional arguments.}
}
\value{
Estimated value for sigma of the given fitted blblm model and, optionally, the confidence interval for sigma at the specified confidence level.
}
\description{
Estimates sigma and, optionally, the confidence interval for sigma for a fitted blblm model.
}
