% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Sigma Value for blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{model fit produced by blblm(formula, data, ...)}

\item{confidence}{logical, TRUE gives a confidence interval, FALSE gives a single estimate.}

\item{level}{numeric, confidence interval level (only if confidence = TRUE).}

\item{...}{}
}
\description{
Outputs the value of sigma for the fitted object (after blblm algorithm)
}
