% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Gets sigma for BLBLM}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{the blblm model to use}

\item{confidence}{whether or not to include a confidence interval}

\item{level}{the confidence level to be used if confidence is \code{TRUE}, defaults to \code{0.95}}

\item{...}{extra arguments}
}
\description{
overloads the standard \code{sigma} functionality to get the error for a \code{blblm} model.
}
\examples{
blblm_model <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, num_workers = 4, parallelize = FALSE, fast = FALSE)
sigma(blblm_model)
sigma(blblm_model, confidence = TRUE)
sigma(blblm_model, confidence = TRUE, level = 0.99)
}
