% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{sigma.blblm is used to obtain the residual standard deviation}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{an R object typically from a linear model}

\item{confidence}{logical. If TRUE will output the confidence interval for the residual standard deviation}

\item{level}{an alpha level for the confidence interval}

\item{...}{additional arguments to be passed}
}
\description{
sigma.blblm is used to obtain the residual standard deviation
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100)
sigma(fit)
sigma(fit, confidence = TRUE)

}
