% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Computing the sigma of the blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level, ...)
}
\arguments{
\item{object}{the fit of blblm}

\item{confidence}{whether users want to use the confidence interval or not}

\item{level}{level of the confidence, Eg.0.95}

\item{...}{Other arguments}
}
\description{
Computing the sigma of the blblm
}
