% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Sigma value of regression}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The result from \code{blblm}}

\item{confidence}{determine if the confience interval is needed. (logical, default is F)}

\item{level}{The significant level (default is 0.95)}

\item{...}{Additional arguments to be passed to the function}
}
\value{
The sigma value(or values from confidence interval)
}
\description{
Sigma value of regression
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
sigma(fit)
sigma(fit, confidence = TRUE)
}
