% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{calculate the sigma for blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblm object}

\item{confidence}{whether do confidence interval. if TRUE, return confidence interval, if FALSE, return sigma only}

\item{level}{confidence level}
}
\value{
sigma
}
\description{
calculate the sigma for blblm
}
