% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{find the standard eror for each blb}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model from blblm}

\item{confidence}{boolean}

\item{level}{numeric, the confidence interval level}

\item{...}{additional arguements/inputs}
}
\value{
the standard error
}
\description{
find the standard eror for each blb
}
