% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Get the sd of blblm}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, nthreads = 1)
}
\arguments{
\item{object}{an object of "blblm"}

\item{confidence}{logic, need confidence interval or not}

\item{level}{the confidence level}

\item{nthreads}{the number of cores you want to use when processing}
}
\value{
When confidence = FALSE, return estiamted sd, else return sd and its upper bound and lower bound
}
\description{
Get the sd of blblm
}
