% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Calculate sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm model}

\item{confidence}{give a confidene interval or not}

\item{level}{confidence level}

\item{...}{other parameters}
}
\description{
Calculate sigma
}
