% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{sigma.blblog}
\alias{sigma.blblog}
\title{find the standard eror for each blb}
\usage{
\method{sigma}{blblog}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model from blblog}

\item{confidence}{boolean}

\item{level}{numeric, the confidence interval level}

\item{...}{additional arguements/inputs}
}
\value{
the standard error
}
\description{
find the standard eror for each blb
}
