% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{sigma.blblogreg}
\alias{sigma.blblogreg}
\title{Compute the residual standard deviation (sigma) of a logistic model}
\usage{
\method{sigma}{blblogreg}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{a blblogreg object, the fitted model}

\item{confidence}{boolean, specify whether to compute the confidence interval}

\item{level}{double, the significance level 1-alpha, 0.95 by default}

\item{...}{other conditions}
}
\value{
return the residual standard deviation (sigma) of the object (the fitted model)
}
\description{
Compute the residual standard deviation (sigma) of a logistic model
}
\examples{
> sigma(blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = Smarket, m = 3, B = 100, parallel = TRUE))
[1] 1.479956
> sigma(blblogreg(Direction ~ Lag1 + Lag2 + Volume, data = Smarket, m = 3, B = 100, parallel = TRUE), confidence = TRUE)
sigma      lwr      upr
1.479956 1.442451 1.517016
}
