% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblogreg}
\alias{sigma.blblogreg}
\title{Used to find the sigma of blblogreg}
\usage{
\method{sigma}{blblogreg}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblogreg function}

\item{confidence}{logical. If confidence = TRUE, we will use alpha in the calculations.}

\item{level}{confidence level. Set at 0.95.}

\item{...}{}
}
\value{
object of class "blblogreg"
}
\description{
Used to find the sigma of blblogreg
}
