% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{Data split}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{dataframe}

\item{m}{number of times to split the dataframe}
}
\description{
splits data in m parts that are approximately equal sizes.
}
\details{
split data into m parts of approximated equal sizes
}
