% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblm_logistic.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{a dataframe}

\item{m}{an integer that is the number of subsets in the data}
}
\value{
seperate dataframes

seperate dataframes
}
\description{
split data into m parts of approximated equal sizes

split data into m parts of approximated equal sizes
}
