% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m, seed = NaN)
}
\arguments{
\item{data}{Data Frame}

\item{m}{Number of Desired Splits}

\item{seed}{Optional seed for testing}
}
\value{
Returns a list of m subsections of data.
}
\description{
split data into m parts of approximated equal sizes
}
