% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{split_data}
\alias{split_data}
\title{Split data into Subsamples
split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{data frame we are working on}

\item{m}{number of subsamples}
}
\description{
Split data into Subsamples
split data into m parts of approximated equal sizes

Split data into Subsamples
split data into m parts of approximated equal sizes
}
