% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{a list of files, or a vector}

\item{m}{prefered size of data seperation}
}
\description{
split data into m parts of approximated equal sizes
}
