% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes.}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{m}{an integer specifying the number of chunks the data will be sliced into.}
}
\value{
a list of length m of sub-sampled data
}
\description{
to be filled
}
