% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{Split data}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{Data you want to split}

\item{m}{number of observations}
}
\description{
split data into m parts of approximated equal sizes
}
