% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm-logistic.R, R/blblm.R
\name{split_data}
\alias{split_data}
\title{Splitting Data into Approximately Equal Sizes
split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{a data frame or a character vector}

\item{m}{an integer giving the number of groups to split the data in}
}
\value{
m data frames of approximately equal sizes

m data frames of approximately equal sizes
}
\description{
Splitting Data into Approximately Equal Sizes
split data into m parts of approximated equal sizes

Splitting Data into Approximately Equal Sizes
split data into m parts of approximated equal sizes
}
