% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{dataframe}

\item{m}{numeric}
}
\value{
several datasets
}
\description{
split data into m parts of approximated equal sizes
}
\examples{
split_data(mtcars, 3)
}
