% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blbglm_cluster.R, R/blblm.R,
%   R/blblm_cluster.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)

split_data(data, m)

split_data(data, m)
}
\description{
split data into m parts of approximated equal sizes

split data into m parts of approximated equal sizes

split data into m parts of approximated equal sizes

split data into m parts of approximated equal sizes
}
