% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{m}{number of subsets for data}
}
\description{
split data into m parts of approximated equal sizes

split data into m parts of approximated equal sizes
}
