% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{the source data}

\item{m}{the number of subsamples}
}
\description{
An internal function to split data into m parts of approximated equal sizes
}
