% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{the data you would like to analyze.}

\item{m}{the number of sub-samples you want to divide your data into.}
}
\description{
split data into m parts of approximated equal sizes
}
