% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{the data you will use to pass into the function}

\item{m}{the number of parts you want your data to be split into.}
}
\description{
split data into m parts of approximated equal sizes
}
