% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R, R/blblm.R
\name{split_data}
\alias{split_data}
\title{Splitting Data}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{a data frame, list or environment containing variables in the model.}

\item{m}{an integer specifying the number of approximated equal size parts to be used for splitting data.}
}
\value{
a list of approximated equal size data.

a list of approximated equal size data.
}
\description{
Split data into m parts of approximated equal sizes.

Split data into m parts of approximated equal sizes.
}
