% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{input data}

\item{m}{number of subsamples}
}
\description{
split data into m parts of approximated equal sizes using random sampling
}
