% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split the data
split data into m subdata}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{the data you want to regression}

\item{m}{number of subdata}
}
\value{
data.frame
}
\description{
split the data
split data into m subdata
}
