% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{data}

\item{m}{number of splits}
}
\value{
list of splited data
}
\description{
Split data into m parts of approximated equal sizes
}
