% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/logistic_regression.R
\name{split_data}
\alias{split_data}
\title{Split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{dataframe}

\item{m}{integer}
}
\value{
split dataframe

split dataframe
}
\description{
Split data into m parts of approximated equal sizes

Split data into m parts of approximated equal sizes
}
