% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{Chosen dataset}

\item{m}{Number of times to split data for bootstrapping}
}
\description{
Splits data into m parts of approximated equal sizes
}
