% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblogreg.R
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{A data set, such as a data frame or list, containing the variables in the model.}

\item{m}{Number indicating how many sub-samples the data will be split into.}
}
\value{
A split data set of m sub-samples of approximated equal sizes.

split data into m parts of approximated equal sizes

A split data set of m sub-samples of approximated equal sizes.

split data into m parts of approximated equal sizes
}
\description{
Splits the given data set into m sub-samples of approximated equal sizes.

Splits the given data set into m sub-samples of approximated equal sizes.
}
