% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R, R/blblog.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)

split_data(data, m)
}
\arguments{
\item{data}{data frame}

\item{m}{numeric}
}
\value{
splitted dataframe

splitted dataframe
}
\description{
split data into m parts of approximated equal sizes
}
