% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{"split_data()" split data into m parts of approximated equal sizes
The sampling is with replacement.}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{dataframe}

\item{m}{integer, number of dataframes you want to split to}
}
\value{
a list of dataframes
}
\description{
"split_data()" split data into m parts of approximated equal sizes
The sampling is with replacement.
}
