% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_sample}
\alias{split_sample}
\title{Randomly split data into m parts}
\usage{
split_sample(data, m, min_subsample_size, even_split)
}
\arguments{
\item{data}{A data object to split.}

\item{m}{The number of splits to create.}

\item{min_subsample_size}{The minimum size of each split.}

\item{even_split}{Whether to create splits so they are as similarly-sized as possible.}
}
\value{
A list containing m subsets of data.
}
\description{
Randomly split data into m parts
}
